% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{accuracy}
\alias{accuracy}
\title{Calculate the accuracy}
\usage{
accuracy(TP, TN, FN, FP)
}
\arguments{
\item{TP}{Number of true positives}

\item{TN}{Number of true negatives}

\item{FN}{Number of false negatives}

\item{FP}{Number of false positives}
}
\value{
accuracy value
}
\description{
Calculate the accuracy
}
\details{
Calculate the accuracy
}
