% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadToDatabaseDiagnostics.R
\name{addMultipleDiagnosePlpToDatabase}
\alias{addMultipleDiagnosePlpToDatabase}
\title{Insert multiple diagnosePlp results saved to a directory into a PLP result schema database}
\usage{
addMultipleDiagnosePlpToDatabase(
  connectionDetails,
  databaseSchemaSettings,
  cohortDefinitions,
  databaseList = NULL,
  resultLocation
)
}
\arguments{
\item{connectionDetails}{A connection details created by using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{databaseSchemaSettings}{A object created by \code{createDatabaseSchemaSettings} with all the settings specifying the result tables}

\item{cohortDefinitions}{(list) A list of cohortDefinitions (each list must contain: name, id)}

\item{databaseList}{(Optional) ...}

\item{resultLocation}{The location of the diagnostic results}
}
\value{
Returns NULL but uploads multiple diagnosePlp results into the database schema specified in databaseSchemaSettings
}
\description{
This function inserts diagnosePlp results into the result schema
}
\details{
This function can be used to upload diagnosePlp results into a database
}
