% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadToDatabase.R
\name{addMultipleRunPlpToDatabase}
\alias{addMultipleRunPlpToDatabase}
\title{Populate the PatientLevelPrediction results tables}
\usage{
addMultipleRunPlpToDatabase(
  conn,
  databaseSchemaSettings = createDatabaseSchemaSettings(resultSchema = "main"),
  cohortDefinitions,
  databaseList = createDatabaseList(cdmDatabaseSchemas = c("cdm_truven_ccae_v123",
    "cdm_madeup_v1")),
  resultLocation = NULL,
  resultLocationVector,
  modelSaveLocation
)
}
\arguments{
\item{conn}{A connection to a database created by using the
function \code{connect} in the
\code{DatabaseConnector} package.}

\item{databaseSchemaSettings}{A object created by \code{createDatabaseSchemaSettings} with all the settings specifying the result tables}

\item{cohortDefinitions}{(list) A list of cohortDefinitions (each list must contain: name, id)}

\item{databaseList}{A list created by \code{createDatabaseList} to specify the databases}

\item{resultLocation}{(string) location of directory where the main package results were saved}

\item{resultLocationVector}{(only used when resultLocation is missing) a vector of locations with development or validation results}

\item{modelSaveLocation}{The location of the file system for saving the models in a subdirectory}
}
\value{
Returns NULL but uploads all the results in resultLocation to the PatientLevelPrediction result tables in resultSchema
}
\description{
This function formats and uploads results that have been generated via an ATLAS prediction package into a database
}
\details{
This function can be used upload PatientLevelPrediction results into a database
}
