% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recalibration.R
\name{addRecalibration}
\alias{addRecalibration}
\title{addRecalibration}
\usage{
addRecalibration(performanceEvaluation, recalibration)
}
\arguments{
\item{performanceEvaluation}{The main result performanceEvaluation}

\item{recalibration}{The recalibration result}
}
\value{
An object of class \code{runPlp} that is recalibrated on the new data
}
\description{
Adds the recalibration results to the main results
}
\details{
Append the recalibration results into the main results
}
