% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadToDatabase.R
\name{addRunPlpToDatabase}
\alias{addRunPlpToDatabase}
\title{Function to add the run plp (development or validation) to database}
\usage{
addRunPlpToDatabase(
  runPlp,
  conn,
  databaseSchemaSettings,
  cohortDefinitions,
  databaseList
)
}
\arguments{
\item{runPlp}{An object of class \code{runPlp} or class \code{externalValidatePlp}}

\item{conn}{A connection to a database created by using the
function \code{connect} in the
\code{DatabaseConnector} package.}

\item{databaseSchemaSettings}{A object created by \code{createDatabaseSchemaSettings} with all the settings specifying the result tables}

\item{cohortDefinitions}{(list) A list of cohortDefinitions (each list must contain: name, id)}

\item{databaseList}{A list created by \code{createDatabaseList} to specify the databases}
}
\value{
Returns a data.frame with the database details
}
\description{
This function adds a runPlp or external validation result into a database
}
\details{
This function is used when inserting results into the PatientLevelPrediction database results schema
}
