% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationSummary.R
\name{averagePrecision}
\alias{averagePrecision}
\title{Calculate the average precision}
\usage{
averagePrecision(prediction)
}
\arguments{
\item{prediction}{A prediction object}
}
\value{
The average precision value
}
\description{
Calculate the average precision
}
\details{
Calculates the average precision from a predition object
}
\examples{
prediction <- data.frame(
  value = c(0.1, 0.2, 0.3, 0.4, 0.5),
  outcomeCount = c(0, 1, 0, 1, 1)
)
averagePrecision(prediction)
}
