% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationSummary.R
\name{calibrationInLarge}
\alias{calibrationInLarge}
\title{Calculate the calibration in large}
\usage{
calibrationInLarge(prediction)
}
\arguments{
\item{prediction}{A prediction dataframe}
}
\value{
data.frame with meanPredictionRisk, observedRisk, and N
}
\description{
Calculate the calibration in large
}
\keyword{internal}
