% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{calibrationLine}
\alias{calibrationLine}
\title{calibrationLine}
\usage{
calibrationLine(prediction, numberOfStrata = 10, ...)
}
\description{
calibrationLine
}
\details{
Calculates the calibration from prediction object
}

