% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censorPlpData.R
\name{censorPlpData}
\alias{censorPlpData}
\title{Filters the plpData based on user specified criteria}
\usage{
censorPlpData(plpData, outcomeTime = NULL, newOutcome = NULL,
  predictionPeriod = NULL, dateInterval = NULL, minPriorObservation = 365,
  minCohortTime = NULL, excludeOutcomeOccurrence = list(`1` = c("inf", 0)),
  classificationCensor = list(insufficientCohortObservation = c("include",
  "include"), insufficientPredictionPeriod = c("include", "include"),
  minPostObservation = NULL, insufficientPostObservation = c("include",
  "include")), survivalCensor = list(useCohortObservation = F,
  usePredictionPeriod = T, maxPostObservation = NULL, useMaxPostObservation =
  F))
}
\arguments{
\item{plpData}{An object of type \code{plpData} - the patient level prediction 
data extracted from the CDM.}

\item{outcomeTime}{An integer - if you need to edit the time from index where you want to predict the
outcome use this parameter.  For example, if you created the outcome table 
by finding the occurrence of the outcome 30 days after cohort start but
you wish to conduct a sensitivity test for this and reduce it to 20 days,
then set outcomeTime=20.}

\item{newOutcome}{A vector of existing outcomeIds - constructs a new outcomeId based on people having all the specified outcomeIds.
For example, you may wish to predict the subset of the people who have the outcome
who also get given a secific treatment.  If you create the outcome with outcomeId 1
and the treatment with outcomeId 2, the set newOutcome =c(1,2) to find all the people
who have outcomeIds 1 and 2, they are then assigned a new outcomeId -1.}

\item{predictionPeriod}{A vector of length 2 with the first value corresponding to the
number of days after index to define the start of the risk prediction period
and the second value corresponding to the number of days after index
defining the end of the risk prediction period.  If this is NULL the the cohort
start and end date will define the risk prediction period.}

\item{dateInterval}{(a vector of 2 dates or NULL) corresponding to the inclusion dates. For
example, if the user inputs c(1990-01-01,2000-01-01) then any people with an index
prior to 1990 or after (Jan 1st 2000 minus minPriorObservaton) will be excluded. If
NULL all dates are included.}

\item{minPriorObservation}{an integer - people with the time in days between their observation start and cohort start 
less than this number will be removed from the data.}

\item{minCohortTime}{an integer - people with the time in days between the cohort start and cohort end
less than this number will be removed from the data.}

\item{excludeOutcomeOccurrence}{A list containing named list members and vectors of two integer values, where the name corresponds
to the outcomeId and the integer vector corresponds to an interval in days for filtering people 
who had the outcomeId recorded during this interval.  For example the list:
list('1'=c(180,40), '4'=c('inf',0)) would find all the people who had the outcomeId 1 recorded 
180 days prior to index up to 40 days after index and filter these people from the data, it would
also find the people who have outcomeId 4 anytime prior to index and filter these people from the data.}

\item{classificationCensor}{A list detailing the exclusion criteria for classification.  The list contains:
\describe{
\item{insufficientCohortObservation}{a character vector of length 2 with each element either 'include' 
or 'exclude' - indicating whether to include or exclude patients whose observation period ends 
before their cohort end.  The first element is applied to people with the outcome (class 1) and the second element is
applied to people without the outcome (class 0)}
\item{insufficientPredictionPeriod}{a character vector of length 2 with each element either 'include' 
or 'exclude' - indicating whether to include or exclude patients whose predictionPeriod falls outside of 
their observation period.  The first element is applied to people with the outcome (class 1) and the second element is
applied to people without the outcome (class 0)}
\item{minPostObservation}{An integer - specifying the required minimum number of days after index (Used by insufficientPostObservation) }
\item{insufficientPostObservation}{a character vector of length 2 with each element either 'include' 
or 'exclude' - indicating whether to include or exclude patients with an index date plus minPostObservation greater than
their observationEndDate.  The first element is applied to people with the outcome (class 1) and the second element is
applied to people without the outcome (class 0)}

}}

\item{survivalCensor}{A list containing the criteria for censoring the data...}
}
\value{
An object of type \code{plpData} containing information on the prediction problem that only contains the
data satisfying the user's specified censoring options. This object will
contain the following data:
\describe{ \item{cohorts}{An ffdf object listing all persons and their prediction periods. This
object will have these fields: row_id (a unique ID per period), person_id, cohort_start_date,
cohort_id, time (number of days in the window).} \item{outcomes}{An ffdf object listing all
outcomes per period. This object will have these fields: row_id, outcome_id, outcome_count,
time_to_event.} \item{exclude}{Either NULL or an ffdf object listing per outcome ID which windows
had the outcome prior to the window. This object will have these fields: rowId, outcomeId.}
\item{covariates}{An ffdf object listing the baseline covariates per person in the cohorts. This is
done using a sparse representation: covariates with a value of 0 are omitted to save space. The
covariates object will have three columns: rowId, covariateId, and covariateValue. }
\item{covariateRef}{An ffdf object describing the covariates that have been extracted.}
\item{metaData}{A list of objects with information on how the plpData object was constructed
and censoring details.  The list member named 'excluded' contains a ffdf of the excluded 
people and reason for exclusion.} }
}
\description{
Filters the data based on user censoring specifications for classification or survival analysis
}
\details{
Users can define a risk period of interest for the prediction of the outcome relative to index or use 
the cohprt dates.  The user can then specify whether they wish to exclude patients who are not observed
during the whole risk period, cohort period or experienced the outcome prior to the risk period.
}

