% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalValidatePlp.R
\name{checkAllSameInModels}
\alias{checkAllSameInModels}
\title{checkAllSameInModels - Check if all settings are the same across models}
\usage{
checkAllSameInModels(settingsList, settingName)
}
\description{
checkAllSameInModels - Check if all settings are the same across models
}
