% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FfHelperFunctions.R
\name{checkffFolder}
\alias{checkffFolder}
\title{Check if the fftempdir is writable}
\usage{
checkffFolder()
}
\description{
Check if the fftempdir is writable
}
\details{
This function checks whether the fftempdir is writable.
If not, it will ask the use to specify a writable folder.
}
