% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatting.R
\name{checkLibsvm}
\alias{checkLibsvm}
\title{Check plpData is in libSVM format and convert if needed}
\usage{
checkLibsvm(plpData, filePath = NULL, convert = TRUE)
}
\arguments{
\item{plpData}{An object of type \code{plpData} with covariate in coo format - the patient level prediction 
data extracted from the CDM.}

\item{filePath}{The path to the directory to output the libsvm files}

\item{convert}{Whether to convert or stop}
}
\value{
return a plpData object in LibSVM format as using for the Python calls
}
\description{
Allows to automatically convert to libsvm format if needed and saves it in the current
directory or a pre-defined path
}
\details{
This function checks if the plpData is in libsvm format
If not it calls convertToLibsvm if convert is TRUE or stops otherwise. 
The converted plpData is automatically saved to allow re-use.
}
\examples{
   
#TODO

}

