% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{checkPlpInstallation}
\alias{checkPlpInstallation}
\title{Check PatientLevelPrediction and its dependencies are correctly installed}
\usage{
checkPlpInstallation(connectionDetails = NULL, python = T)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the function
\code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{python}{Whether to test the python models}
}
\description{
Check PatientLevelPrediction and its dependencies are correctly installed
}
\details{
This function checks whether PatientLevelPrediction and its dependencies are correctly installed. This will
check the database connectivity, some models, and large data object
handling (ff).
}
