% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{checkSurvivalPackages}
\alias{checkSurvivalPackages}
\title{Check if the required packages for survival analysis are installed}
\usage{
checkSurvivalPackages()
}
\description{
Check if the required packages for survival analysis are installed
}
