% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalValidatePlp.R
\name{checkValidateExternalInputs}
\alias{checkValidateExternalInputs}
\title{checkValidateExternalInputs - Check the inputs for validateExternal}
\usage{
checkValidateExternalInputs(
  validationDesignList,
  databaseDetails,
  logSettings,
  outputFolder
)
}
\description{
checkValidateExternalInputs - Check the inputs for validateExternal
}
