% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMultiplePlp.R
\name{combinePlpModelSettings}
\alias{combinePlpModelSettings}
\title{combine two objects specifying multiple Plp model settings}
\usage{
combinePlpModelSettings(plpModelSetting1, plpModelSetting2)
}
\arguments{
\item{plpModelSetting1}{A combination of model, covariate and population settings}

\item{plpModelSetting2}{A combination of model, covariate and population settings}
}
\value{
A list containing a dataframe settingLookupTable containing all the model, covariate and popualtion combination details, 
a list models containing all the model settings, a list covariateSettings containing all the covariate settings and a list 
populationSettings containing all the population settings.
}
\description{
combine two objects specifying multiple Plp model settings
}
\details{
Takes two output of running createPlpModelSettings() and combined them
}
