% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationSummary.R
\name{computeAuprc}
\alias{computeAuprc}
\title{Compute the area under the Precision-Recall curve}
\usage{
computeAuprc(prediction)
}
\arguments{
\item{prediction}{A prediction object as generated using the
\code{\link{predict}} functions.}
}
\value{
A numeric value containing the AUPRC
}
\description{
Compute the area under the Precision-Recall curve
}
\details{
Computes the area under the Precision-Recall curve for the predicted scores, given the true observed
outcomes.
}
\examples{
prediction <- data.frame(
  value = c(0.1, 0.2, 0.3, 0.4, 0.5),
  outcomeCount = c(0, 1, 0, 1, 1))
computeAuprc(prediction)
}
