% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CovariatePrevalence.R
\name{computeCovariateMeans}
\alias{computeCovariateMeans}
\title{Compute covariate means}
\usage{
computeCovariateMeans(plpData, cohortId = NULL, outcomeId = NULL)
}
\arguments{
\item{plpData}{An object of type \code{plpData}.}

\item{cohortId}{The ID of the specific cohort for which to compute the means.}

\item{outcomeId}{The ID of the specific outcome for which to compute the subgroup means.}
}
\description{
Compute covariate means
}

