% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{computePreferenceScore}
\alias{computePreferenceScore}
\title{Compute the preference score}
\usage{
computePreferenceScore(prediction)
}
\arguments{
\item{prediction}{A prediction object as generated using the
\code{\link{predictProbabilities}} function.}
}
\value{
The preference score
}
\description{
Compute the preference score
}
\details{
Calculates the preference score from a predition object
}

