% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{configurePython}
\alias{configurePython}
\title{Configure a Python environment manually}
\usage{
configurePython(envname = "PLP", envtype = NULL, condaPythonVersion = "3.11")
}
\arguments{
\item{envname}{A string for the name of the virtual environment (default is 'PLP')}

\item{envtype}{An option for specifying the environment as 'conda' or 'python'. If NULL then the default is 'conda' for Windows users and 'python' for non-Windows users}

\item{condaPythonVersion}{String, Python version to use when creating a conda environment}
}
\value{
Location of the created conda or virtual Python environment
}
\description{
Configure a Python environment manually
}
\details{
PatientLevelPrediction normally lets \code{reticulate} manage Python requirements
automatically when Python-backed models are used. This helper is retained for
users who need to create a conda or virtualenv environment manually, for
example in offline or locked-down environments.
}
\examples{
\dontshow{ # dontrun reason: don't modify environment in examples }
\dontrun{ 
 configurePython(envname="PLP", envtype="conda")
}
}
