% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatting.R
\name{convertToLibsvm}
\alias{convertToLibsvm}
\title{Convert the plpData in COO format into the sparse libSVM format}
\usage{
convertToLibsvm(plpData, filePath = NULL)
}
\arguments{
\item{plpData}{An object of type \code{plpData} with covariate in coo format - the patient level prediction 
data extracted from the CDM.}

\item{filePath}{The path to the directory to output the libsvm files}
}
\value{
Returns an object of type \code{plpData}, containing information on the cohorts, their
outcomes, and baseline covariates. Information about multiple outcomes can be captured at once for
efficiency reasons. This object is a list with the following components: \describe{
\item{outcomes}{A data frame listing the outcomes per person, including the time to event, and
the outcome id. Outcomes are not yet filtered based on risk window, since this is done at
a later stage.} \item{cohorts}{A data frame listing the persons in each cohort, listing their
exposure status as well as the time to the end of the observation period and time to the end of the
cohort (usually the end of the exposure era).} \item{covariates}{A character object pointing to a directory
containing a libsvm file with the baseline covariates per person in the cohort. This is done using a sparse representation:
covariates with a value of 0 are omitted to save space.} \item{covariateRef}{An ffdf object describing the covariates that have been extracted.}
\item{metaData}{A list of objects with information on how the cohortMethodData object was
constructed.} }
}
\description{
Converts the standard plpData to libSVM format
}
\details{
This function converts the covariate file from ffdf in COO format into a libsvm format and
returns a plpData object with the covariate item representing the directory where the 
libsvm file is saved.
}
\examples{
   
#TODO

}

