% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackagePlp.R
\name{createCohort}
\alias{createCohort}
\title{createCohort - Loads all the cohort sql in a network study and creates the cohorts}
\usage{
createCohort(cohortDetails, cohortLocation, connectionDetails,
  cdmDatabaseSchema, cohortDatabaseSchema, cohortTable,
  oracleTempSchema = cdmDatabaseSchema, package)
}
\arguments{
\item{cohortDetails}{A dataframe containing two columns: cohortName and cohortId (if missing then the skeleton default is used when available)}

\item{cohortLocation}{A string specifying the location of the cohort sql files (uses default skeleton location if missing)}

\item{connectionDetails}{The connection details}

\item{cdmDatabaseSchema}{A string specifying the CDM database schema e.g., database.dbo}

\item{cohortDatabaseSchema}{A string specifying the cohort database schema e.g., cohort_database.dbo}

\item{cohortTable}{A string specifying the cohort table}

\item{oracleTempSchema}{Temp oracle schema}

\item{package}{The name of the package}
}
\value{
A data frame with the cohortName, cohortId, size
}
\description{
This function finds the sql files in a network study package, loads, renders and translates the sql then 
implements it
}
\details{
This is used by people running network studies using the package skeleton to enable users to create cohorts 
on their platform
}
