% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetConceptCovariates.R
\name{createConceptCovariateSettings}
\alias{createConceptCovariateSettings}
\title{Create Concept covariate settings}
\usage{
createConceptCovariateSettings(conceptList)
}
\arguments{
\item{conceptList}{A list of lists - each inner list contains two objects: conceptSet a vector of conceptSetIds 
and prior an integer specifying the number of days prior to index to search for the concepts in the set}
}
\value{
An object of type \code{conceptCovariateSettings}, to be used in other functions.
}
\description{
Create Concept covariate settings
}
\details{
creates an object specifying how covariates should be contructed from data in the CDM model.
}

