% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDefaultCovariates.R
\name{createCovariateSettings}
\alias{createCovariateSettings}
\title{Create covariate settings}
\usage{
createCovariateSettings(useCovariateCohortIdIs1 = FALSE,
  useCovariateDemographics = TRUE, useCovariateDemographicsGender = TRUE,
  useCovariateDemographicsRace = TRUE,
  useCovariateDemographicsEthnicity = TRUE,
  useCovariateDemographicsAge = TRUE, useCovariateDemographicsYear = TRUE,
  useCovariateDemographicsMonth = TRUE,
  useCovariateConditionOccurrence = TRUE,
  useCovariateConditionOccurrence365d = TRUE,
  useCovariateConditionOccurrence30d = FALSE,
  useCovariateConditionOccurrenceInpt180d = FALSE,
  useCovariateConditionEra = FALSE, useCovariateConditionEraEver = FALSE,
  useCovariateConditionEraOverlap = FALSE,
  useCovariateConditionGroup = FALSE,
  useCovariateConditionGroupMeddra = FALSE,
  useCovariateConditionGroupSnomed = FALSE,
  useCovariateDrugExposure = FALSE, useCovariateDrugExposure365d = FALSE,
  useCovariateDrugExposure30d = FALSE, useCovariateDrugEra = FALSE,
  useCovariateDrugEra365d = FALSE, useCovariateDrugEra30d = FALSE,
  useCovariateDrugEraOverlap = FALSE, useCovariateDrugEraEver = FALSE,
  useCovariateDrugGroup = FALSE, useCovariateProcedureOccurrence = FALSE,
  useCovariateProcedureOccurrence365d = FALSE,
  useCovariateProcedureOccurrence30d = FALSE,
  useCovariateProcedureGroup = FALSE, useCovariateObservation = FALSE,
  useCovariateObservation365d = FALSE, useCovariateObservation30d = FALSE,
  useCovariateObservationCount365d = FALSE, useCovariateMeasurement = FALSE,
  useCovariateMeasurement365d = FALSE, useCovariateMeasurement30d = FALSE,
  useCovariateMeasurementCount365d = FALSE,
  useCovariateMeasurementBelow = FALSE,
  useCovariateMeasurementAbove = FALSE, useCovariateConceptCounts = FALSE,
  useCovariateRiskScores = FALSE, useCovariateRiskScoresCharlson = FALSE,
  useCovariateRiskScoresDCSI = FALSE, useCovariateRiskScoresCHADS2 = FALSE,
  useCovariateRiskScoresCHADS2VASc = FALSE,
  useCovariateInteractionYear = FALSE, useCovariateInteractionMonth = FALSE,
  excludedCovariateConceptIds = c(), includedCovariateConceptIds = c(),
  deleteCovariatesSmallCount = 100)
}
\arguments{
\item{useCovariateCohortIdIs1}{A boolean value (TRUE/FALSE) to determine if a
covariate should be contructed for whether the
cohort ID is 1 (currently primarily used in
CohortMethod).}

\item{useCovariateDemographics}{A boolean value (TRUE/FALSE) to determine if
demographic covariates (age in 5-yr increments,
gender, race, ethnicity, year of index date, month
of index date) will be created and included in
future models.}

\item{useCovariateDemographicsGender}{A boolean value (TRUE/FALSE) to determine if gender
should be included in the model.}

\item{useCovariateDemographicsRace}{A boolean value (TRUE/FALSE) to determine if race
should be included in the model.}

\item{useCovariateDemographicsEthnicity}{A boolean value (TRUE/FALSE) to determine if
ethnicity should be included in the model.}

\item{useCovariateDemographicsAge}{A boolean value (TRUE/FALSE) to determine if age
(in 5 year increments) should be included in the
model.}

\item{useCovariateDemographicsYear}{A boolean value (TRUE/FALSE) to determine if
calendar year should be included in the model.}

\item{useCovariateDemographicsMonth}{A boolean value (TRUE/FALSE) to determine if
calendar month should be included in the model.}

\item{useCovariateConditionOccurrence}{A boolean value (TRUE/FALSE) to determine if
covariates derived from CONDITION_OCCURRENCE table
will be created and included in future models.}

\item{useCovariateConditionOccurrence365d}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of condition in 365d
window prior to or on cohort index date.  Only
applicable if useCovariateConditionOccurrence =
TRUE.}

\item{useCovariateConditionOccurrence30d}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of condition in 30d
window prior to or on cohort index date.  Only
applicable if useCovariateConditionOccurrence =
TRUE.}

\item{useCovariateConditionOccurrenceInpt180d}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of condition within
inpatient type in 180d window prior to or on cohort
index date.  Only applicable if
useCovariateConditionOccurrence = TRUE.}

\item{useCovariateConditionEra}{A boolean value (TRUE/FALSE) to determine if
covariates derived from CONDITION_ERA table will be
created and included in future models.}

\item{useCovariateConditionEraEver}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of condition era anytime
prior to or on cohort index date.  Only applicable
if useCovariateConditionEra = TRUE.}

\item{useCovariateConditionEraOverlap}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of condition era that
overlaps the cohort index date.  Only applicable if
useCovariateConditionEra = TRUE.}

\item{useCovariateConditionGroup}{A boolean value (TRUE/FALSE) to determine if all
CONDITION_OCCURRENCE and CONDITION_ERA covariates
should be aggregated or rolled-up to higher-level
concepts based on vocabluary classification.}

\item{useCovariateConditionGroupMeddra}{A boolean value (TRUE/FALSE) to determine if all
CONDITION_OCCURRENCE and CONDITION_ERA covariates
should be aggregated or rolled-up to higher-level
concepts based on the MEDDRA classification.}

\item{useCovariateConditionGroupSnomed}{A boolean value (TRUE/FALSE) to determine if all
CONDITION_OCCURRENCE and CONDITION_ERA covariates
should be aggregated or rolled-up to higher-level
concepts based on the SNOMED classification.}

\item{useCovariateDrugExposure}{A boolean value (TRUE/FALSE) to determine if
covariates derived from DRUG_EXPOSURE table will be
created and included in future models.}

\item{useCovariateDrugExposure365d}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of drug in 365d window
prior to or on cohort index date.  Only applicable
if useCovariateDrugExposure = TRUE.}

\item{useCovariateDrugExposure30d}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of drug in 30d window
prior to or on cohort index date.  Only applicable
if useCovariateDrugExposure = TRUE.}

\item{useCovariateDrugEra}{A boolean value (TRUE/FALSE) to determine if
covariates derived from DRUG_ERA table will be
created and included in future models.}

\item{useCovariateDrugEra365d}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of drug era in 365d
window prior to or on cohort index date.  Only
applicable if useCovariateDrugEra = TRUE.}

\item{useCovariateDrugEra30d}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of drug era in 30d window
prior to or on cohort index date.  Only applicable
if useCovariateDrugEra = TRUE.}

\item{useCovariateDrugEraOverlap}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of drug era that overlaps
the cohort index date.  Only applicable if
useCovariateDrugEra = TRUE.}

\item{useCovariateDrugEraEver}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of drug era anytime prior
to or on cohort index date.  Only applicable if
useCovariateDrugEra = TRUE.}

\item{useCovariateDrugGroup}{A boolean value (TRUE/FALSE) to determine if all
DRUG_EXPOSURE and DRUG_ERA covariates should be
aggregated or rolled-up to higher-level concepts of
drug classes based on vocabluary classification.}

\item{useCovariateProcedureOccurrence}{A boolean value (TRUE/FALSE) to determine if
covariates derived from PROCEDURE_OCCURRENCE table
will be created and included in future models.}

\item{useCovariateProcedureOccurrence365d}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of procedure in 365d
window prior to or on cohort index date.  Only
applicable if useCovariateProcedureOccurrence =
TRUE.}

\item{useCovariateProcedureOccurrence30d}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of procedure in 30d
window prior to or on cohort index date.  Only
applicable if useCovariateProcedureOccurrence =
TRUE.}

\item{useCovariateProcedureGroup}{A boolean value (TRUE/FALSE) to determine if all
PROCEDURE_OCCURRENCE covariates should be
aggregated or rolled-up to higher-level concepts
based on vocabluary classification.}

\item{useCovariateObservation}{A boolean value (TRUE/FALSE) to determine if
covariates derived from OBSERVATION table will be
created and included in future models.}

\item{useCovariateObservation365d}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of observation in 365d
window prior to or on cohort index date.  Only
applicable if useCovariateObservation = TRUE.}

\item{useCovariateObservation30d}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of observation in 30d
window prior to or on cohort index date.  Only
applicable if useCovariateObservation = TRUE.}

\item{useCovariateObservationCount365d}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for the count of each observation concept in
365d window prior to or on cohort index date.  Only
applicable if useCovariateObservation = TRUE.}

\item{useCovariateMeasurement}{A boolean value (TRUE/FALSE) to determine if
covariates derived from OBSERVATION table will be
created and included in future models.}

\item{useCovariateMeasurement365d}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of measurement in 365d
window prior to or on cohort index date.  Only
applicable if useCovariateMeasurement = TRUE.}

\item{useCovariateMeasurement30d}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of measurement in 30d
window prior to or on cohort index date.  Only
applicable if useCovariateMeasurement = TRUE.}

\item{useCovariateMeasurementCount365d}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for the count of each measurement concept in
365d window prior to or on cohort index date.  Only
applicable if useCovariateMeasurement = TRUE.}

\item{useCovariateMeasurementBelow}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of measurement with a
numeric value below normal range for latest value
within 180d of cohort index.  Only applicable if
useCovariateMeasurement = TRUE (CDM v5+) or
useCovariateObservation = TRUE (CDM v4).}

\item{useCovariateMeasurementAbove}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of measurement with a
numeric value above normal range for latest value
within 180d of cohort index.  Only applicable if
useCovariateMeasurement = TRUE (CDM v5+) or
useCovariateObservation = TRUE (CDM v4).}

\item{useCovariateConceptCounts}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
count the number of concepts that a person has
within each domain (CONDITION, DRUG, PROCEDURE,
OBSERVATION)}

\item{useCovariateRiskScores}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
calculate various Risk Scores, including Charlson,
DCSI.}

\item{useCovariateRiskScoresCharlson}{A boolean value (TRUE/FALSE) to determine if the
Charlson comorbidity index should be included in
the model.}

\item{useCovariateRiskScoresDCSI}{A boolean value (TRUE/FALSE) to determine if the
DCSI score should be included in the model.}

\item{useCovariateRiskScoresCHADS2}{A boolean value (TRUE/FALSE) to determine if the
CHADS2 score should be included in the model.}

\item{useCovariateRiskScoresCHADS2VASc}{A boolean value (TRUE/FALSE) to determine if the
CHADS2VASc score should be included in the model.}

\item{useCovariateInteractionYear}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
represent interaction terms between all other
covariates and the year of the cohort index date.}

\item{useCovariateInteractionMonth}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
represent interaction terms between all other
covariates and the month of the cohort index date.}

\item{excludedCovariateConceptIds}{A list of concept IDs that should NOT be used to
construct covariates.}

\item{includedCovariateConceptIds}{A list of concept IDs that should be used to
construct covariates.}

\item{deleteCovariatesSmallCount}{A numeric value used to remove covariates that
occur in both cohorts fewer than
deleteCovariateSmallCounts time.}
}
\value{
An object of type \code{defaultCovariateSettings}, to be used in other functions.
}
\description{
Create covariate settings
}
\details{
creates an object specifying how covariates should be contructed from data in the CDM model.
}

