% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadToDatabase.R
\name{createDatabaseList}
\alias{createDatabaseList}
\title{Create a data frame with the database details}
\usage{
createDatabaseList(cdmDatabaseSchemas, acronyms, versions, descriptions, types)
}
\arguments{
\item{cdmDatabaseSchemas}{(string vector) A vector of the cdmDatabaseSchemas used in the study}

\item{acronyms}{(optional string vector) A vector of the acronyms for the cdmDatabaseSchemas}

\item{versions}{(optional string vector) A vector of the database versions for the cdmDatabaseSchemas}

\item{descriptions}{(optional string vector) A vector of the database descriptions for the cdmDatabaseSchemas}

\item{types}{(optional string vector) A vector of the data types (e.g., claims or EHR) for the cdmDatabaseSchemas}
}
\value{
Returns a data.frame with the database details
}
\description{
This function creates a data.frame with the details of the databases used in the study
}
\details{
This function is used when inserting database details into the PatientLevelPrediction database results schema
}
