% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunPlpHelpers.R
\name{createDefaultExecuteSettings}
\alias{createDefaultExecuteSettings}
\title{Creates default list of settings specifying what parts of runPlp to execute}
\usage{
createDefaultExecuteSettings()
}
\value{
list with TRUE for split, preprocess, model development and covariate summary
}
\description{
Creates default list of settings specifying what parts of runPlp to execute
}
\details{
runs split, preprocess, model development and covariate summary
}
