% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalValidatePlp.R
\name{createDownloadTasks}
\alias{createDownloadTasks}
\title{createDownloadTasks
create download tasks based on unique combinations of targetId and
restrictPlpDataSettings. It adds all covariateSettings and outcomes that
have that targetId and restrictPlpDataSettings. This is used to avoid
downloading the same data multiple times.}
\usage{
createDownloadTasks(validationDesignList)
}
\arguments{
\item{validationDesignList}{A list of validationDesign objects}
}
\value{
A dataframe where each row is a downloadTask
}
\description{
createDownloadTasks
create download tasks based on unique combinations of targetId and
restrictPlpDataSettings. It adds all covariateSettings and outcomes that
have that targetId and restrictPlpDataSettings. This is used to avoid
downloading the same data multiple times.
}
\keyword{internal}
