% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{createEvaluateSubgroup}
\alias{createEvaluateSubgroup}
\title{run multiple evaluations of the population}
\usage{
createEvaluateSubgroup(
  plpResult,
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortJsonDirectory,
  outputFolder
)
}
\arguments{
\item{plpResult}{The result of running runPlp()}

\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where the
cohort data used to define the at risk cohort is available.
Requires read permissions to this database.}

\item{cohortJsonDirectory}{Location of the directory containing the JSON files for the desired subgroup cohorts}

\item{outputFolder}{The folder to save the results too}
}
\value{
plpResult evaluation object for the subgroup
}
\description{
run multiple evaluations of the population
}
\details{
create subgroups on database instance and run the subgroup analyses
}
