% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportExport.R
\name{createExistingModelSql}
\alias{createExistingModelSql}
\title{Apply an existing logistic regression prediction model}
\usage{
createExistingModelSql(modelTable, modelNames, interceptTable, covariateTable,
  type = "logistic", analysisId = 112, covariateSettings, asFunctions = F,
  customCovariates = NULL, e = environment(), covariateValues = F)
}
\arguments{
\item{modelTable}{A dataframe or list of dataframes with columns: modelId, modelCovariateId, coefficientValue all doubles}

\item{modelNames}{A name used in the covariate function names (no spaces)}

\item{interceptTable}{A dataframe or list of dataframes with the columns: modelId, interceptValue}

\item{covariateTable}{A dataframe or list of dataframes with columns: modelCovariateId, covariateId (the mapping of covariate_id to standard covariates)}

\item{type}{The type of model: logistic or linear/score}

\item{analysisId}{The covariate analysis_id (default 112)}

\item{covariateSettings}{The settings for the standard covariates (needs for temporal settings)}

\item{asFunctions}{If T then return two functions}

\item{customCovariates}{enables custome SQL to be used to create custom covariates}

\item{e}{The environment to output the covariate setting functions to}

\item{covariateValues}{boolean Whether to also download the covariates that make up the risk score}
}
\description{
Apply an existing logistic regression prediction model
}
\details{
This function is used to create custom covariates corresponding to existing models
}
