% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExistingPython.R
\name{createFeatureEngineeringMapColumnsSettings}
\alias{createFeatureEngineeringMapColumnsSettings}
\title{Create settings that enable you to convert from standard covariateIds to
model covariate names - this is useful when implementing a model developed
outside of the OHDSI tools.}
\usage{
createFeatureEngineeringMapColumnsSettings(columnMap)
}
\arguments{
\item{columnMap}{A data.frame containing the columns: covariateId the covariate ID from FeatureExtraction and
modelCovariateIdName which is the column name used when fitting the model.}
}
\value{
An object of class \code{featureEngineeringSettings} that will convert column names
}
\description{
Create settings that enable you to convert from standard covariateIds to
model covariate names - this is useful when implementing a model developed
outside of the OHDSI tools.
}
\details{
This function creates settings that let you rename the covariates in the plpData object
}
