% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering.R
\name{createFeatureEngineeringSettings}
\alias{createFeatureEngineeringSettings}
\title{Create the settings for defining any feature engineering that will be done}
\usage{
createFeatureEngineeringSettings(type = "none")
}
\arguments{
\item{type}{(character) Choice of:  \itemize{
\item{'none'}{ No feature engineering - this is the default }
}}
}
\value{
An object of class \code{featureEngineeringSettings}
}
\description{
Create the settings for defining any feature engineering that will be done
}
\details{
Returns an object of class \code{featureEngineeringSettings} that specifies the sampling function that will be called and the settings
}
