% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearningCurve.R
\name{createLearningCurvePar}
\alias{createLearningCurvePar}
\title{createLearningCurvePar}
\usage{
createLearningCurvePar(
  population,
  plpData,
  modelSettings,
  testSplit = "stratified",
  testFraction = 0.25,
  trainFractions = c(0.25, 0.5, 0.75),
  trainEvents = NULL,
  splitSeed = NULL,
  nfold = 3,
  indexes = NULL,
  verbosity = "TRACE",
  minCovariateFraction = 0.001,
  normalizeData = T,
  saveDirectory = getwd(),
  savePlpData = F,
  savePlpResult = F,
  savePlpPlots = F,
  saveEvaluation = F,
  timeStamp = FALSE,
  analysisId = "lc-",
  cores = NULL
)
}
\arguments{
\item{population}{The population created using \code{createStudyPopulation()}
that will be used to develop the model.}

\item{plpData}{An object of type \code{plpData} - the patient level
prediction data extracted from the CDM.}

\item{modelSettings}{An object of class \code{modelSettings} created using
one of the function. Currently only one model is supported:
\itemize{
  \item{\code{setLassoLogisticRegression} - a lasso logistic regression
    model}
}}

\item{testSplit}{Specifies the type of evaluation used. Can be either 
\code{'person'} or \code{'time'}. The value \code{'time'} finds the date
that splots the population into the testing and training fractions
provided. Patients with an index after this date are assigned to the test
set and patients with an index prior to this date are assigned to the
training set. The value \code{'person'} splits the data randomly into
testing and training sets according to fractions provided. The split is
stratified by the class label.}

\item{testFraction}{The fraction of the data, which will be used as the 
testing set in the patient split evaluation.}

\item{trainFractions}{A list of training fractions to create models for.
Note, providing \code{trainEvents} will override your input to
\code{trainFractions}.}

\item{trainEvents}{Events have shown to be determinant of model performance.
Therefore, it is recommended to provide \code{trainEvents} rather than
\code{trainFractions}. Note, providing \code{trainEvents} will override
your input to \code{trainFractions}. The format should be as follows:
\itemize{
  \item{\code{c(500, 1000, 1500)} - a list of training events}
}#' @param splitSeed The seed used to split the testing and training set when
using a 'person' type split}

\item{nfold}{The number of folds used in the cross validation (default = 
\code{3}).}

\item{indexes}{A dataframe containing a rowId and index column where the 
index value of -1 means in the test set, and positive integer represents
the cross validation fold (default is \code{NULL}).}

\item{verbosity}{Sets the level of the verbosity. If the log level is at or
higher in priority than the logger threshold, a message will print. The 
levels are:
\itemize{
  \item{\code{DEBUG} - highest verbosity showing all debug statements}
  \item{\code{TRACE} - showing information about start and end of steps}
  \item{\code{INFO} - show informative messages (default)}
  \item{\code{WARN} - show warning messages}
  \item{\code{ERROR} - show error messages}
  \item{\code{FATAL} - be silent except for fatal errors}
}}

\item{minCovariateFraction}{Minimum covariate prevalence in population to
avoid removal during preprocssing.}

\item{normalizeData}{Whether to normalise the data}

\item{saveDirectory}{Location to save log and results}

\item{savePlpData}{Whether to save the plpData}

\item{savePlpResult}{Whether to save the plpResult}

\item{savePlpPlots}{Whether to save the plp plots}

\item{saveEvaluation}{Whether to save the plp performance csv files}

\item{timeStamp}{Include a timestamp in the log}

\item{analysisId}{The analysis unique identifier}

\item{cores}{The number of cores to use}
}
\value{
A learning curve object containing the various performance measures
 obtained by the model for each training set fraction. It can be plotted
 using \code{plotLearningCurve}.
}
\description{
Creates a learning curve in parallel, which can be plotted using
 the \code{plotLearningCurve()} function. Currently this functionality is
 only supported by Lasso Logistic Regression.
}
\examples{
\dontrun{
# define model
modelSettings = setLassoLogisticRegression()

# register parallel backend
registerParallelBackend()

# create learning curve
learningCurve <- createLearningCurvePar(population,
                                        plpData,
                                        modelSettings)
# plot learning curve
plotLearningCurve(learningCurve)
}

}
