% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportExport.R
\name{createLrSql}
\alias{createLrSql}
\title{Convert logistic regression model to sql code...}
\usage{
createLrSql(models, modelNames, covariateConstructionName = "prediction",
  model_table = "#model_table", analysis_id = 111, e = environment())
}
\arguments{





}
\description{
Convert logistic regression model to sql code...
}
\details{
This function is used to create custom covariates for a logistic regression model
(currently only supports, demographics/conditions/drug/procedures/observations and measurement concepts)
}
