% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessingData.R
\name{createNormalization}
\alias{createNormalization}
\title{Create the settings for normalizing the data}
\usage{
createNormalization(type = "minmax")
}
\arguments{
\item{type}{The type of normalization to use, either "minmax" or "robust"}
}
\value{
An object of class \code{featureEngineeringSettings}
}
\description{
Create the settings for normalizing the data
}
