% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering.R
\name{createNormalizer}
\alias{createNormalizer}
\title{Create the settings for normalizing the data @param type The type of normalization to use, either "minmax" or "robust"}
\usage{
createNormalizer(type = "minmax")
}
\arguments{
\item{type}{The type of normalization to use, either "minmax" or "robust"}
}
\value{
An object of class \code{featureEngineeringSettings}

An object of class \code{featureEngineeringSettings}'
}
\description{
Create the settings for normalizing the data @param type The type of normalization to use, either "minmax" or "robust"
}
