% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePlpDocument.R
\name{createPlpDocument}
\alias{createPlpDocument}
\title{createPlpDocument}
\usage{
createPlpDocument(plpResult = NULL, plpData = NULL,
  targetName = "<target population>", outcomeName = "<outcome>",
  characterisationSettings = list(demo = T, utilization = T, condition = T,
  conditionNumber = 10, drug = T, drugNumber = 10, observation = F,
  observationNumber = 10, procedure = F, procedureNumber = 10, measurement = F,
  measurementNumber = 10, include = NULL), includeTrain = TRUE,
  includeTest = TRUE, includePredictionPicture = TRUE,
  includeAttritionPlot = TRUE, outputLocation = file.path(getwd(),
  "plp_document.docx"))
}
\arguments{
\item{plpResult}{An object of type \code{plpResult} returned by running runPlp()}

\item{plpData}{The plpData}

\item{targetName}{A string with the target description name}

\item{outcomeName}{A string with the outcome description name}

\item{characterisationSettings}{A list containing the settings to determine what covariates are included into the data summary table}

\item{includeTrain}{Whether to include the train set performance}

\item{includeTest}{Whether to include the test set performance}

\item{includePredictionPicture}{Whether to include a picture detailing the prediction problem}

\item{includeAttritionPlot}{Whether to include the attrition plot}

\item{outputLocation}{The location to write the document to}
}
\value{
A work document containing the selected outputs within the user's directory at location specified in outputLocation
}
\description{
Creates a template for a prediction paper with the characteristics/results filled in
}
\details{
The function creates a word document containing the analysis details, data summary and prediction model results.
}
