% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePlpDocument.R
\name{createPlpJournalDocument}
\alias{createPlpJournalDocument}
\title{createPlpJournalDocument}
\usage{
createPlpJournalDocument(plpResult = NULL, plpData = NULL,
  targetName = "<target population>", outcomeName = "<outcome>",
  characterisationSettings = list(demo = T, utilization = T, condition = T,
  conditionNumber = 10, drug = T, drugNumber = 10, observation = F,
  observationNumber = 10, procedure = F, procedureNumber = 10, measurement = F,
  measurementNumber = 10, include = NULL), includeTrain = FALSE,
  includeTest = TRUE, includePredictionPicture = TRUE,
  includeAttritionPlot = TRUE, outputLocation = file.path(getwd(),
  "plp_journal_document.docx"))
}
\arguments{
\item{plpResult}{An object of type \code{plpResult} returned by running runPlp()}

\item{plpData}{The plpData}

\item{targetName}{A string with the target description name}

\item{outcomeName}{A string with the outcome description name}

\item{characterisationSettings}{A list containing the settings to determine what covariates are included into the data summary table}

\item{includeTrain}{Whether to include the train set performance}

\item{includeTest}{Whether to include the test set performance}

\item{includePredictionPicture}{Whether to include a picture detailing the prediction problem}

\item{outputLocation}{The location to write the document to}

\item{includeAttrictionPlot}{Whether to include the attriction plot}

\item{includeAnalysisPlot}{Whether to include the analysis flowchart plot}
}
\value{
A work document containing the selected outputs within the user's directory at location specified in outputLocation
}
\description{
Creates a template for a prediction journal paper with the characteristics/results filled in
}
\details{
The function creates a word document containing the analysis details, data summary and prediction model results.
}
