% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMultiplePlp.R
\name{createPlpModelSettings}
\alias{createPlpModelSettings}
\title{create a an object specifying the multiple Plp model settings}
\usage{
createPlpModelSettings(modelList, covariateSettingList, populationSettingList)
}
\arguments{
\item{modelList}{A list of model settings}

\item{covariateSettingList}{A list of covariate settings}

\item{populationSettingList}{A list of population settings}
}
\value{
A list containing a dataframe settingLookupTable containing all the model, covariate and popualtion combination details, 
a list models containing all the model settings, a list covariateSettings containing all the covariate settings and a list 
populationSettings containing all the population settings.
}
\description{
create a an object specifying the multiple Plp model settings
}
\details{
Takes a list of models, covariates, population and returns the cartesian product combining all
settings.
}
