% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePlpDocument.R
\name{createPlpReport}
\alias{createPlpReport}
\title{createPlpReport}
\usage{
createPlpReport(plpResult = NULL, plpValidation = NULL, plpData = NULL,
  targetName = "<target population>", outcomeName = "<outcome>",
  targetDefinition = NULL, outcomeDefinition = NULL,
  outputLocation = file.path(getwd(), "plp_report.docx"), save = T)
}
\arguments{
\item{plpResult}{An object of type \code{plpResult} returned by running runPlp()}

\item{plpValidation}{An object of type \code{validatePlp} returned by running externalValidatePlp()}

\item{plpData}{The plpData}

\item{targetName}{A string with the target description name}

\item{outcomeName}{A string with the outcome description name}

\item{targetDefinition}{The cohort details}

\item{outcomeDefinition}{The cohort details}

\item{outputLocation}{The location to write the document to}

\item{save}{If false the output of the function of the function is the document rather than creating the document in outputLocation}
}
\value{
A work document containing the selected outputs within the user's directory at location specified in outputLocation
}
\description{
Creates a word document report of the prediction
}
\details{
The function creates a word document containing the analysis details, data summary and prediction model results.
}
