% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExistingPython.R
\name{createSciKitLearnModel}
\alias{createSciKitLearnModel}
\title{Plug an existing scikit learn python model into the
PLP framework}
\usage{
createSciKitLearnModel(
  modelLocation = "/model",
  covariateMap = data.frame(columnId = 1:2, covariateId = c(1, 2), modelCovariateIdName =
    c("pred_1", "pred_2")),
  covariateSettings,
  populationSettings,
  isPickle = TRUE
)
}
\arguments{
\item{modelLocation}{The location of the folder that contains the model as model.pkl}

\item{covariateMap}{A data.frame with the columns: columnId specifying the column order for the
covariate, covariateId the covariate ID from FeatureExtraction and modelCovariateIdName which is the
column name used when fitting the model.  For example, if you had a column called 'age' in your model and this was the 3rd
column when fitting the model, then the values for columnId would be 3, covariateId would be 1002 (the covariateId for age in years) and
modelCovariateIdName would be 'age'.}

\item{covariateSettings}{The settings for the standardized covariates}

\item{populationSettings}{The settings for the population, this includes the time-at-risk settings and
and inclusion criteria.}

\item{isPickle}{If the model is saved as a pickle set this to T if it is a json set this to F}
}
\value{
An object of class plpModel, this is a list that contains: model (the location of the model.pkl),
preprocessing (settings for mapping the covariateIds to the model column mames), modelDesign (specification
of the model design), trainDetails (information about the model fitting) and covariateImportance.  You can use the output
as an input in PatientLevelPrediction::predictPlp to apply the model and calculate the risk for patients.
}
\description{
Plug an existing scikit learn python model into the
PLP framework
}
\details{
This function lets users add an existing scikit learn that is saved as model.pkl
into PLP format.  covariateMap is a mapping between standard covariateIds and the model column names
and order are required in addition to pythonModelLocation, the location of the model that must be saved
as `model.pkl`.  The user also needs to specify the covariate settings and population settings as these
are used to determine the standard PLP model design.
}
