% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{createSimpleImputer}
\alias{createSimpleImputer}
\title{Create Simple Imputer settings}
\usage{
createSimpleImputer(method = "mean", missingThreshold = 0.3)
}
\arguments{
\item{method}{The method to use for imputation, either "mean" or "median"}

\item{missingThreshold}{The threshold for missing values to be imputed vs removed}
}
\value{
The settings for the single imputer of class `featureEngineeringSettings`
}
\description{
This function creates the settings for a simple imputer
which imputes missing values with the mean or median
}
