% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{createSklearnIterativeImputer}
\alias{createSklearnIterativeImputer}
\title{Create scikit-learn Iterative Imputer settings}
\usage{
createSklearnIterativeImputer(
  missingThreshold = 0.3,
  methodSettings = list(),
  addMissingIndicator = FALSE
)
}
\arguments{
\item{missingThreshold}{The threshold for missing values to remove a feature}

\item{methodSettings}{A list of settings for sklearn \code{IterativeImputer}.
Supported settings are:
\itemize{
\item maxIter
\item tol
\item samplePosterior
\item nNearestFeatures
\item initialStrategy
\item imputationOrder
\item skipComplete
\item randomState
\item minValue
\item maxValue
}}

\item{addMissingIndicator}{Add a binary missingness indicator per feature that
passes the imputation missingness threshold.}
}
\value{
The settings for the sklearn iterative imputer of class \code{featureEngineeringSettings}
}
\description{
This function creates settings for a dense iterative imputer
powered by scikit-learn's \code{IterativeImputer} through \code{reticulate}.
}
\examples{
\dontshow{ # dontrun reason: requires python and scikit-learn }
\dontrun{
createSklearnIterativeImputer(
  missingThreshold = 0.3,
  methodSettings = list(maxIter = 5, nNearestFeatures = 20)
)
}
}
