% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering.R
\name{createSplineSettings}
\alias{createSplineSettings}
\title{Create the settings for adding a spline for continuous variables}
\usage{
createSplineSettings(continousCovariateId, knots, analysisId = 683)
}
\arguments{
\item{continousCovariateId}{The covariateId to apply splines to}

\item{knots}{Either number of knots of vector of split values}

\item{analysisId}{The analysisId to use for the spline covariates}
}
\value{
An object of class \code{featureEngineeringSettings}
}
\description{
Create the settings for adding a spline for continuous variables
}
\details{
Returns an object of class \code{featureEngineeringSettings} that specifies the sampling function that will be called and the settings
}
\examples{
# create splines for age (1002) with 5 knots
createSplineSettings(continousCovariateId = 1002, knots = 5, analysisId = 683)
}
