% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{createTempModelLoc}
\alias{createTempModelLoc}
\title{Create a temporary model location}
\usage{
createTempModelLoc()
}
\value{
A string for the location of the temporary model location
}
\description{
Create a temporary model location
}
