% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetTextCovariates.R
\name{createTextCovariateSettings}
\alias{createTextCovariateSettings}
\title{Create text covariate settings}
\usage{
createTextCovariateSettings(language = "eng", removeNegations = TRUE,
  deleteCovariatesSmallCount = 100)
}
\arguments{
\item{language}{Specify the language of the free-text.}

\item{removeNegations}{Remove negated text prior to constructing features.}

\item{deleteCovariatesSmallCount}{A numeric value used to remove covariates that occur in both
                                    cohorts fewer than deleteCovariateSmallCounts time.}
}
\value{
An object of type \code{covariateSettings}, to be used in other functions.
}
\description{
Create text covariate settings
}
\details{
creates an object specifying how covariates should be constructed from text in notes table in the
CDM model.
}

