% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlmNetModels.R
\name{cvGlmNet}
\alias{cvGlmNet}
\title{cvGlmNet}
\usage{
cvGlmNet(dataMatrix, labels, param, covariateMap)
}
\arguments{
\item{dataMatrix}{R sparse matrix object}

\item{labels}{population dataframe with the labels}

\item{covariateMap}{Dataframe that maps covariates to columns in the sparse matrix}

\item{hyperParamGrid}{Hyperparameters to try}
}
\description{
performs hyperparameter search using cross validation for glmnet models
}
