% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalValidatePlp.R
\name{deDuplicateCovariateData}
\alias{deDuplicateCovariateData}
\title{deplucateCovariateData - Remove duplicate covariate data
when downloading data with multiple different covariateSettings sometimes
there will be duplicated analysisIds which need to be removed}
\usage{
deDuplicateCovariateData(covariateData)
}
\arguments{
\item{covariateData}{The covariate data Andromeda object}
}
\value{
The deduplicated covariate data
}
\description{
deplucateCovariateData - Remove duplicate covariate data
when downloading data with multiple different covariateSettings sometimes
there will be duplicated analysisIds which need to be removed
}
\keyword{internal}
