% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describePlpData.R
\name{describePlpData}
\alias{describePlpData}
\title{describePlpData}
\usage{
describePlpData(plpData, covariateVals = NULL, cdmDatabase, outcomeId = 2,
  agehist = TRUE, plot = T, plotFile = NULL, saveTable = T,
  tableFile = NULL, outcomeName = "Pregnancy with GDM",
  cohortName = "Total pregnancy", perYear = T)
}
\arguments{
\item{plpData}{An object of type \code{plpData}.}

\item{covariateVals}{A vector of the covariateIds to focus on}

\item{cdmDatabase}{The database the data comes from}

\item{outcomeId}{The outcome of interest}

\item{agehist}{Whether to plot a histogram of ages}

\item{plot}{Whether to plot results}

\item{plotFile}{Location to save plot}

\item{saveTable}{Whether to save results}

\item{tableFile}{Location to save results table}

\item{outcomeName}{Outcome name to be used on plots}

\item{cohortName}{Cohort name to be used on plots}

\item{perYear}{Whether to do descriptive stats per year}
}
\description{
describePlpData
}
\details{
Describes the data
}

