% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developModel.R
\name{developModel}
\alias{developModel}
\title{developModel - Train and elavuate the model}
\usage{
developModel(plpData, modelSettings = list(model = "lr-lasso", param = NULL,
  cohortId = NULL, outcomeId = 2), featureSettings = list(analysisSelector =
  NULL, covariateSelector = NULL, wrapper = list(method = "lr-lasso", variance =
  0.01), matrixFactor = F), validationFraction = 0.2,
  fileLoc = file.path(getwd(), "models"), type = c("year", "both",
  "patient"))
}
\arguments{
\item{plpData}{An object of type \code{plpData} - the patient level prediction
data extracted from the CDM.}

\item{modelSettings}{A list defining the model to train, the cohortId, the outcomeId,
                                        preprocessing options and model parameters.

                                        The potential models are:
                                        Logistic regression with lasso regularisation: 'lr-lasso'
                                        Random forest: 'rf'
                                        Gradient boosting machineL 'gbm'
                                        Logistic regression with elastic new regularisation: 'lr-enet'

                                        The potential preprocessing settings are:
                                        Wrapper feature selection using lasso logistic regression: 'lr-lasso'
                                        Selecting only the condition/drug era features: 'allEra'

                                        The parameter setting depend on the model-
                                        model='lr-lasso' has the parameter val specificying the initial variance
                                        model='rf' has the parameters ntree, max_depth and mtry
                                        model='gbm' has the parameters ntree, bal (class balance), nrowSample (fraction of training data people to use per tree),
                                               ncolSample (fraction of training data features to include into each tree)
                                        model='lr-enet' has the parameters alpha, ...

                                        example: list(model='gbm', cohortId=NULL, outcomeId=c(1,2),
                                                      preprocess='lr-lasso', param=list(ntree=50, bal=F, nrowsample=0.6))
                                        Would train a gradient boosting machine to predict the outcomes 1 and 2
                                        using the features selected by logistic regression with lasso regularisation
                                        with the model settings ntree 50, no class label balance and 0.6 training data rows
                                        used per tree.}

\item{validationFraction}{The fraction of the data to be used as the validation set in the patient
split evaluation.}

\item{fileLoc}{The path to the directory where the models will be saved}

\item{type}{A subset of c('year','both','patient') specifying the type of evaluation used.
'year' splits the date prior to 2013 into the training set and post 2013 into the test set
'both' splits the data by the year 2013 but removes any patient in the test set from the training set
'patient' splits the data into test (1-validationFraction of the data) and
train (validationFraction of the data) sets.  The split is stratified by the class label.}
}
\value{
An object containing the model or location where the model is save, the data selection settings, the preprocessing
and training settings as well as various performance measures obtained by the model.
\describe{
\item{model}{The trained prediction model}
\item{dataSummary}{A list detailing the size of the train/test sets and outcome prevalence}
\item{evalType}{The type of evaluation that was performed}
\item{prediction}{The model prediction on the test set }
\item{performance}{A list detailing the performance of the model}}
}
\description{
Trains various machine learning models and elavulates them
}
\details{
Users can define a risk period of interest for the prediction of the outcome relative to index or use
the cohprt dates.  The user can then specify whether they wish to exclude patients who are not observed
during the whole risk period, cohort period or experienced the outcome prior to the risk period.
}

