% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{diagnostic}
\alias{diagnostic}
\title{diagnostic - Investigates the prediction problem settings - use before training a model}
\usage{
diagnostic(
  plpData = NULL,
  cdmDatabaseName,
  connectionDetails,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  cohortId,
  outcomeIds,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  outcomeDatabaseSchema = cohortDatabaseSchema,
  outcomeTable = cohortTable,
  cdmVersion = 5,
  riskWindowStart = 1,
  startAnchor = "cohort start",
  riskWindowEnd = 365,
  endAnchor = "cohort start",
  outputFolder = NULL,
  sampleSize = NULL
)
}
\arguments{
\item{plpData}{The data object to do the diagnostic on - if NULL you need to specify the connection settings below}

\item{cdmDatabaseName}{Name of the database}

\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you want
all temporary tables to be managed. Requires create/insert
permissions to this database.}

\item{cohortId}{A unique identifier to define the at risk cohorts. CohortId is
used to select the cohort_concept_id in the cohort-like table.}

\item{outcomeIds}{A list of cohort_definition_ids used to define outcomes.}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where the
cohort data used to define the at risk cohort is available.
If cohortTable = DRUG_ERA, cohortDatabaseSchema is not used
by assumed to be cdmSchema.  Requires read permissions to this
database.}

\item{cohortTable}{The tablename that contains the at risk cohort.  If
cohortTable <> DRUG_ERA, then expectation is cohortTable has
format of COHORT table: cohort_concept_id, SUBJECT_ID,
COHORT_START_DATE, COHORT_END_DATE.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where
the data used to define the outcome cohorts is available. If
cohortTable = CONDITION_ERA, exposureDatabaseSchema is not
used by assumed to be cdmSchema.  Requires read permissions
to this database.}

\item{outcomeTable}{The tablename that contains the outcome cohorts.  If
outcomeTable <> CONDITION_OCCURRENCE, then expectation is
outcomeTable has format of COHORT table:
COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE,
COHORT_END_DATE.}

\item{cdmVersion}{Define the OMOP CDM version used: currently support "4" and "5".}

\item{riskWindowStart}{The start of the risk window (in days) relative to the \code{startAnchor}.}

\item{startAnchor}{The anchor point for the start of the risk window. Can be "cohort start" or "cohort end".}

\item{riskWindowEnd}{The end of the risk window (in days) relative to the \code{endAnchor} parameter}

\item{endAnchor}{The anchor point for the end of the risk window. Can be "cohort start" or "cohort end".}

\item{outputFolder}{Location to save results for shiny app}

\item{sampleSize}{Sample from the target population}
}
\value{
An object containing the model or location where the model is save, the data selection settings, the preprocessing
and training settings as well as various performance measures obtained by the model.

\item{distribution}{list for each O of a data.frame containing: i) Time to observation end distribution, ii) Time from observation start distribution, iii) Time to event distribution and iv) Time from last prior event to index distribution (only for patients in T who have O before index) }
\item{incident}{list for each O of incidence of O in T during TAR}
\item{characterization}{list for each O of Characterization of T, TnO, Tn~O}
}
\description{
This function runs a set of prediction diagnoses to help pick a suitable T, O, TAR and determine 
whether the prediction problem is worth executing.
}
\details{
Users can define set of Ts, Os, databases and population settings.  A list of data.frames containing details such as
follow-up time distribution, time-to-event information, characteriszation details, time from last prior event, 
observation time distribution.
}
\examples{
\dontrun{
#******** EXAMPLE 1 ********* 
} 
}
