% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{diagnosticOddsRatio}
\alias{diagnosticOddsRatio}
\title{Calculate the diagnostic odds ratio}
\usage{
diagnosticOddsRatio(TP, TN, FN, FP)
}
\arguments{
\item{TP}{Number of true positives}

\item{TN}{Number of true negatives}

\item{FN}{Number of false negatives}

\item{FP}{Number of false positives}
}
\value{
diagnosticOddsRatio value
}
\description{
Calculate the diagnostic odds ratio
}
\details{
Calculate the diagnostic odds ratio
}
