% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CyclopsModels.R
\name{doCyclopsCVPenalty}
\alias{doCyclopsCVPenalty}
\title{do simple CV to determine best penalty manually}
\usage{
doCyclopsCVPenalty(
  data,
  prior,
  fixedCoefficients,
  startingCoefficients,
  penaltyRatio = 0.1,
  nTries = 10
)
}
\arguments{
\item{data}{The training data}

\item{prior}{Cyclops prior to use}

\item{fixedCoefficients}{What coefficients (if any) should be fixed}

\item{startingCoefficients}{What coefficients (if any) should have some starting value}

\item{penaltyRatio}{This controls the lowest penalty to try as a ratio of `BIC` penalty.
Trying very low penalties will increase computation times a lot.}

\item{nTries}{How many penalties to try. Default: 10}
}
\description{
do simple CV to determine best penalty manually
}
\details{
Will try a sequence of penalties from `BIC` down to `penaltyRatio` * `BIC`. How many penalties
to try is determined by `nTries.` Will use cross-validation to determine optimal penalty
based on `AUC`.
}
