% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttritionDiagramPlp.R
\name{drawAttritionDiagramPlp}
\alias{drawAttritionDiagramPlp}
\title{Draw the attrition diagram}
\usage{
drawAttritionDiagramPlp(attrition, targetLabel = "Target Population",
  outcomeLabel = "Outcome Count", fileName = NULL)
}
\arguments{
\item{attrition}{The table of attrition details return from the population attr(popualtion, 'metaData')$attrition}

\item{targetLabel}{A label to us for the treated cohort.}

\item{outcomeLabel}{A label to us for the comparator cohort.}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'.
See the function \code{ggsave} in the ggplot2 package for supported file
formats.}
}
\value{
A ggplot object. Use the \code{\link[ggplot2]{ggsave}} function to save to file in a different
format.
}
\description{
\code{drawAttritionDiagramPlp} draws the attition diagram, showing how many people were excluded from
the study population, and for what reasons.
}
