% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMultiplePlp.R
\name{evaluateMultiplePlp}
\alias{evaluateMultiplePlp}
\title{externally validate the multiple plp models across new datasets}
\usage{
evaluateMultiplePlp(analysesLocation, outputLocation, connectionDetails,
  validationSchemaTarget, validationSchemaOutcome, validationSchemaCdm,
  databaseNames, validationTableTarget, validationTableOutcome,
  validationIdTarget = NULL, validationIdOutcome = NULL,
  oracleTempSchema = NULL, verbosity = "INFO", keepPrediction = F,
  sampleSize = NULL)
}
\arguments{
\item{analysesLocation}{The location where the multiple plp analyses are}

\item{outputLocation}{The location to save to validation results}

\item{connectionDetails}{The connection details for extracting the new data}

\item{validationSchemaTarget}{A string or list of strings specifying the database containing the target cohorts}

\item{validationSchemaOutcome}{A string or list of strings specifying the database containing the outcome cohorts}

\item{validationSchemaCdm}{A string or list of strings specifying the database containing the cdm}

\item{databaseNames}{A string of lift of strings specifying sharing friendly database names corresponding to validationSchemaCdm}

\item{validationTableTarget}{A string or list of strings specifying the table containing the target cohorts}

\item{validationTableOutcome}{A string or list of strings specifying the table containing the outcome cohorts}

\item{validationIdTarget}{An iteger or list of integers specifying the cohort id for the target cohorts}

\item{validationIdOutcome}{An iteger or list of integers specifying the cohort id for the outcome cohorts}

\item{oracleTempSchema}{The temp oracle schema requires read/write}

\item{verbosity}{Sets the level of the verbosity. If the log level is at or higher in priority than the logger threshold, a message will print. The levels are:
\itemize{
\item{DEBUG}{Highest verbosity showing all debug statements}
\item{TRACE}{Showing information about start and end of steps}
\item{INFO}{Show informative information (Default)}
\item{WARN}{Show warning messages}
\item{ERROR}{Show error messages}
\item{FATAL}{Be silent except for fatal errors}
}}

\item{keepPrediction}{Whether to keep the predicitons for the new data}

\item{sampleSize}{If not NULL, the number of people to sample from the target cohort}
}
\description{
This function loads all the models in a multiple plp analysis folder and
validates the models on new data
}
\details{
Users need to input a location where the results of the multiple plp analyses
are found and the connection and database settings for the new data
}
