% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluatePlp.R
\name{evaluatePlp}
\alias{evaluatePlp}
\title{evaluatePlp}
\usage{
evaluatePlp(plpPredict, plpData, sparse = T)
}
\arguments{
\item{plpPredict}{The patient level prediction model's prediction}

\item{plpData}{An object of type \code{plpData} - the patient level prediction
data extracted from the CDM.}

\item{sparse}{(boolean) Whether the metrics should be calculated in sparse format}
}
\value{
A list containing the performance values
}
\description{
Evaluates the performance of the patient level prediction model
}
\details{
The function calculates various metrics to measure the performance of the model
}

