% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackagePlp.R
\name{exportPlpResult}
\alias{exportPlpResult}
\title{exportPlpResult exports an object returned by runPlp into a network study package while removing 
sensitive information from the object}
\usage{
exportPlpResult(plpResult, modelName, packageName, gitHubLocation, n = NULL,
  includeEvaluationStatistics = T, includeThresholdSummary = T,
  includeDemographicSummary = T, includeCalibrationSummary = T,
  includePredictionDistribution = T, includeCovariateSummary = F)
}
\arguments{
\item{plpResult}{The model to be saved into the package}

\item{modelName}{The name of the model}

\item{packageName}{The network study package name}

\item{gitHubLocation}{The github directory}

\item{n}{If not null, the minimum number of people required for a row to be included}

\item{includeEvaluationStatistics}{Whether to include includeEvaluationStatistics evaluation}

\item{includeThresholdSummary}{Whether to include thresholdSummary evaluation}

\item{includeDemographicSummary}{Whether to include demographicSummary evaluation}

\item{includeCalibrationSummary}{Whether to include calibrationSummary evaluation}

\item{includePredictionDistribution}{Whether to include predictionDistribution evaluation}

\item{includeCovariateSummary}{Whether to include covariateSummary evaluation}
}
\value{
The location of the saved model
}
\description{
This function should be used to export a prediciton model and performance information into a 
network study for others to implement on new data for external validation
}
\details{
This is a helper function to perform external validation
}
