% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalValidatePlp.R
\name{externalValidatePlp}
\alias{externalValidatePlp}
\title{externalValidatePlp - Validate a model on new databases}
\usage{
externalValidatePlp(plpResult, connectionDetails, validationSchemaTarget,
  validationSchemaOutcome, validationSchemaCdm, databaseNames,
  validationTableTarget = "cohort", validationTableOutcome = "cohort",
  validationIdTarget = NULL, validationIdOutcome = NULL,
  oracleTempSchema = NULL, verbosity = "INFO", keepPrediction = F,
  sampleSize = NULL)
}
\arguments{
\item{plpResult}{The object returned by runPlp() containing the trained model}

\item{connectionDetails}{The connection details for extracting the new data}

\item{validationSchemaTarget}{A string or list of strings specifying the database containing the target cohorts}

\item{validationSchemaOutcome}{A string or list of strings specifying the database containing the outcome cohorts}

\item{validationSchemaCdm}{A string or list of strings specifying the database containing the cdm}

\item{databaseNames}{A string of lift of strings specifying sharing friendly database names corresponding to validationSchemaCdm}

\item{validationTableTarget}{A string or list of strings specifying the table containing the target cohorts}

\item{validationTableOutcome}{A string or list of strings specifying the table containing the outcome cohorts}

\item{validationIdTarget}{An iteger or list of integers specifying the cohort id for the target cohorts}

\item{validationIdOutcome}{An iteger or list of integers specifying the cohort id for the outcome cohorts}

\item{oracleTempSchema}{The temp oracle schema requires read/write}

\item{verbosity}{Sets the level of the verbosity. If the log level is at or higher in priority than the logger threshold, a message will print. The levels are:
\itemize{
\item{DEBUG}{Highest verbosity showing all debug statements}
\item{TRACE}{Showing information about start and end of steps}
\item{INFO}{Show informative information (Default)}
\item{WARN}{Show warning messages}
\item{ERROR}{Show error messages}
\item{FATAL}{Be silent except for fatal errors}
}}

\item{keepPrediction}{Whether to keep the predicitons for the new data}

\item{sampleSize}{If not NULL, the number of people to sample from the target cohort}
}
\value{
A list containing the performance for each validation_schema
}
\description{
This function extracts data using a user specified connection and cdm_schema, applied the model and then calcualtes the performance
}
\details{
Users need to input a trained model (the output of runPlp()) and new database connections. The function will return a list of length equal to the 
number of cdm_schemas input with the performance on the new data
}
