% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{extractDatabaseToCsv}
\alias{extractDatabaseToCsv}
\title{Exports all the results from a database into csv files}
\usage{
extractDatabaseToCsv(
  connectionDetails,
  databaseSchemaSettings = createDatabaseSchemaSettings(resultSchema = "main"),
  csvFolder
)
}
\arguments{
\item{connectionDetails}{The connectionDetails for the result database}

\item{databaseSchemaSettings}{The result database schema settings}

\item{csvFolder}{Location to save the csv files}
}
\description{
Exports all the results from a database into csv files
}
\details{
Extracts the results from a database into a set of csv files
}
