% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalValidatePlp.R
\name{extractModelDesigns}
\alias{extractModelDesigns}
\title{extractModelDesigns - Extract all modelDesigns from a list of plpModels}
\usage{
extractModelDesigns(plpModelList)
}
\description{
extractModelDesigns - Extract all modelDesigns from a list of plpModels
}
