% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalValidatePlp.R
\name{extractUniqueCombinations}
\alias{extractUniqueCombinations}
\title{extractUniqueCombinations 
create download tasks based on unique combinations of targetId and
restrictPlpDataSettings. This is used to avoid downloading the same data
multiple times.}
\usage{
extractUniqueCombinations(validationDesignList)
}
\arguments{
\item{validationDesignList}{A list of validationDesign objects}
}
\value{
A list of download tasks
}
\description{
extractUniqueCombinations 
create download tasks based on unique combinations of targetId and
restrictPlpDataSettings. This is used to avoid downloading the same data
multiple times.
}
\keyword{internal}
