% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{f1Score}
\alias{f1Score}
\title{Calculate the f1Score}
\usage{
f1Score(TP, TN, FN, FP)
}
\arguments{
\item{TP}{Number of true positives}

\item{TN}{Number of true negatives}

\item{FN}{Number of false negatives}

\item{FP}{Number of false positives}
}
\value{
f1Score value
}
\description{
Calculate the f1Score
}
\details{
Calculate the f1Score
}
