% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{falseDiscoveryRate}
\alias{falseDiscoveryRate}
\title{Calculate the falseDiscoveryRate}
\usage{
falseDiscoveryRate(TP, TN, FN, FP)
}
\arguments{
\item{TP}{Number of true positives}

\item{TN}{Number of true negatives}

\item{FN}{Number of false negatives}

\item{FP}{Number of false positives}
}
\value{
falseDiscoveryRate value
}
\description{
Calculate the falseDiscoveryRate
}
\details{
Calculate the falseDiscoveryRate
}
