% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{falseNegativeRate}
\alias{falseNegativeRate}
\title{Calculate the falseNegativeRate}
\usage{
falseNegativeRate(TP, TN, FN, FP)
}
\arguments{
\item{TP}{Number of true positives}

\item{TN}{Number of true negatives}

\item{FN}{Number of false negatives}

\item{FP}{Number of false positives}
}
\value{
falseNegativeRate  value
}
\description{
Calculate the falseNegativeRate
}
\details{
Calculate the falseNegativeRate
}
