% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{falsePositiveRate}
\alias{falsePositiveRate}
\title{Calculate the falsePositiveRate}
\usage{
falsePositiveRate(TP, TN, FN, FP)
}
\arguments{
\item{TP}{Number of true positives}

\item{TN}{Number of true negatives}

\item{FN}{Number of false negatives}

\item{FP}{Number of false positives}
}
\value{
falsePositiveRate  value
}
\description{
Calculate the falsePositiveRate
}
\details{
Calculate the falsePositiveRate
}
