% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitGLMModel.R
\name{fitGLMModel}
\alias{fitGLMModel}
\title{Fit a predictive model}
\usage{
fitGLMModel(population, plpData, modelType = "logistic",
  excludeCovariateIds = c(), includeCovariateIds = c(),
  prior = createPrior("laplace", useCrossValidation = TRUE),
  control = createControl(cvType = "auto", fold = 3, startingVariance = 0.01,
  tolerance = 2e-06, cvRepetitions = 1, selectorType = "byPid", noiseLevel =
  "silent", threads = -1, maxIterations = 3000))
}
\arguments{
\item{population}{A population object generated by \code{createStudyPopulation}, potentially filtered by other functions.}

\item{plpData}{An object of type \code{plpData} as generated using
\code{getDbPlpData}.}

\item{modelType}{The type of outcome model that will be used. Possible values are
"logistic", "poisson", or "cox".}

\item{excludeCovariateIds}{Exclude these covariates from the outcome model.}

\item{includeCovariateIds}{Include only these covariates in the outcome model.}

\item{prior}{The prior used to fit the model. See
\code{\link[Cyclops]{createPrior}} for details.}

\item{control}{The control object used to control the cross-validation used to
determine the hyperparameters of the prior (if applicable). See
\code{\link[Cyclops]{createControl}} for details.}
}
\description{
Fit a predictive model
}
