% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlmNetModels.R
\name{fitGlmNet}
\alias{fitGlmNet}
\title{fitGlmNet}
\usage{
fitGlmNet(trainData, modelSettings, search = "grid", analysisId)
}
\arguments{
\item{trainData}{The training data received through fitPlp}

\item{modelSettings}{The model settings objects}

\item{search}{Hyperparameter search strategy}

\item{analysisId}{Analysis Id}
}
\description{
fits a glmnet model
}
