% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developModel.R
\name{fitPlp}
\alias{fitPlp}
\title{fitModel}
\usage{
fitPlp(model, param, featureSettings, outcomeId, cohortId, data, loc)
}
\arguments{
\item{model}{A character string specifiying the machine learning model to train.

                                        The potential models are:
                                        \description{
                                        \item{'lr-lasso'}{Logistic regression with lasso regularisation}
                                        \item{'rf'}{Random forest}
                                        \item{'gbm'}{Gradient boosting machines}
                                        \item{'lr-enet'}{Logistic regression with elastic new regularisation}
                                        }}

\item{param}{The parameter setting depend on the model-
model='lr-lasso' has the parameter val specificying the initial variance
model='rf' has the parameters ntree, max_depth and mtry
model='gbm' has the parameters ntree, bal (class balance), nrowSample (fraction of training data people to use per tree),
       ncolSample (fraction of training data features to include into each tree)
model='lr-enet' has the parameters alpha, ...}

\item{featureSettings}{A list containing any parameters for the feature selection/engineering.
\description{
\item{...}{The initial variance setting for the lasso logistic regression}
}}

\item{outcomeId}{The outcomeId the user is aiming to predict}

\item{cohortId}{The id of the cohort being used.  Default is NULL.}

\item{data}{An object of type \code{plpData} - the patient level prediction
data extracted from the CDM.}

\item{loc}{The path to the directory where the model will be saved}
}
\value{
An object of class plpModel
\describe{
\item{model}{The trained prediction model}
\item{modelLoc}{The path to where the model is saved}
\item{trainAuc}{The AUC obtained on the training set}
\item{trainCalibration}{The calibration obtained on the training set}
\item{modelSettings}{A list specifiying the model, preprocessing, outcomeId and cohortId}
\item{features}{The features used by the model}}
}
\description{
Train various models using a default parameter gird search or user specified parameters
}
\details{
The user can define the machine learning model to train (regularised logistic regression, random forest,
gradient boosting machine, neural network and )
}

