% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FitModel.R
\name{fitPredictiveModel}
\alias{fitPredictiveModel}
\title{Fit a predictive model}
\usage{
fitPredictiveModel(cohortsData, covariateData, outcomeData, modelType = "lr",
  cohortConceptId = NULL, outcomeConceptId = NULL,
  prior = createPrior("laplace", exclude = c(0), useCrossValidation = TRUE),
  control = createControl(noiseLevel = "silent", cvType = "auto",
  startingVariance = 0.1))
}
\arguments{
\item{covariateData}{An object of type \code{covariateData}.}

\item{outcomeData}{An object of type \code{outcomeData}.}

\item{modelType}{The type of predictive model}

\item{prior}{The prior used to fit the model. See \code{\link[Cyclops]{createPrior}} for details.}

\item{control}{The control object used to control the cross-validation used to determine the
hyperparameters of the prior (if applicable). See \code{\link[Cyclops]{createControl}}  for details.}
}
\description{
Fit a predictive model
}

