% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalValidatePlp.R
\name{fromDesignOrModel}
\alias{fromDesignOrModel}
\title{fromDesignOrModel - Check if the design has the setting, if not use the model's}
\usage{
fromDesignOrModel(validationDesign, modelDesigns, settingName)
}
\arguments{
\item{validationDesign}{The validationDesign object}

\item{modelDesigns}{A list of modelDesign objects}

\item{settingName}{The name of the setting to check}
}
\value{
The updated design
}
\description{
fromDesignOrModel - Check if the design has the setting, if not use the model's
}
\keyword{internal}
