% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSettings.R
\name{GBMclassifier}
\alias{GBMclassifier}
\title{Create setting for gradient boosting machine model}
\usage{
GBMclassifier(rsampRate = 0.9, csampRate = 1, ntrees = c(10, 100),
  bal = F, nbins = 20, max_depth = 4, min_rows = 2, learn_rate = 0.1)
}
\arguments{
\item{rsampRate}{The fraction of rows to include in each tree during training}

\item{csampRate}{The fraction of features to include in each tree during training}

\item{ntrees}{The number of trees to build}

\item{bal}{Whether to balance the training set classes}

\item{nbins}{Number of bins used in continuous variables?}

\item{max_depth}{Maximum number of interactions - a large value will lead to slow model training}

\item{min_rows}{The minimum number of rows required at each end node of the tree}

\item{learn_rate}{The boosting learn rate}
}
\description{
Create setting for gradient boosting machine model
}
\examples{
model.gbm <- GBMclassifier(rsampRate=c(0.5,0.9,1),csampRate=1, ntrees=c(10,100), bal=c(F,T),
                           max_depth=c(4,5), learn_rate=c(0.1,0.01))

}

