% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{getCalibration}
\alias{getCalibration}
\title{Get a sparse summary of the calibration}
\usage{
getCalibration(prediction, numberOfStrata = 10, truncateFraction = 0.01)
}
\arguments{
\item{prediction}{A prediction object as generated using the
\code{\link{predict}} functions.}

\item{numberOfStrata}{The number of strata in the plot.}

\item{truncateFraction}{This fraction of probability values will be ignored when plotting, to
avoid the x-axis scale being dominated by a few outliers.}
}
\value{
A dataframe with the calibration summary
}
\description{
Get a sparse summary of the calibration
}
\details{
Generates a sparse summary showing the predicted probabilities and the observed fractions. Predictions are
stratefied into equally sized bins of predicted probabilities.
}
