% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadToDatabase.R
\name{getCohortDefinitionJson}
\alias{getCohortDefinitionJson}
\title{Get the cohort definition from a list of definitions}
\usage{
getCohortDefinitionJson(cohortDefinitions, cohortId)
}
\arguments{
\item{cohortDefinitions}{A list of cohortDefinitions}

\item{cohortId}{The cohortId to extract the cohortDefinition for}
}
\value{
Returns a list with the cohort definition R object
}
\description{
This function extracts the cohort definition from a list
}
\details{
This function extracts the cohort definition from a list
}
