% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{getCovariateData}
\alias{getCovariateData}
\title{Get the covaridate data for a cohort table}
\usage{
getCovariateData(connection, cdmDatabaseSchema,
  oracleTempSchema = cdmDatabaseSchema, cohortTable = "#cohort_person",
  cdmVersion = 5, covariateSettings)
}
\arguments{
\item{connection}{Can also use an existing connection rather than the connectionDetails}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you want
all temporary tables to be managed. Requires create/insert
permissions to this database.}

\item{cohortTable}{The temp table containing the cohort of people}

\item{cdmVersion}{The version of the CDM (default 5)}

\item{covariateSettings}{An object of type \code{covariateSettings} as created using the
\code{createCovariateSettings} function in the
\code{FeatureExtraction} package.}
}
\value{
Returns the covariates for the people in the temp table
}
\description{
This function executes some SQL to extract covaraite data for a cohort table
}
\details{

}
