% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalValidatePlp.R
\name{getData}
\alias{getData}
\title{getData - Get the plpData for the validation}
\usage{
getData(design, database, outputFolder, allCovSettings)
}
\description{
getData - Get the plpData for the validation
}
