% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadToDatabase.R
\name{getDatabaseDetail}
\alias{getDatabaseDetail}
\title{Get the database details}
\usage{
getDatabaseDetail(databaseList, databaseSchema)
}
\arguments{
\item{databaseList}{A list created by \code{createDatabaseList} to specify the databases}

\item{databaseSchema}{The database schema string}
}
\value{
Returns a list with the database details
}
\description{
This function extracts the database details from a list of details
}
\details{
This function extracts the database details from a list of details
}
