% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetCohorts.R
\name{getDbCohortData}
\alias{getDbCohortData}
\title{Get cohorts of interest}
\usage{
getDbCohortData(connectionDetails = NULL, connection = NULL,
  cdmDatabaseSchema, oracleTempSchema = cdmDatabaseSchema,
  useExistingCohortPerson = FALSE, cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort", cohortIds = c(0, 1), cdmVersion = "4")
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{connection}{A connection to the server containing the schema as created using the
\code{connect} function in the \code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM instance.
Requires read permissions to this database. On SQL Server, this should
specifiy both the database and the schema, so for example
'cdm_instance.dbo'.}

\item{oracleTempSchema}{A schema where temp tables can be created in Oracle.#'}

\item{useExistingCohortPerson}{Does the temporary table \code{cohort_person}
already exists? Can only be used when the \code{connection} parameter
is not NULL.}

\item{cohortDatabaseSchema}{If not using an existing \code{cohort_person} temp table, where is the
source cohort table located? Note that on SQL Server, one should
include both the database and schema, e.g. "cdm_schema.dbo".}

\item{cohortTable}{If not using an existing temp table, what is the name of the table
holding the cohort?}

\item{cohortIds}{If not using an existing temp table, what is the name of the source
cohort table?}

\item{cdmVersion}{Define the OMOP CDM version used:  currently support "4" and "5".}
}
\value{
An object of type \code{cohortData} containing information on who are in the cohorts.
}
\description{
Gets the cohorts of interest from the database
}

