% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetCovariates.R
\name{getDbCovariateData}
\alias{getDbCovariateData}
\title{Get covariate information from the database}
\usage{
getDbCovariateData(connectionDetails = NULL, connection = NULL,
  oracleTempSchema = NULL, cdmDatabaseSchema,
  useExistingCohortPerson = FALSE, cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort", cohortIds = c(0, 1), covariateSettings,
  cdmVersion = "4")
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{connection}{A connection to the server containing the schema as created using
the \code{connect} function in the \code{DatabaseConnector}
package.}

\item{oracleTempSchema}{A schema where temp tables can be created in Oracle.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema, so
for example 'cdm_instance.dbo'.}

\item{useExistingCohortPerson}{Does the temporary table \code{cohort_person} already exists? Can
only be used when the \code{connection} parameter is not NULL.}

\item{cohortDatabaseSchema}{If not using an existing \code{cohort_person} temp table, where is
the source cohort table located? Note that on SQL Server, one
should include both the database and schema, e.g. 'cdm_schema.dbo'.}

\item{cohortTable}{If not using an existing \code{cohort_person} temp table, what is
the name of the source cohort table?}

\item{cohortIds}{The IDs of the cohortsin the cohort table for which we want to
build covariates.}

\item{covariateSettings}{An object of type \code{covariateSettings} as created using the
\code{\link{createCovariateSettings}} function.}

\item{cdmVersion}{Define the OMOP CDM version used: currently support "4" and "5".}
}
\value{
Returns an object of type \code{covariateData}, containing information on the baseline covariates.
Information about multiple outcomes can be captured at once for efficiency reasons. This object is
a list with the following components: \describe{ \item{covariates}{An ffdf object listing the
baseline covariates per person in the two cohorts. This is done using a sparse representation:
covariates with a value of 0 are omitted to save space.} \item{covariateRef}{An ffdf object
describing the covariates that have been extracted.} \item{metaData}{A list of objects with
information on how the covariateData object was constructed.} }
}
\description{
Constructs a large set of covariates for one or more cohorts using data in the CDM schema.
}
\details{
This function uses the data in the CDM to construct a large set of covariates for the provided
cohorts. The cohorts are assumed to be in a table with the same structure as the cohort table in
the OMOP CDM. The subject_id in this table must refer to person_ids in the CDM. One person can
occurr multiple times, but the combination of subject_id and cohort_start_date is assumed to be
unique.
}

