% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetOutcomes.R
\name{getDbOutcomeData}
\alias{getDbOutcomeData}
\title{Get outcomes for persons in the cohort}
\usage{
getDbOutcomeData(connectionDetails = NULL, connection = NULL,
  cdmDatabaseSchema, oracleTempSchema = cdmDatabaseSchema,
  useExistingCohortPerson = FALSE, cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort", cohortConceptIds = c(0, 1),
  outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "condition_occurrence", outcomeConceptIds = c(),
  outcomeConditionTypeConceptIds = "", firstOutcomeOnly = FALSE)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using
the function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{connection}{A connection to the server containing the schema as created
using the \code{connect} function in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On
SQL Server, this should specifiy both the database and the
schema, so for example 'cdm_instance.dbo'.}

\item{oracleTempSchema}{A schema where temp tables can be created in Oracle.#'}

\item{useExistingCohortPerson}{Does the temporary table \code{cohort_person} already
exists? Can only be used when the \code{connection}
parameter is not NULL.}

\item{cohortDatabaseSchema}{If not using an existing \code{cohort_person} temp table,
where is the source cohort table located? Note that on SQL
Server, one should include both the database and schema,
e.g. "cdm_schema.dbo".}

\item{cohortTable}{If not using an existing temp table, what is the name of the
table holding the cohort?}

\item{cohortConceptIds}{If not using an existing temp table, what is the name of the
source cohort table?}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where
the data used to define the outcome cohorts is available. If
exposureTable = CONDITION_ERA, exposureDatabaseSchema is not
used by assumed to be cdmSchema.  Requires read permissions
to this database.}

\item{outcomeTable}{The tablename that contains the outcome cohorts.  If
outcomeTable <> CONDITION_OCCURRENCE, then expectation is
outcomeTable has format of COHORT table: COHORT_CONCEPT_ID,
SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE.}

\item{outcomeConceptIds}{A list of CONCEPT_IDs used to define outcomes.  If
outcomeTable = CONDITION_OCCURRENCE, the list is a set of
ancestor CONCEPT_IDs, and all occurrences of all descendant
concepts will be selected.  If outcomeTable <>
CONDITION_OCCURRENCE, the list contains records found in
COHORT_DEFINITION_ID field.}

\item{outcomeConditionTypeConceptIds}{A list of TYPE_CONCEPT_ID values that will restrict
condition occurrences.  Only applicable if outcomeTable =
CONDITION_OCCURRENCE.}

\item{firstOutcomeOnly}{Only keep the first outcome per person?}
}
\value{
An object of type \code{outcomeData} containing information on the outcomes in the cohort(s).
}
\description{
Gets the outcomes for the specified cohort(s).
}
\details{
For the specified cohorts, retrieve the outcomes of interest during cohort start and end date.
Either a \code{connectionDetails} or a \code{connection} object has to be specified.
}

