% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetPlpData.R
\name{getDbPlpData}
\alias{getDbPlpData}
\title{Get outcomes for persons in the cohort}
\usage{
getDbPlpData(connectionDetails = NULL, cdmDatabaseSchema,
  oracleTempSchema = NULL, cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort", cohortIds = c(0, 1), washoutWindow = 183,
  useCohortEndDate = TRUE, windowPersistence = 0, covariateSettings,
  outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "condition_occurrence", outcomeIds = c(),
  outcomeConditionTypeConceptIds = "", firstOutcomeOnly = FALSE,
  cdmVersion = "4")
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using
the function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On
SQL Server, this should specifiy both the database and the
schema, so for example 'cdm_instance.dbo'.}

\item{oracleTempSchema}{A schema where temp tables can be created in Oracle.}

\item{cohortDatabaseSchema}{Where is the source cohort table located? Note that on SQL
Server, one should include both the database and schema,
e.g. "cdm_schema.dbo".}

\item{cohortTable}{What is the name of the table holding the cohort?}

\item{cohortIds}{The IDs of the cohorts for which we want to create models.}

\item{washoutWindow}{The mininum required continuous observation time prior to
index date for a person to be included in the cohort.}

\item{useCohortEndDate}{Use the cohort end date as the basis for the end of the risk
window? If FALSE, the cohort start date will be used
instead.}

\item{windowPersistence}{The number of days the risk window should persist.}

\item{covariateSettings}{An object of type \code{covariateSettings} as created using
the \code{\link{createCovariateSettings}} function.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where
the data used to define the outcome cohorts is available. If
outcomeTable = CONDITION_ERA, outcomeDatabaseSchema is not
used.  Requires read permissions to this database.}

\item{outcomeTable}{The tablename that contains the outcome cohorts.  If
outcomeTable <> CONDITION_OCCURRENCE, then expectation is
outcomeTable has format of COHORT table: COHORT_CONCEPT_ID,
SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE.}

\item{outcomeIds}{A list of ids used to define outcomes.  If outcomeTable =
CONDITION_OCCURRENCE, the list is a set of ancestor
CONCEPT_IDs, and all occurrences of all descendant concepts
will be selected.  If outcomeTable <> CONDITION_OCCURRENCE,
the list contains records found in COHORT_DEFINITION_ID
field.}

\item{outcomeConditionTypeConceptIds}{A list of TYPE_CONCEPT_ID values that will restrict
condition occurrences.  Only applicable if outcomeTable =
CONDITION_OCCURRENCE.}

\item{firstOutcomeOnly}{Only keep the first outcome per person?}

\item{cdmVersion}{Define the OMOP CDM version used: currently support "4" and
                                        "5".}
}
\value{
An object of type \code{plpData} containing information on the prediction problem. This object will
contain the following data:
\describe{ \item{cohorts}{An ffdf object listing all persons and their prediction periods. This
object will have these fields: row_id (a unique ID per period), person_id, cohort_start_date,
cohort_id, time (number of days in the window).} \item{outcomes}{An ffdf object listing all
outcomes per period. This object will have these fields: row_id, outcome_id, outcome_count,
time_to_event.} \item{exclude}{Either NULL or an ffdf object listing per outcome ID which windows
had the outcome prior to the window. This object will have these fields: rowId, outcomeId.}
\item{covariates}{An ffdf object listing the baseline covariates per person in the cohorts. This is
done using a sparse representation: covariates with a value of 0 are omitted to save space. The
covariates object will have three columns: rowId, covariateId, and covariateValue. }
\item{covariateRef}{An ffdf object describing the covariates that have been extracted.}
\item{metaData}{A list of objects with information on how the plpData object was constructed.} }
}
\description{
Get all the data for the prediction problem from the server.
}
\details{
For the specified cohorts, retrieve the outcomes of interest and covariates to be used for the
prediction problem.
}

