% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{getNetBenefit}
\alias{getNetBenefit}
\title{Calculate net benefit in an efficient way}
\usage{
getNetBenefit(plpResult, evalType)
}
\arguments{
\item{plpResult}{A plp result object as generated using the \code{\link{runPlp}} function.}

\item{evalType}{The evaluation type column}
}
\value{
A data frame with the net benefit, treat all and treat none values
}
\description{
Calculate net benefit in an efficient way
}
\keyword{internal}
