% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalValidatePlp.R
\name{getPopulation}
\alias{getPopulation}
\title{getPopulation - Get the population for the validationDesign}
\usage{
getPopulation(validationDesign, modelDesigns, plpData)
}
\arguments{
\item{validationDesign}{The validationDesign objects}

\item{modelDesigns}{A list of modelDesign objects}

\item{plpData}{The plpData object}
}
\value{
The population dataframe
}
\description{
getPopulation - Get the population for the validationDesign
}
\keyword{internal}
