% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionDistribution.R
\name{getPredictionDistribution_binary}
\alias{getPredictionDistribution_binary}
\title{Calculates the prediction distribution}
\usage{
getPredictionDistribution_binary(prediction, evalColumn, ...)
}
\arguments{
\item{prediction}{A prediction object}

\item{evalColumn}{A column that is used to stratify the results}

\item{...}{Other inputs}
}
\value{
The 0.00, 0.1, 0.25, 0.5, 0.75, 0.9, 1.00 quantile pf the prediction,
the mean and standard deviation per class
}
\description{
Calculates the prediction distribution
}
\details{
Calculates the quantiles from a predition object
}
