% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThresholdSummary.R
\name{getThresholdSummary_binary}
\alias{getThresholdSummary_binary}
\title{Calculate all measures for sparse ROC when prediction is bianry classification}
\usage{
getThresholdSummary_binary(prediction, evalColumn, ...)
}
\arguments{
\item{prediction}{A prediction object}

\item{evalColumn}{A column that is used to stratify the results}

\item{...}{Other inputs}
}
\value{
A data.frame with all the measures
}
\description{
Calculate all measures for sparse ROC when prediction is bianry classification
}
\details{
Calculates the TP, FP, TN, FN, TPR, FPR, accuracy, PPF, FOR and Fmeasure
from a prediction object
}
