% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePlp.R
\name{getTPR}
\alias{getTPR}
\title{getTPR}
\usage{
getTPR(roc, FPR = 0.05)
}
\arguments{
\item{roc}{A dataframe containing TPR and FPR at range of thresholds}
}
\value{
TPR at specified FPR
}
\description{
Extracts TPR for specified FPR
}
\details{

}

