% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSettings.R
\name{GLMclassifier}
\alias{GLMclassifier}
\title{Create setting for generalised linear model with elastic-new regularisation}
\usage{
GLMclassifier(alpha = 0.5, lambda = 1e-06, lambda_search = T,
  nlambdas = 100, lambda_min_ratio = 1/10000)
}
\arguments{
\item{alpha}{...}

\item{lambda}{...}

\item{lambda_search}{...}

\item{nlambdas}{...}

\item{lambda_min_ratio}{...}
}
\description{
Create setting for generalised linear model with elastic-new regularisation
}
\examples{
model.glm <- GLMclassifier(alpha=c(0.5,0.1,0.9))
}

