% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradientBoostingMachine.R
\name{gradientBoostingMachine.set}
\alias{gradientBoostingMachine.set}
\title{Create setting for gradient boosting machine model using gbm_xgboost implementation}
\usage{
gradientBoostingMachine.set(ntrees = c(10, 100), nthread = 20,
  max_depth = 6, min_rows = 20, learn_rate = 0.1)
}
\arguments{
\item{ntrees}{The number of trees to build}

\item{nthread}{The number of computer threads to (how many cores do you have?)}

\item{max_depth}{Maximum number of interactions - a large value will lead to slow model training}

\item{min_rows}{The minimum number of rows required at each end node of the tree}

\item{learn_rate}{The boosting learn rate}
}
\description{
Create setting for gradient boosting machine model using gbm_xgboost implementation
}
\examples{
model.gbm <- gradientBoostingMachine.set(ntrees=c(10,100), nthread=20,
                           max_depth=c(4,6), learn_rate=c(0.1,0.3))

}

