% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{grepCovariateNames}
\alias{grepCovariateNames}
\title{Extract covariate names}
\usage{
grepCovariateNames(pattern, object)
}
\arguments{
\item{pattern}{A regular expression with which to name covariate names}

\item{object}{An R object of type \code{plpData} or \code{covariateData}.}
}
\value{
Returns a \code{data.frame} containing information about covariates that match a regular
expression.  This \code{data.frame} has the following columns: \describe{
\item{covariateId}{Numerical identifier for use in model fitting using these covariates}
\item{covariateName}{Text identifier} \item{analysisId}{Analysis identifier} \item{conceptId}{OMOP
common data model concept identifier, or 0} }
}
\description{
Extracts covariate names using a regular-expression.
}
\details{
This function extracts covariate names that match a regular-expression for a
\code{plpData} or \code{covariateData} object.
}
