% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFromCsv.R
\name{insertCsvToDatabase}
\alias{insertCsvToDatabase}
\title{Function to insert results into a database from csvs}
\usage{
insertCsvToDatabase(
  csvFolder,
  connectionDetails,
  databaseSchemaSettings,
  modelSaveLocation,
  csvTableAppend = ""
)
}
\arguments{
\item{csvFolder}{The location to the csv folder with the plp results}

\item{connectionDetails}{A connection details for the plp results database that the csv results will be inserted into}

\item{databaseSchemaSettings}{A object created by \code{createDatabaseSchemaSettings} with all the settings specifying the result tables to insert the csv results into}

\item{modelSaveLocation}{The location to save any models from the csv folder - this should be the same location you picked when inserting other models into the database}

\item{csvTableAppend}{A string that appends the csv file names}
}
\value{
Returns a data.frame indicating whether the results were inported into the database
}
\description{
This function converts a folder with csv results into plp objects and loads 
them into a plp result database
}
\details{
The user needs to have plp csv results in a single folder and an existing plp 
result database
}
