% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{insertDbPopulation}
\alias{insertDbPopulation}
\title{Insert a population into a database}
\usage{
insertDbPopulation(population, cohortIds = 1, connectionDetails,
  cohortDatabaseSchema, cohortTable = "cohort", createTable = FALSE,
  dropTableIfExists = TRUE, cdmVersion = "5")
}
\arguments{
\item{population}{Either an object of type \code{plpData} or a population object generated by functions
like \code{createStudyPopulation}.}

\item{cohortIds}{The IDs to be used for the treated and comparator cohort, respectively.}

\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cohortDatabaseSchema}{The name of the database schema where the data will be written.
Requires write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{cohortTable}{The name of the table in the database schema  where the data will be written.}

\item{createTable}{Should a new table be created? If not, the data will be inserted into an existing
table.}

\item{dropTableIfExists}{If \code{createTable = TRUE} and the table already exists it will be overwritten.}

\item{cdmVersion}{Define the OMOP CDM version used: currently support "4" and "5".}
}
\description{
Insert a population into a database
}
\details{
Inserts a population table into a database. The table in the database will have the same structure as the
'cohort' table in the Common Data Model.
}
