% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadToDatabaseModelDesign.R
\name{insertModelDesignInDatabase}
\alias{insertModelDesignInDatabase}
\title{Insert a model design into a PLP result schema database}
\usage{
insertModelDesignInDatabase(
  object,
  conn,
  databaseSchemaSettings,
  cohortDefinitions
)
}
\arguments{
\item{object}{An object of class modelDesign, runPlp or externalValidatePlp}

\item{conn}{A connection to a database created by using the
function \code{connect} in the
\code{DatabaseConnector} package.}

\item{databaseSchemaSettings}{A object created by \code{createDatabaseSchemaSettings} with all the settings specifying the result tables}

\item{cohortDefinitions}{(list) A list of cohortDefinitions (each list must contain: name, id)}
}
\value{
Returns NULL but uploads the model design into the database schema specified in databaseSchemaSettings
}
\description{
This function inserts a model design and all the settings into the result schema
}
\details{
This function can be used to upload a model design into a database
}
