% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadToDatabase.R
\name{insertResultsToSqlite}
\alias{insertResultsToSqlite}
\title{Create sqlite database with the results}
\usage{
insertResultsToSqlite(
  resultLocation,
  cohortDefinitions,
  databaseList = createDatabaseList(cdmDatabaseSchemas = c("cdm_truven_ccae_v123",
    "cdm_madeup_v1")),
  sqliteLocation = file.path(resultLocation, "sqlite")
)
}
\arguments{
\item{resultLocation}{(string) location of directory where the main package results were saved}

\item{cohortDefinitions}{(list) A list of cohortDefinitions (each list must contain: name, id)}

\item{databaseList}{A list created by \code{createDatabaseList} to specify the databases}

\item{sqliteLocation}{(string) location of directory where the sqlite database will be saved}
}
\value{
Returns the location of the sqlite database file
}
\description{
This function create an sqlite database with the PLP result schema and inserts all results
}
\details{
This function can be used upload PatientLevelPrediction results into an sqlite database
}
