% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{interpretInstallCode}
\alias{interpretInstallCode}
\title{Tells you the package issue}
\usage{
interpretInstallCode(response)
}
\arguments{
\item{response}{The response code from checkPlpInstallation()}
}
\description{
Tells you the package issue
}
\details{
This function prints any issues found during the checkPlpInstallation() call
}
