% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.R
\name{KNN.set}
\alias{KNN.set}
\title{Create setting for knn model}
\usage{
KNN.set(k = 1000, indexFolder = getwd())
}
\arguments{
\item{k}{The number of neighbors to consider}

\item{indexFolder}{The directory where the results and intermediate steps are output}
}
\description{
Create setting for knn model
}
\examples{
model.knn <- KNN.set(k=c(3,100,1000))
}

