% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSettings.R
\name{KNNclassifier}
\alias{KNNclassifier}
\title{Create setting for knn model}
\usage{
KNNclassifier(k = 1000, indexFolder = getwd())
}
\arguments{
\item{k}{The number of neighbors to consider}

\item{indexFolder}{The directory where the results and intermediate steps are output}
}
\description{
Create setting for knn model
}
\examples{
model.knn <- KNNclassifier(k=c(3,100,1000))
}

