% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassoLogisticRegression.R
\name{lassoLogisticRegression.set}
\alias{lassoLogisticRegression.set}
\title{Create setting for lasso logistic regression}
\usage{
lassoLogisticRegression.set(variance = 0.01)
}
\arguments{
\item{variance}{a single value or vector of values to be used to train multiple models and the model with the
best performance on the cross validation set is choosen}
}
\description{
Create setting for lasso logistic regression
}
\examples{
model.lr <- logisticRegressionModel()
}

