% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SklearnClassifierHelpers.R
\name{listCartesian}
\alias{listCartesian}
\title{Cartesian product}
\usage{
listCartesian(allList)
}
\arguments{
\item{allList}{a list of lists}
}
\value{
A list with all possible combinations from the input list of lists
}
\description{
Computes the Cartesian product of all the combinations of elements in a list
}
\examples{
listCartesian(list(list(1, 2), list(3, 4))
}
