% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetCohorts.R
\name{loadCohortData}
\alias{loadCohortData}
\title{Load the cohorts data from a folder}
\usage{
loadCohortData(file, readOnly = FALSE)
}
\arguments{
\item{file}{The name of the folder containing the data.}

\item{readOnly}{If true, the data is opened read only.}
}
\value{
An object of class cohortData
}
\description{
\code{loadCohortData} loads an object of type cohortData from a folder in the file system.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}

