% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetCohorts.R
\name{loadCohortsData}
\alias{loadCohortsData}
\title{Load the cohorts data from a folder}
\usage{
loadCohortsData(file, readOnly = FALSE)
}
\arguments{
\item{file}{The name of the folder containing the data.}

\item{readOnly}{If true, the data is opened read only.}
}
\value{
An object of class cohortsData
}
\description{
\code{loadCohortsData} loads an object of type cohortsData from a folder in the file system.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}
\examples{
#todo
}

