% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensemble.R
\name{loadEnsemblePlpModel}
\alias{loadEnsemblePlpModel}
\title{loads the Ensmeble plp model and return a model list}
\usage{
loadEnsemblePlpModel(dirPath)
}
\arguments{
\item{dirPath}{The location of the model}
}
\description{
loads the Ensmeble plp model and return a model list
}
\details{
Loads a plp model list that was saved using \code{savePlpModel()}
}
