% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensemble.R
\name{loadEnsemblePlpResult}
\alias{loadEnsemblePlpResult}
\title{loads the Ensemble plp results}
\usage{
loadEnsemblePlpResult(dirPath)
}
\arguments{
\item{dirPath}{The location of the model}
}
\description{
loads the Ensemble plp results
}
\details{
Loads a plp model list that was saved using \code{saveEnsemblePlpResults()}
}
