% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{loadEvaluation}
\alias{loadEvaluation}
\title{Loads the evalaution dataframe}
\usage{
loadEvaluation(location)
}
\arguments{
\item{location}{The directory where the evaluation was saved}
}
\description{
Loads the evalaution dataframe
}
\details{
Loads the evaluation
}

