% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{loadPlpFromCsv}
\alias{loadPlpFromCsv}
\title{Loads parts of the plp result saved as csv files for transparent sharing}
\usage{
loadPlpFromCsv(dirPath)
}
\arguments{
\item{dirPath}{The directory with the results as csv files}
}
\description{
Loads parts of the plp result saved as csv files for transparent sharing
}
\details{
Load the main results from csv files into a runPlp object
}
