% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMultiplePlp.R
\name{loadPredictionAnalysisList}
\alias{loadPredictionAnalysisList}
\title{Load the multiple prediction json settings from a file}
\usage{
loadPredictionAnalysisList(predictionAnalysisListFile)
}
\arguments{
\item{predictionAnalysisListFile}{The prediciton specification json extracted from atlas.}
}
\description{
Load the multiple prediction json settings from a file
}
\details{
This function interprets a json with the multiple prediction settings and creates a list 
that can be combined with connection settings to run a multiple prediction study
}
\examples{
\dontrun{
predictionAnalysisList <- loadPredictionAnalysisList('./predictionStudyAnalyses.json')
predictionAnalysisList$connectionDetails = connectionDetails
predictionAnalysisList$cdmDatabaseSchema = cdmDatabaseSchema
predictionAnalysisList$cdmDatabaseName = cdmDatabaseName
predictionAnalysisList$oracleTempSchema = oracleTempSchema
predictionAnalysisList$cohortDatabaseSchema = cohortDatabaseSchema
predictionAnalysisList$cohortTable = cohortTable
predictionAnalysisList$outcomeDatabaseSchema = outcomeDatabaseSchema
predictionAnalysisList$outcomeTable = outcomeTable
predictionAnalysisList$cdmVersion = cdmVersion
predictionAnalysisList$outputFolder = outputFolder
result <- do.call(runPlpAnalyses, predictionAnalysisList)
}

}
