% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadData.R
\name{loadTrainTestData}
\alias{loadTrainTestData}
\title{Load the cohort data from a folder}
\usage{
loadTrainTestData(file, readOnly = TRUE)
}
\arguments{
\item{file}{The name of the folder containing the data.}

\item{readOnly}{If true, the data is opened read only.}
}
\value{
Data object before runModelDevelopment.
}
\description{
\code{loadTrainTestData} loads an data object before runModelDevelopment from a folder in the file.
system.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}
\examples{
# todo

}
