% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developModel.R
\name{makeRandomString}
\alias{makeRandomString}
\title{makeRandomString}
\usage{
makeRandomString(n = 1, lenght = 12)
}
\arguments{
\item{n}{An integer - the number of random string to generate}

\item{length}{An integer - the number of characters for each string}
}
\value{
A list containing n random strings with the number of characters specified by the use input length
}
\description{
A function for making a random string
}
\details{
The function creates n random strings of size length
}

