% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatting.R
\name{MapIds}
\alias{MapIds}
\title{Map covariate and row Ids so they start from 1}
\usage{
MapIds(covariateData, cohort = NULL, mapping = NULL)
}
\arguments{
\item{covariateData}{a covariateData object}

\item{cohort}{if specified rowIds restricted to the ones in cohort}

\item{mapping}{A pre defined mapping to use}
}
\description{
this functions takes covariate data and a cohort/population and remaps
the covariate and row ids, restricts to pop and saves/creates mapping
}
