% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering.R
\name{minMaxNormalize}
\alias{minMaxNormalize}
\title{A function that normalizes continous features to have values between 0 and 1}
\usage{
minMaxNormalize(trainData, featureEngineeringSettings, done = FALSE)
}
\arguments{
\item{trainData}{The training data to be normalized}

\item{featureEngineeringSettings}{The settings for the normalization}

\item{done}{Whether the data has already been normalized (bool)}
}
\value{
The normalized data
}
\description{
A function that normalizes continous features to have values between 0 and 1
}
\details{
uses value - min / (max - min) to normalize the data
}
\keyword{internal}
