% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveBayes.R
\name{naiveBayes.set}
\alias{naiveBayes.set}
\title{Create setting for naive bayes model with python}
\usage{
naiveBayes.set()
}
\description{
Create setting for naive bayes model with python
}
\examples{
\dontrun{
model.nb <- naiveBayes.set()
}
}

