% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSettings.R
\name{NBclassifier_python}
\alias{NBclassifier_python}
\title{Create setting for naive bayes model with python}
\usage{
NBclassifier_python()
}
\description{
Create setting for naive bayes model with python
}
\examples{
model.nb <- NBclassifie_pythonr()
}

