% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{negativeLikelihoodRatio}
\alias{negativeLikelihoodRatio}
\title{Calculate the negativeLikelihoodRatio}
\usage{
negativeLikelihoodRatio(TP, TN, FN, FP)
}
\arguments{
\item{TP}{Number of true positives}

\item{TN}{Number of true negatives}

\item{FN}{Number of false negatives}

\item{FP}{Number of false positives}
}
\value{
negativeLikelihoodRatio value
}
\description{
Calculate the negativeLikelihoodRatio
}
\details{
Calculate the negativeLikelihoodRatio
}
