% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{negativePredictiveValue}
\alias{negativePredictiveValue}
\title{Calculate the negativePredictiveValue}
\usage{
negativePredictiveValue(TP, TN, FN, FP)
}
\arguments{
\item{TP}{Number of true positives}

\item{TN}{Number of true negatives}

\item{FN}{Number of false negatives}

\item{FP}{Number of false positives}
}
\value{
negativePredictiveValue value
}
\description{
Calculate the negativePredictiveValue
}
\details{
Calculate the negativePredictiveValue
}
