% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newDataEval.R
\name{newDataEval}
\alias{newDataEval}
\title{Make predictions using model on new data}
\usage{
newDataEval(newData, plpModel)
}
\arguments{
\item{newData}{An object of type \code{plpData}.}

\item{plpModel}{A model returned by calling developModel2()}
}
\description{
Make predictions using model on new data
}
\details{
Computes the performance of the model on new data
}

