% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NJMIM.R
\name{njmimSettings}
\alias{njmimSettings}
\title{Create settings using normalized joint mutual information maximization (NJMIM) for feature selection}
\usage{
njmimSettings(k = 20)
}
\arguments{
\item{k}{number of features to select}
}
\description{
Create settings using normalized joint mutual information maximization (NJMIM) for feature selection
}
