% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetCovariates.R
\name{normalizeCovariates}
\alias{normalizeCovariates}
\title{Normalize covariate values}
\usage{
normalizeCovariates(covariates)
}
\arguments{
\item{covariates}{An ffdf object as generated using the \code{\link{getDbCovariateData}}
                    function.#'}
}
\description{
Normalize covariate values
}
\details{
Normalize covariate values by dividing by the max. This is to avoid numeric problems when fitting
models.
}

