% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagePlp.R
\name{packageResults}
\alias{packageResults}
\title{Package the results for sharing with OHDSI researchers}
\usage{
packageResults(mainFolder, includeROCplot = T, includeCalibrationPlot = T,
  includePRPlot = T, includeTable1 = T, includeThresholdSummary = T,
  includeDemographicSummary = T, includeCalibrationSummary = T,
  includePredictionDistribution = T, includeCovariateSummary = T,
  removeLessThanN = T, N = 5)
}
\arguments{
\item{mainFolder}{The location of the folder with the standard output}

\item{includeROCplot}{Whether to include ROC plot}

\item{includeCalibrationPlot}{Whether to include calibration plot}

\item{includePRPlot}{Whether to include precision recall plot}

\item{includeTable1}{Whether to include table1}

\item{includeThresholdSummary}{Whether to include thresholdSummary evaluation}

\item{includeDemographicSummary}{Whether to include demographicSummary evaluation}

\item{includeCalibrationSummary}{Whether to include calibrationSummary evaluation}

\item{includePredictionDistribution}{Whether to include predictionDistribution evaluation}

\item{includeCovariateSummary}{Whether to include covariateSummary evaluation}

\item{removeLessThanN}{Whether to remove any entry with less than N people}

\item{N}{If removeLessThanN is TRUE the value for N}
}
\description{
Package the results for sharing with OHDSI researchers
}
\details{
This function packages the results.
}
