% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developModel.R
\name{paramSettings}
\alias{paramSettings}
\title{paramSettings}
\usage{
paramSettings(model = "gbm", bal = F, ntrees = 100, rsampRate = 1,
  csampRate = 1, learn_rate = 0.01, nbins = 10, max_depth = 4,
  min_rows = 20, mtries = -1, alpha = 0.5, lambda = 1e-06,
  lambda_search = T, nlambdas = 50, lambda_min_ratio = 0.001)
}
\arguments{
\item{bal}{A vector of boolean values - specifying whether to balance the class labels during training}

\item{ntrees}{A vector of integers -specifying the number of trees to train}

\item{rsampRate}{A vector of values between 0 and 1 specifying the fraction of rows to use for each tree}

\item{csampRate}{A vector of values between 0 and 1 specifying the fraction of features to use for each tree}
}
\value{
A list the parameters expanded out like a grid ready to be investigated during model training
}
\description{
A function for specifying the hyperparameters to create a grid search when training the GBM model
}
\details{
The function takes a list of the model's hyperparameters and values to investigate while training the model
}

