% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PatientLevelPrediction.R
\docType{package}
\name{PatientLevelPrediction}
\alias{PatientLevelPrediction}
\alias{PatientLevelPrediction-package}
\title{PatientLevelPrediction}
\description{
PatientLevelPrediction
}

