% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CovariatePrevalence.R
\name{plotCovariateDifferenceOfTopVariables}
\alias{plotCovariateDifferenceOfTopVariables}
\title{Plot variables with largest standardized difference}
\usage{
plotCovariateDifferenceOfTopVariables(means, n = 20, maxNameWidth = 100,
  fileName = NULL)
}
\arguments{
\item{means}{A data frame created by the \code{computeCovariateMeans} funcion.}

\item{n}{Count of variates to plot.}

\item{maxNameWidth}{Covariate names longer than this number of characters are truncated to create
a nicer plot.}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'. See
                      the function \code{ggsave} in the ggplot2 package for supported file formats.}
}
\value{
A ggplot object. Use the \code{\link[ggplot2]{ggsave}} function to save to file in a different
format.
}
\description{
Create a plot showing those variables having the largest standardized difference between the group
having the outcome and the group that doesn't have the outcome. Requires running
\code{computeCovariateMeans} first.
}

