% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plotPredictedPDF}
\alias{plotPredictedPDF}
\title{Plot the Predicted probability density function, showing prediction overlap between true and false cases}
\usage{
plotPredictedPDF(
  plpResult,
  typeColumn = "evaluation",
  saveLocation = NULL,
  fileName = "PredictedPDF.png"
)
}
\arguments{
\item{plpResult}{A plp result object as generated using the \code{\link{runPlp}} function.}

\item{typeColumn}{The name of the column specifying the evaluation type}

\item{saveLocation}{Directory to save plot (if NULL plot is not saved)}

\item{fileName}{Name of the file to save to plot, for example
'plot.png'. See the function \code{ggsave} in the ggplot2 package for
supported file formats.}
}
\value{
A ggplot object. Use the \code{\link[ggplot2]{ggsave}} function to save to file in a different
format.
}
\description{
Plot the Predicted probability density function, showing prediction overlap between true and false cases
}
\details{
Create a plot showing the predicted probability density function, showing prediction overlap between true and false cases
}
