% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plotVariableScatterplot}
\alias{plotVariableScatterplot}
\title{Plot the variable importance scatterplot}
\usage{
plotVariableScatterplot(covariateSummary, fileName = NULL)
}
\arguments{
\item{covariateSummary}{A prediction object as generated using the
\code{\link{runPlp}} function.}

\item{fileName}{Name of the file where the plot should be saved, for example
'plot.png'. See the function \code{ggsave} in the ggplot2 package for
supported file formats.}
}
\value{
A ggplot object. Use the \code{\link[ggplot2]{ggsave}} function to save to file in a different
format.
}
\description{
Plot the variable importance scatterplot
}
\details{
Create a plot showing the variable importance scatterplot
#'
}
