% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatientLevelPrediction.R
\docType{data}
\name{plpDataSimulationProfile}
\alias{plpDataSimulationProfile}
\title{A simulation profile}
\format{A data frame containing the following elements:
\describe{
  \item{covariatePrevalence}{prevalence of all covariates}
  \item{outcomeModels}{regression model parameters to simulate outcomes}
  \item{metaData}{settings used to simulate the profile}
  \item{covariateRef}{covariateIds and covariateNames}
  \item{timePrevalence}{time window}
  \item{exclusionPrevalence}{prevalence of exclusion of covariates}
}}
\usage{
data(plpDataSimulationProfile)
}
\description{
A simulation profile
}
\keyword{datasets}
