% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{pmmFit}
\alias{pmmFit}
\title{Predictive mean matching using lasso}
\usage{
pmmFit(data, k = 5)
}
\arguments{
\item{k}{The number of donors to use for matching (default 5)}

\item{numericData}{An andromeda object with the following fields:
xObs: covariates table for observed data
xMiss: covariates table for missing data
yObs: outcome variable that we want to impute}
}
\description{
Predictive mean matching using lasso
}
\keyword{internal}
