% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadPlpDbResults.R
\name{populatePlpResultTables}
\alias{populatePlpResultTables}
\title{Populate the PatientLevelPrediction results tables}
\usage{
populatePlpResultTables(
  conn,
  resultSchema,
  stringAppendToTables = "",
  targetDialect = "postgresql",
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  packageName,
  studyJsonList,
  studyName = "",
  studyDescription = "",
  researcherName = "",
  researcherEmail = "",
  researcherOrg = "",
  databaseName = NULL,
  databaseAcronym = NULL,
  databaseVersion = 1,
  databaseDescription = NULL,
  databaseType = NULL,
  valDatabases = list(ccae = list(name = "CCAE", description = "", version = 1, type =
    "US Claims")),
  resultLocation = NULL,
  resultPattern = "",
  validationLocation = file.path(resultLocation, "Validation"),
  addInternalValidation = T,
  addExternalValidation = T,
  gsubVal = NULL,
  removePattern = NULL
)
}
\arguments{
\item{conn}{A connection to a database created by using the
function \code{connect} in the
\code{DatabaseConnector} package.}

\item{resultSchema}{(string) The name of the database schema that the result tables will be created.}

\item{stringAppendToTables}{(string) A string that appends to the PatientLevelPrediction result tables}

\item{targetDialect}{(string) The database management system being used}

\item{tempEmulationSchema}{(string) The temp schema used when the database management system is oracle}

\item{packageName}{(string) The name of the ATLAS R package used to generate the results (this is used to extract cohort jsons)}

\item{studyJsonList}{(list) A list of lists per cohort with the cohort_name, cohort_id and cohort_json}

\item{studyName}{(string) A reference study name}

\item{studyDescription}{(string) A description of the study}

\item{researcherName}{(string) Name of the researcher who developed the study}

\item{researcherEmail}{(string) Email of the researcher who developed the study}

\item{researcherOrg}{(string) Organisation of the researcher who developed the study}

\item{databaseName}{(string) name of the database used to develop the model/s}

\item{databaseAcronym}{(string) acronym of the database used to develop the model/s}

\item{databaseVersion}{(int) Version of the database used to develop the model/s}

\item{databaseDescription}{(string) Description of the database used to develop the model/s}

\item{databaseType}{(string) Type of the database used to develop the model/s (e.g., claims)}

\item{valDatabases}{(list) A named list with details of the external validation databases.  Needs to contain: name, description, version, type.}

\item{resultLocation}{(string) location of directory where the main package results were saved}

\item{resultPattern}{(string) A string to match to select models of interest}

\item{validationLocation}{(string) location of directory where the validation package results were saved}

\item{addInternalValidation}{(boolean) Whether the internval validation results should be uploaded}

\item{addExternalValidation}{(boolean) Whether the externval validation results should be uploaded}

\item{gsubVal}{(string) Remove patterns from the result name}

\item{removePattern}{(string) Restrict to result names with this pattern}
}
\value{
Returns NULL but uploads all the results in resultLocation to the PatientLevelPrediction result tables in resultSchema
}
\description{
This function formats and uploads results that have been generated via an ATLAS prediction package into a database
}
\details{
This function can be used upload PatientLevelPrediction results into a database
}
