% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{positiveLikelihoodRatio}
\alias{positiveLikelihoodRatio}
\title{Calculate the positiveLikelihoodRatio}
\usage{
positiveLikelihoodRatio(TP, TN, FN, FP)
}
\arguments{
\item{TP}{Number of true positives}

\item{TN}{Number of true negatives}

\item{FN}{Number of false negatives}

\item{FP}{Number of false positives}
}
\value{
positiveLikelihoodRatio value
}
\description{
Calculate the positiveLikelihoodRatio
}
\details{
Calculate the positiveLikelihoodRatio
}
