% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{positivePredictiveValue}
\alias{positivePredictiveValue}
\title{Calculate the positivePredictiveValue}
\usage{
positivePredictiveValue(TP, TN, FN, FP)
}
\arguments{
\item{TP}{Number of true positives}

\item{TN}{Number of true negatives}

\item{FN}{Number of false negatives}

\item{FP}{Number of false positives}
}
\value{
positivePredictiveValue value
}
\description{
Calculate the positivePredictiveValue
}
\details{
Calculate the positivePredictiveValue
}
