% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict.predictiveModel}
\alias{predict.predictiveModel}
\title{Create predictive probabilities}
\usage{
\method{predict}{predictiveModel}(predictiveModel, cohortsData, covariateData)
}
\arguments{
\item{predictiveModel}{An object of type \code{predictiveModel} as generated using \code\link{{fitPredictiveModel}}.}

\item{cohortsData}{An object of type \code{cohortsData} as generated using \code\link{{getDbCohortsData}}.}

\item{covariateData}{An object of type \code{covariateData} as generated using \code\link{{getDbCovarteData}}.}
}
\description{
Create predictive probabilities
}
\details{
Note that the cohortsData and covariateData objects need to come from the same population.
}

