% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlmNetModels.R
\name{predictGlmNet}
\alias{predictGlmNet}
\title{predictGlmNet}
\usage{
predictGlmNet(plpModel, data, cohort)
}
\arguments{
\item{plpModel}{An object of type \code{predictiveModel} as generated using
\code{\link{fitPlp}}.}

\item{data}{The new plpData containing the covariateData for the new population}

\item{cohort}{The cohort to calculate the prediction for}
}
\description{
predicts a glmnet model
}
