% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developModel.R
\name{predictPlp}
\alias{predictPlp}
\title{predictPlp}
\usage{
predictPlp(plpModel, plpData)
}
\arguments{
\item{plpModel}{An object of type \code{plpModel} - a patient level prediction model}

\item{plpData}{An object of type \code{plpData} - the patient level prediction
data extracted from the CDM.}
}
\value{
An ffdf object containing the prediction for each person in the cohort
}
\description{
Predict the risk of the outcome using the input plpModel for the input plpData
}
\details{
The function applied the trained model on the plpData to make predictions
}

