% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Predict.R
\name{predictProbabilities}
\alias{predictProbabilities}
\title{Create predictive probabilities}
\usage{
predictProbabilities(predictiveModel, plpData)
}
\arguments{
\item{predictiveModel}{An object of type \code{predictiveModel} as generated using
\code{\link{fitPredictiveModel}}.}

\item{plpData}{An object of type \code{plpData} as generated using
\code{\link{getDbPlpData}}.}
}
\description{
Create predictive probabilities
}
\details{
Note that the cohortData and covariateData objects need to come from the same population.
}

