% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperparameterSettings.R
\name{prepareHyperparameterGrid}
\alias{prepareHyperparameterGrid}
\title{Prepare Hyperparameter}
\usage{
prepareHyperparameterGrid(paramDefinition, hyperSettings)
}
\arguments{
\item{paramDefinition}{A list defining the hyperparameters and their possible values.}

\item{hyperSettings}{An object of class \code{hyperparameterSettings} created using \code{createHyperparameterSettings}.}
}
\value{
An iterator object with methods to get the next hyperparameter combination and finalize the search.
}
\description{
Prepare Hyperparameter
}
