% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData.R
\name{print.plpData}
\alias{print.plpData}
\title{Print a plpData object}
\usage{
\method{print}{plpData}(x, ...)
}
\arguments{
\item{x}{The plpData object to print}

\item{...}{Additional arguments}
}
\value{
A message describing the object
}
\description{
Print a plpData object
}
