% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest.R
\name{randomForest.set}
\alias{randomForest.set}
\title{Create setting for random forest model with python (very fast)}
\usage{
randomForest.set(mtries = -1, ntrees = c(10, 500), max_depth = 17,
  varImp = T)
}
\arguments{
\item{mtries}{The number of features to include in each tree (-1 defaults to square root of total features)}

\item{ntrees}{The number of trees to build}

\item{max_depth}{Maximum number of interactions - a large value will lead to slow model training}

\item{varImp}{Perform an initial variable selection prior to fitting the model to select the useful variables}
}
\description{
Create setting for random forest model with python (very fast)
}
\examples{
\dontrun{
model.rf <- randomForest.set(mtries=c(-1,5,20),  ntrees=c(10,100), 
                           max_depth=c(5,20))
}                           
}

