% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSplitting.R
\name{randomSplitter}
\alias{randomSplitter}
\title{Split data into random subsets stratified by class}
\usage{
randomSplitter(population, test = 0.3, train = NULL, nfold = 3,
  seed = NULL)
}
\arguments{
\item{population}{An object created using createStudyPopulation().}

\item{test}{A real number between 0 and 1 indicating the test set fraction of the data}

\item{train}{A real number between 0 and 1 indicating the train set fraction of the data.
If not set train is equal to 1 - test}

\item{nfold}{An integer >= 1 specifying the number of folds used in cross validation}

\item{seed}{If set a fixed seed is used, otherwise a random split is performed}
}
\value{
A dataframe containing the columns: rowId and index
}
\description{
Split data into random subsets stratified by class
}
\details{
Returns a dataframe of rowIds and indexes with a -1 index indicating the rowId belongs to the test
set and a positive integer index value indicating the rowId's cross valiation fold within the train
set.
}
