% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalibrateModel.R
\name{recalibrateModel}
\alias{recalibrateModel}
\title{recalibrateModel - Re-calibrate model probabilities.}
\usage{
recalibrateModel(
  plpResult,
  recalibrationMethod,
  logSettings = createLogSettings(verbosity = "DEBUG", timeStamp = T, logName =
    "plpLogRecalibration"),
  saveDirectory
)
}
\description{
Recalibrate model
}
\details{
This function takes as input a plpResult as generated from PatientLevelPrediction::loadPlpResult() and returns again a model
with re-calibrated probabilities.
}
