% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recalibration.R
\name{recalibratePlpRefit}
\alias{recalibratePlpRefit}
\title{recalibratePlpRefit}
\usage{
recalibratePlpRefit(plpModel, newPopulation, newData, returnModel = FALSE)
}
\arguments{
\item{plpModel}{The trained plpModel (runPlp$model)}

\item{newPopulation}{The population created using createStudyPopulation() who will have their risks predicted}

\item{newData}{An object of type \code{plpData} - the patient level prediction
data extracted from the CDM.}

\item{returnModel}{Logical: return the refitted model}
}
\value{
An prediction dataframe with the predictions of the recalibrated model added
}
\description{
Recalibrating a model by refitting it
}
