% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recalibration.R
\name{recalibrationInTheLarge}
\alias{recalibrationInTheLarge}
\title{recalibrationInTheLarge}
\usage{
recalibrationInTheLarge(prediction, columnType = "evaluationType")
}
\arguments{
\item{prediction}{A prediction dataframe}

\item{analysisId}{The model analysisId}

\item{typeColumn}{The column name where the strata types are specified}

\item{method}{Method used to recalibrate ('recalibrationInTheLarge' or 'weakRecalibration' )}
}
\value{
An prediction dataframe with the recalibrated predictions added
}
\description{
Recalibrate a model using the recalibrationInTheLarge method which calculates a single correction factor
for the average predicted risks to match the average observed risks
}
\keyword{internal}
