% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{registerParallelBackend}
\alias{registerParallelBackend}
\title{registerParallelBackend - Registers a parallel backend}
\usage{
registerParallelBackend(cores = NULL, logical = TRUE)
}
\arguments{
\item{cores}{the number of cores to use for multi core operations}

\item{logical}{whether to consider logical or physical cores}
}
\description{
registerParallelBackend registers a parallel backend for multi core
processing It will detect the number of cores available automatically unless
specified.
}
\examples{
# detect logical cores automatically
registerParallelBackend()

# use four physical cores
numCores <- 4
registerParallelBackend(numCores, logical = FALSE)

}
