% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{registerSequentialBackend}
\alias{registerSequentialBackend}
\title{registerSequentialBackend - Registers a sequential backend}
\usage{
registerSequentialBackend()
}
\description{
registerSequentialBackend registers a sequential backend for single core
processing.
}
\examples{
# register a sequential backend
registerSequentialBackend()
}
