% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering.R
\name{removeRareFeatures}
\alias{removeRareFeatures}
\title{A function that removes rare features from the data}
\usage{
removeRareFeatures(trainData, featureEngineeringSettings, done = FALSE)
}
\arguments{
\item{trainData}{The data to be normalized}

\item{featureEngineeringSettings}{The settings for the normalization}

\item{done}{Whether to find and remove rare features or remove them only (bool)}
}
\value{
The data with rare features removed
}
\description{
A function that removes rare features from the data
}
\details{
removes features that are present in less than a certain fraction of the population
}
\keyword{internal}
