% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSettings.R
\name{RFclassifier}
\alias{RFclassifier}
\title{Create setting for random forest model}
\usage{
RFclassifier(mtries = -1, ntrees = c(10, 500), rsampRate = 0.9,
  csampRate = 1, bal = F, nbins = 20, max_depth = 17, min_rows = 2)
}
\arguments{
\item{mtries}{The number of features to include in each tree (-1 defaults to square root of total features)}

\item{ntrees}{The number of trees to build}

\item{rsampRate}{The fraction of rows to include in each tree during training}

\item{csampRate}{The fraction of features to include in each tree during training}

\item{bal}{Whether to balance the training set classes}

\item{nbins}{Number of bins used in continuous variables?}

\item{max_depth}{Maximum number of interactions - a large value will lead to slow model training}

\item{min_rows}{The minimum number of rows required at each end node of the tree}
}
\description{
Create setting for random forest model
}
\examples{
model.rf <- RFclassifier(mtries=c(-1,5,20), rsampRate=c(0.5,0.9,1),csampRate=1, ntrees=c(10,100), bal=c(F,T),
                           max_depth=c(5,20))
}

