% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetCohorts.R
\name{saveCohortData}
\alias{saveCohortData}
\title{Save the cohort data to folder}
\usage{
saveCohortData(cohortData, file)
}
\arguments{
\item{cohortData}{An object of type \code{cohortData} as generated using \code{getDbcohortData}.}

\item{file}{The name of the folder where the data will be written. The folder should
not yet exist.}
}
\description{
\code{saveCohortData} saves an object of type cohortData to folder.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}
\examples{
#todo
}

