% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetCohorts.R
\name{saveCohortsData}
\alias{saveCohortsData}
\title{Save the cohort data to folder}
\usage{
saveCohortsData(cohortsData, file)
}
\arguments{
\item{cohortsData}{An object of type \code{cohortsData} as generated using \code{getDbCohortsData}.}

\item{file}{The name of the folder where the data will be written. The folder should
not yet exist.}
}
\description{
\code{saveCohortsData} saves an object of type outcomeData to folder.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}
\examples{
#todo
}

