% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetCovariates.R
\name{saveCovariateData}
\alias{saveCovariateData}
\title{Save the covariate data to folder}
\usage{
saveCovariateData(covariateData, file)
}
\arguments{
\item{covariateData}{An object of type \code{covariateData} as generated using
\code{getDbCovariateData}.}

\item{file}{The name of the folder where the data will be written. The folder should not
                       yet exist.}
}
\description{
\code{saveCovariateData} saves an object of type covariateData to folder.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}
\examples{
# todo
}

