% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensemble.R
\name{saveEnsemblePlpModel}
\alias{saveEnsemblePlpModel}
\title{saves the Ensmeble plp model}
\usage{
saveEnsemblePlpModel(ensembleModel, dirPath)
}
\arguments{
\item{ensembleModel}{The ensemble model to save}

\item{dirPath}{The location to save the model}
}
\description{
saves the Ensmeble plp model
}
\details{
Saves a plp ensemble model
}
