% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensemble.R
\name{saveEnsemblePlpResult}
\alias{saveEnsemblePlpResult}
\title{saves the Ensemble plp results}
\usage{
saveEnsemblePlpResult(ensembleResult, dirPath)
}
\arguments{
\item{ensembleResult}{The ensemble result}

\item{dirPath}{The location to save the ensemble results}
}
\description{
saves the Ensemble plp results
}
\details{
Saves a plp ensemble results
}
