% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{saveEvaluation}
\alias{saveEvaluation}
\title{Saves the evalaution dataframe to csv}
\usage{
saveEvaluation(evaluation, location)
}
\arguments{
\item{evaluation}{The evaluation object}

\item{location}{The directory to save the csv}
}
\description{
Saves the evalaution dataframe to csv
}
\details{
Saves the evaluation on new data to the input location
}

