% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLibSVM.R
\name{saveLibSVM}
\alias{saveLibSVM}
\title{Save the plpData in the sparse libSVM format}
\usage{
saveLibSVM(population, plpData, filePath, mapping = NULL, silent = F)
}
\arguments{
\item{population}{The population created using createStudyPopulation() who will be used to develop the model}

\item{plpData}{An object of type \code{plpData} - the patient level prediction 
data extracted from the CDM.}

\item{filePath}{The path to the directory to output the files}

\item{silent}{Whether to turn off progress reporting}

\item{mappings}{An ffdf containing originalCovariateId (old) and covariateId (new) columns specifying the old to new mapping}
}
\description{
Converts the plpData to libSVM format and saves the data in the user specified directory
}
\details{
Given the plpData and a directory the libSVM format data will be saved into the directory.  The file plpData.txt 
contains the plpData in libSVM format, file covRef.txt is the covariate reference dataframe, the file rowId.txt is a 
vector of the rowIds of each row in the plpData file.  The plpData.txt can be loaded into h2o, oython or spark for 
running efficient machine learning techniques
}
\examples{
   
# To convert plpData into libSVM and save results to C:\\plpData

}

