% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetOutcomes.R
\name{saveOutcomeData}
\alias{saveOutcomeData}
\title{Save the outcome data to folder}
\usage{
saveOutcomeData(outcomeData, file)
}
\arguments{
\item{outcomeData}{An object of type \code{outcomeData} as generated using \code{getDbOutcomeData}.}

\item{file}{The name of the folder where the data will be written. The folder should
not yet exist.}
}
\description{
\code{saveOutcomeData} saves an object of type outcomeData to folder.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}

