% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetPlpData.R
\name{savePlpData}
\alias{savePlpData}
\title{Save the PatientLevelPrediction data to folder}
\usage{
savePlpData(plpData, file)
}
\arguments{
\item{file}{The name of the folder where the data will be written. The folder should
                          not yet exist.}

\item{cohortMethodData}{An object of type \code{plpData} as generated using
\code{getDbPlPData}.}
}
\description{
\code{saveCohortMethodData} saves an object of type cohortMethodData to folder.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}
\examples{
# todo
}

