% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{savePlpToCsv}
\alias{savePlpToCsv}
\title{Save parts of the plp result as a csv for transparent sharing}
\usage{
savePlpToCsv(result, dirPath)
}
\arguments{
\item{result}{An object of class runPlp with development or validation results}

\item{dirPath}{The directory the save the results as csv files}
}
\description{
Save parts of the plp result as a csv for transparent sharing
}
\details{
Saves the main results as a csv (these files can be read by the shiny app)
}
