% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMultiplePlp.R
\name{savePredictionAnalysisList}
\alias{savePredictionAnalysisList}
\title{Saves a json prediction settings given R settings}
\usage{
savePredictionAnalysisList(workFolder = "inst/settings/", cohortIds,
  cohortNames, outcomeIds, outcomeNames, covariateSettingsList,
  populationSettingList, modelSettingList, maxSampleSize = NULL,
  washoutPeriod = 0, minCovariateFraction = 0, normalizeData = T,
  testSplit = "person", testFraction = 0.25, splitSeed = 1, nfold = 3,
  verbosity = "INFO")
}
\arguments{
\item{workFolder}{Location to save json specification}

\item{cohortIds}{Vector of target population cohort ids}

\item{cohortNames}{Vector of names for each cohortId}

\item{outcomeIds}{Vector of outcome cohort ids}

\item{outcomeNames}{Vector of names for each outcomeId}

\item{covariateSettingsList}{A list of covariate settings}

\item{populationSettingList}{A list of population settings}

\item{modelSettingList}{A list of model settings}

\item{maxSampleSize}{If not NULL then max number of target population to sample for model training}

\item{washoutPeriod}{Minimum prior observation for each person in target pop to be included}

\item{minCovariateFraction}{Minimum covariate fraction to include}

\item{normalizeData}{Whether to normalise data}

\item{testSplit}{Split by person or time}

\item{testFraction}{Fractiuon of data to use for test set}

\item{splitSeed}{Seed used in test split}

\item{nfold}{Number of folds used when training model}

\item{verbosity}{Level of log details}
}
\description{
Saves a json prediction settings given R settings
}
\details{
This function interprets a json with the multiple prediction settings and creates a list 
that can be combined with connection settings to run a multiple prediction study
}
