% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadData.R
\name{saveTrainTestData}
\alias{saveTrainTestData}
\title{Save the cohort data to folder}
\usage{
saveTrainTestData(data, file, test = T, envir = NULL, overwrite = F)
}
\arguments{
\item{data}{Data object before runModelDevelopment.}

\item{file}{The name of the folder where the data will be written. The folder should
not yet exist.}

\item{envir}{The environment for to evaluate variables when saving.}

\item{overwrite}{Whether to force overwrite an existing file}
}
\description{
\code{saveTrainTestData} saves an data object before runModelDevelopment to folder.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}
\examples{
# todo

}
