% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{sensitivity}
\alias{sensitivity}
\title{Calculate the sensitivity}
\usage{
sensitivity(TP, TN, FN, FP)
}
\arguments{
\item{TP}{Number of true positives}

\item{TN}{Number of true negatives}

\item{FN}{Number of false negatives}

\item{FP}{Number of false positives}
}
\value{
sensitivity value
}
\description{
Calculate the sensitivity
}
\details{
Calculate the sensitivity
}
