% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesBridgeModel.R
\name{setBayesBridge}
\alias{setBayesBridge}
\title{setBayesBridge}
\usage{
setBayesBridge(
  nIter = 1000,
  burnIter = 250,
  bridgeExponent = 0.5,
  regularizingSlabSize = 1,
  thin = 1,
  globalScale = 0.1,
  coefSamplerType = "cg",
  seed = NULL,
  useGPU = FALSE
)
}
\arguments{
\item{nIter}{total number of MCMC iterations i.e. burn-ins + saved 
posterior draws}

\item{burnIter}{number of burn-in iterations}

\item{bridgeExponent}{exponent for bridge prior}

\item{regularizingSlabSize}{Numeric: Standard deviation of the Gaussian 
tail-regularizer on the bridge prior.}

\item{thin}{Number of iterations per saved samples for “thinning” 
MCMC to reduce the output size.}

\item{globalScale}{Numeric: Reference prior for a scale parameter}

\item{coefSamplerType}{An object to specify the sampling method to update regression coefficients:
\itemize{
                                       \item{None}{ Chooses a method via a crude heuristic based on model type}
                                       \item{cholesky}{ Cholesky decomposition based sampler}
                                       \item{cg}{ Conjugate gradient sampler preferred over Cholesky for linear and logistic models with large+sparse design matrix}
                                       \item{hmc}{ Hamilton Monte Carlo for other models}
                                       }}

\item{useGPU}{if model should be fit using GPU and cupy}
}
\description{
setBayesBridge
}
