% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatBoostPython.R
\name{setCatBoost}
\alias{setCatBoost}
\title{Create settings for CatBoost classifier using Python catboost}
\usage{
setCatBoost(
  iterations = c(100, 300),
  depth = c(4, 6, 8),
  learningRate = c(0.05, 0.1, 0.3),
  l2LeafReg = c(3),
  randomStrength = c(1),
  borderCount = c(254),
  rsm = c(1),
  subsample = c(1),
  earlyStoppingRounds = c(25),
  threadCount = 20,
  verbose = FALSE,
  seed = sample(1e+07, 1)
)
}
\arguments{
\item{iterations}{Number of boosting iterations.}

\item{depth}{Depth of trees.}

\item{learningRate}{Learning rate.}

\item{l2LeafReg}{L2 regularization coefficient.}

\item{randomStrength}{Score standard deviation multiplier.}

\item{borderCount}{Number of splits for numerical features.}

\item{rsm}{Fraction of features used at each split.}

\item{subsample}{Fraction of samples used for each tree.}

\item{earlyStoppingRounds}{Stop if metric does not improve for this many rounds.}

\item{threadCount}{Number of threads to use.}

\item{verbose}{Verbose output level.}

\item{seed}{Random seed.}
}
\value{
A modelSettings object that can be used by \code{runPlp}.
}
\description{
Create settings for CatBoost classifier using Python catboost
}
\examples{
\dontshow{ # dontrun reason: requires python's catboost, checkCatBoostPython() will error without it }
\dontrun{
modelCatBoost <- setCatBoost(
  iterations = c(100, 300),
  depth = c(4, 6, 8),
  learningRate = c(0.05, 0.1, 0.3)
)
}
}
