% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covNN.R
\name{setCovNN}
\alias{setCovNN}
\title{Create setting for CovNN model}
\usage{
setCovNN(batch_size = 1000, outcome_weight = 2, dropout = 0,
  epochs = 10, filters = 3, kernel_size = 10,
  loss = "binary_crossentropy", seed = NULL)
}
\arguments{
\item{batch_size}{The number of samples to used in each batch during model training}

\item{outcome_weight}{The weight assined to the outcome (make greater than 1 to reduce unballanced label issue)}

\item{dropout}{[currently not used] the dropout rate for regularisation}

\item{epochs}{The number of times data is used to train the model (e.g., epoches=1 means data only used once to train)}

\item{filters}{The number of columns output by each convolution}

\item{kernel_size}{The number of time dimensions used for each convolution}

\item{loss}{The loss function implemented}

\item{seed}{The random seed}
}
\description{
Create setting for CovNN model
}
\examples{
\dontrun{
model.CovNN <- setCovNN()
}
}
