% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovNN2.R
\name{setCovNN2}
\alias{setCovNN2}
\title{Create setting for CovNN2 model - convolution across input and time - https://arxiv.org/pdf/1608.00647.pdf}
\usage{
setCovNN2(
  batchSize = 1000,
  outcomeWeight = 1,
  lr = 1e-05,
  decay = 1e-06,
  dropout = 0,
  epochs = 10,
  filters = 3,
  kernelSize = 10,
  loss = "binary_crossentropy",
  seed = NULL
)
}
\arguments{
\item{batchSize}{The number of samples to used in each batch during model training}

\item{outcomeWeight}{The weight assined to the outcome (make greater than 1 to reduce unballanced label issue)}

\item{lr}{The learning rate}

\item{decay}{The decay of the learning rate}

\item{dropout}{[currently not used] the dropout rate for regularisation}

\item{epochs}{The number of times data is used to train the model (e.g., epoches=1 means data only used once to train)}

\item{filters}{The number of columns output by each convolution}

\item{kernelSize}{The number of time dimensions used for each convolution}

\item{loss}{The loss function implemented}

\item{seed}{The random seed}
}
\description{
Create setting for CovNN2 model - convolution across input and time - https://arxiv.org/pdf/1608.00647.pdf
}
\examples{
\dontrun{
model.CovNN <- setCovNN()
}
}
