% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepNN.R
\name{setDeepNN}
\alias{setDeepNN}
\title{Create setting for DeepNN model}
\usage{
setDeepNN(
  units = list(c(128, 64), 128),
  layer_dropout = c(0.2),
  lr = c(1e-04),
  decay = c(1e-05),
  outcome_weight = c(1),
  batch_size = c(100),
  epochs = c(100),
  seed = NULL
)
}
\arguments{
\item{units}{The number of units of the deep network - as a list of vectors}

\item{layer_dropout}{The layer dropout rate (regularisation)}

\item{lr}{Learning rate}

\item{decay}{Learning rate decay over each update.}

\item{outcome_weight}{The weight of the outcome class in the loss function}

\item{batch_size}{The number of data points to use per training batch}

\item{epochs}{Number of times to iterate over dataset}

\item{seed}{Random seed used by deep learning model}
}
\description{
Create setting for DeepNN model
}
\examples{
\dontrun{
model <- setDeepNN()
}
}
