% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBMSurvival.R
\name{setGBMSurvival}
\alias{setGBMSurvival}
\title{Create setting for GBM Survival with python
#' @description
This creates a setting for fitting  GBM surivial model.  You need sksurv python install.  To install this open your command line and type: conda install -c sebp scikit-survival}
\usage{
setGBMSurvival(loss = "coxph", learningRate = 0.1, nEstimators = c(100),
  criterion = "friedman_mse", minSamplesSplit = 2, minSamplesLeaf = 1,
  minWeightFractionLeaf = 0, maxDepth = c(3, 10, 17),
  minImpuritySplit = NULL, minImpurityDecrease = 0, maxFeatures = NULL,
  maxLeafNodes = NULL, presort = "auto", subsample = 1, dropoutRate = 0,
  seed = NULL, quiet = F)
}
\arguments{
\item{loss}{A string specifying the loss function to minimise (default: 'coxph' )}

\item{learningRate}{A double specifying the learning rate (controls convergence speed)}

\item{nEstimators}{An integer specifying how many trees to build}

\item{criterion}{Default: 'friedman_mse'}

\item{minSamplesSplit}{An integer specifying min samples per tree split (complexity)}

\item{minSamplesLeaf}{An integer specifying min samples per leaf (complexity)}

\item{minWeightFractionLeaf}{Lookup}

\item{maxDepth}{An integer specifying the max depth of trees (complexity)}

\item{minImpuritySplit}{A double or NULL specifying the minimum impurity split}

\item{minImpurityDecrease}{will add}

\item{maxFeatures}{will add}

\item{maxLeafNodes}{will add}

\item{presort}{will add}

\item{subsample}{will add}

\item{dropoutRate}{will add}

\item{seed}{will add}

\item{quiet}{will add}
}
\description{
Create setting for GBM Survival with python
#' @description
This creates a setting for fitting  GBM surivial model.  You need sksurv python install.  To install this open your command line and type: conda install -c sebp scikit-survival
}
\details{
Pick the hyper-parameters you want to do a grid search for
}
\examples{
\dontrun{
gbmSurv <- setGBMSurvival(learningRate=c(0.1,0.01), nEstimators =c(10,50,100),
 maxDepth=c(4,10,17), seed = 2)
}
}
