% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CyclopsSettings.R
\name{setIterativeHardThresholding}
\alias{setIterativeHardThresholding}
\title{Create setting for lasso logistic regression}
\usage{
setIterativeHardThresholding(
  K = 10,
  penalty = "bic",
  seed = sample(1e+05, 1),
  exclude = c(),
  forceIntercept = F,
  fitBestSubset = FALSE,
  initialRidgeVariance = 0.01,
  tolerance = 1e-08,
  maxIterations = 10000,
  threshold = 1e-06,
  delta = 0,
  nTries = 10
)
}
\arguments{
\item{K}{The maximum number of non-zero predictors}

\item{penalty}{Specifies the IHT penalty; possible values are `BIC` or `AIC` or a numeric value
If set to `auto` it will do CV to determine best penalty}

\item{seed}{An option to add a seed when training the model}

\item{exclude}{A vector of numbers or covariateId names to exclude from prior}

\item{forceIntercept}{Logical: Force intercept coefficient into regularization}

\item{fitBestSubset}{Logical: Fit final subset with no regularization}

\item{initialRidgeVariance}{integer or character vector. If set to `auto` will fit Ridge regression using
cross validation to determine best `initialRidgeVariance` value.}

\item{tolerance}{numeric}

\item{maxIterations}{integer}

\item{threshold}{numeric}

\item{delta}{numeric}

\item{nTries}{If `penalty` is `auto`, how many penalties to include in grid search}
}
\description{
Create setting for lasso logistic regression
}
\examples{
model.lr <- setLassoLogisticRegression()
}
