% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.R
\name{setKNN}
\alias{setKNN}
\title{Create setting for knn model}
\usage{
setKNN(k = 1000, indexFolder = file.path(getwd(), "knn"))
}
\arguments{
\item{k}{The number of neighbors to consider}

\item{indexFolder}{The directory where the results and intermediate steps are output}
}
\description{
Create setting for knn model
}
\examples{
\dontrun{
model.knn <- setKNN(k=10000)
}
}
