% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRTorch.R
\name{setLRTorch}
\alias{setLRTorch}
\title{Create setting for logistics regression model with python}
\usage{
setLRTorch(w_decay = c(0.003, 0.005, 0.007), epochs = c(20, 50, 100),
  class_weight = 0)
}
\arguments{
\item{w_decay}{The l2 regularisation}

\item{epochs}{The number of epochs}

\item{class_weight}{weight the class in imblanced data}

\item{seed}{A seed for the model}
}
\description{
Create setting for logistics regression model with python
}
\examples{
\dontrun{
model.lrTorch <- setLRTorch()
}
}
