% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLPTorch.R
\name{setMLPTorch}
\alias{setMLPTorch}
\title{Create setting for neural network model with python}
\usage{
setMLPTorch(size = 200, epochs = 10, seed = 0)
}
\arguments{
\item{size}{The number of hidden nodes}

\item{seed}{A seed for the model}

\item{alpha}{The l2 regularisation}
}
\description{
Create setting for neural network model with python
}
\examples{
\dontrun{
model.mlpTorch <- setMLPTorch()
}
}
