% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNNTorch.R
\name{setRNNTorch}
\alias{setRNNTorch}
\title{Create setting for RNN model with python}
\usage{
setRNNTorch(hidden_size = c(50, 100), epochs = c(20, 50), seed = 0,
  time_window = 12, class_weight = 0)
}
\arguments{
\item{hidden_size}{The hidden size}

\item{epochs}{The number of epochs}

\item{seed}{A seed for the model}

\item{time_window}{A time window to get temporal data}
}
\description{
Create setting for RNN model with python
}
\examples{
\dontrun{
model.rnnTorch <- setRNNTorch()
}
}
