% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNNTorch2.R
\name{setRNNTorch2}
\alias{setRNNTorch2}
\title{Create setting for RNN model with python}
\usage{
setRNNTorch2(hidden_size = c(50, 100), epochs = c(20, 50), seed = 0,
  class_weight = 0, type = "RNN")
}
\arguments{
\item{hidden_size}{The hidden size}

\item{epochs}{The number of epochs}

\item{seed}{A seed for the model}

\item{class_weight}{The class weight used for imbalanced data: 
 0: Inverse ratio between positives and negatives
-1: Focal loss}

\item{type}{It can be normal 'RNN', 'BiRNN' (bidirectional RNN) and 'GRU'}
}
\description{
Create setting for RNN model with python
}
\examples{
\dontrun{
model.rnnTorch <- setRNNTorch()
}
}
