% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SageMakerBinary.R
\name{setSagemakerBinary}
\alias{setSagemakerBinary}
\title{Create setting for sagemaker model}
\usage{
setSagemakerBinary(classifier = "xgboost", bucket, prefix = "data", roleArn,
  otherparams = NULL, seed = NULL)
}
\arguments{
\item{classifier}{The name of the sagemaker binary classifier to use (pick from: knn, xgboost or linear-learner)}

\item{bucket}{The s3 bucker string to save data for model training}

\item{prefix}{The s3 subdirectory for the data}

\item{roleArn}{The amazon roleArn}

\item{otherparams}{Other parameters for training (currently not working)}

\item{seed}{The seed for the training}
}
\description{
Create setting for sagemaker model
}
\examples{
\dontrun{
model.sm <- setSagemakerBinary(classifier='gxboost',  bucket='ohdsi3')
}                           
}
