% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableSelection.R
\name{setStepwiseSelection}
\alias{setStepwiseSelection}
\title{Selects features using stepwise selection}
\usage{
setStepwiseSelection(
  modelSettings,
  selectMethod = "backward",
  nInitialVariables = 50,
  nVariables = 20,
  stepSize = 1
)
}
\arguments{
\item{modelSettings}{settings of model to use in fit after selecting variables}

\item{selectMethod}{`backward` or `forward` selection}

\item{nInitialVariables}{# of variables to select initially}

\item{nVariables}{amount of variables to select, default `20`}

\item{stepSize}{How many variables to add/remove in each step}
}
\description{
Selects features using stepwise selection
}
