% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMultiplePlp.R
\name{setTimeAtRisk}
\alias{setTimeAtRisk}
\title{setTimeAtRisk}
\usage{
setTimeAtRisk(includeAllOutcomes = T, firstExposureOnly = F,
  washoutPeriod = 0, removeSubjectsWithPriorOutcome = T,
  priorOutcomeLookback = 99999, riskWindowStart = 1,
  addExposureDaysToStart = F, riskWindowEnd = 365,
  addExposureDaysToEnd = F, requireTimeAtRisk = T,
  minTimeAtRisk = riskWindowEnd - riskWindowStart)
}
\arguments{
\item{includeAllOutcomes}{Do you want to include people who have the outcome but are not observed for the whole at risk period?}

\item{firstExposureOnly}{Only consider the first time occurence of the outcome?}

\item{washoutPeriod}{The minimum prior observation in days a person required to be included}

\item{removeSubjectsWithPriorOutcome}{Remove people who have the outcome some period before the time at risk?}

\item{priorOutcomeLookback}{The number of days prior to investigate for the variable removeSubjectsWithPriorOutcome}

\item{riskWindowStart}{The number of days after the at risk popualtion subject's index date to start the time at risk period}

\item{addExposureDaysToStart}{Should the risk window start be relative to the index end date instead?}

\item{riskWindowEnd}{The number of days after the at risk popualtion subject's index date to end the time at risk period}

\item{addExposureDaysToEnd}{Should the risk window end be relative to the index end date instead?}

\item{requireTimeAtRisk}{Should you only include people with a minimum time at risk period?}

\item{minTimeAtRisk}{If requireTimeAtRisk is TRUE, then this is the minimum number of days a person must be at risk}
}
\description{
create the timeAtRisks for the multiple analysis studies
}

