% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableSelection.R
\name{setUnivariateSelection}
\alias{setUnivariateSelection}
\title{Selects features based on univariate statistics}
\usage{
setUnivariateSelection(
  modelSettings = PatientLevelPrediction::setLassoLogisticRegression(),
  corMethod = "pearson",
  nVariables = 50
)
}
\arguments{
\item{modelSettings}{settings of model to use in fit after selecting variables}

\item{corMethod}{which type of correlation to use, default `pearson`}

\item{nVariables}{amount of variables to select, default `50`}
}
\description{
Selects features based on univariate statistics
}
