% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{specificity}
\alias{specificity}
\title{Calculate the specificity}
\usage{
specificity(TP, TN, FN, FP)
}
\arguments{
\item{TP}{Number of true positives}

\item{TN}{Number of true negatives}

\item{FN}{Number of false negatives}

\item{FP}{Number of false positives}
}
\value{
specificity value
}
\description{
Calculate the specificity
}
\details{
Calculate the specificity
}
