% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developModelFramework.R
\name{spliter}
\alias{spliter}
\title{spliter}
\usage{
spliter(plpData, type, frac)
}
\arguments{
\item{plpData}{An object of type \code{plpData} - the patient level prediction
data extracted from the CDM.}

\item{type}{The type of train/test split:
\itemize{
\item{'year'}{Split the data based on cohort start date -
test:pre 2013/train:post2013 }
\item{'both'}{Split the data on year 2013 but exclude any train
people who are in the test set.}
\item{'patient'}{Split the data into (1-frac) train set and frac test set.
This is stratified by the outcome}
}}

\item{frac}{The fraction of people who will go into the test set}
}
\value{
A list with the training set as the first element and the testing set as the second element
}
\description{
Various train/test splitting techniques
}
\details{
The function splits the plpData into test/train sets
}

