% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackagePlp.R
\name{standardOutput}
\alias{standardOutput}
\title{standardOutput - takes the output of runPlp or evaluatePlp and converts it into the standardised output
for a network study - three directories (plots, results, summary)}
\usage{
standardOutput(result, table1, outputLocation, studyName, databaseName,
  cohortName, outcomeName)
}
\arguments{
\item{result}{The result of the network study}

\item{table1}{the table1 result}

\item{outputLocation}{The location where the results will be saved - need to have write access}

\item{studyName}{The name of the network study}

\item{databaseName}{The name of the cdm database}

\item{cohortName}{The name of the target population cohort}

\item{outcomeName}{The name of the outcome cohort}
}
\value{
The location of the saved results
}
\description{
This function saves the plp study results into standardised output
}
\details{
This is used to ensure each study collects results consistently
}
