% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableSelection.R
\name{stepwiseSettings}
\alias{stepwiseSettings}
\title{Create settings using stepwise selection for feature selection}
\usage{
stepwiseSettings(
  k = 20,
  selectMethod = "backward",
  kStart = 100,
  stepSize = 1,
  modelSettings = PatientLevelPrediction::setLassoLogisticRegression()
)
}
\arguments{
\item{k}{number of variables to select, default `20`}

\item{selectMethod}{`backward` or `forward` selection}

\item{kStart}{number of variables to select initially}

\item{stepSize}{How many variables to add/remove in each step}

\item{modelSettings}{settings of model to use in fit after selecting variables}
}
\description{
Create settings using stepwise selection for feature selection
}
