% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{subgroupEvaluation}
\alias{subgroupEvaluation}
\title{Evaluate a subgroup of the population}
\usage{
subgroupEvaluation(prediction, subPopulation)
}
\arguments{
\item{prediction}{the prediction object found in the plpResult object}

\item{subPopulation}{the subPopulation of interestof type 'population'}
}
\value{
plpResult evaluation object for the subgroup
}
\description{
Evaluate a subgroup of the population
}
\details{
run evaluatePlp in a specific subgroup
}
