% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSplitting.R
\name{subjectSplitter}
\alias{subjectSplitter}
\title{Split data when patients are in the data multiple times such that the same patient is always either in the 
train set or the test set (the same patient cannot be in both the test and train set at different times)}
\usage{
subjectSplitter(population, test = 0.3, train = NULL, nfold = 3,
  seed = NULL)
}
\arguments{
\item{population}{An object created using createStudyPopulation().}

\item{test}{A real number between 0 and 1 indicating the test set fraction of the data}

\item{train}{A real number between 0 and 1 indicating the train set fraction of the data.
If not set train is equal to 1 - test}

\item{nfold}{An integer >= 1 specifying the number of folds used in cross validation}

\item{seed}{If set a fixed seed is used, otherwise a random split is performed}
}
\value{
A dataframe containing the columns: rowId and index
}
\description{
Split data when patients are in the data multiple times such that the same patient is always either in the 
train set or the test set (the same patient cannot be in both the test and train set at different times)
}
\details{
Returns a dataframe of rowIds and indexes with a -1 index indicating the rowId belongs to the test
set and a positive integer index value indicating the rowId's cross valiation fold within the train
set.
}
