% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackagePlp.R
\name{submitResults}
\alias{submitResults}
\title{submitResults - sends a zipped folder to the OHDSI network study repository}
\usage{
submitResults(exportFolder, key, secret)
}
\arguments{
\item{exportFolder}{The path to the folder containing the study results compressed file.}

\item{key}{The key string as provided by the study coordinator}

\item{secret}{The secret string as provided by the study coordinator}
}
\value{
TRUE if the upload was successful.
}
\description{
This function takes as input a zipped folder location and sends it to the OHDSI amazon repository
}
\details{
This is used at the end of a network study to submit the results once a user has checked the folder
}
