% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData.R
\name{summary.plpData}
\alias{summary.plpData}
\title{Summarize a plpData object}
\usage{
\method{summary}{plpData}(object, ...)
}
\arguments{
\item{object}{The plpData object to summarize}

\item{...}{Additional arguments}
}
\value{
A summary of the object containing the number of people, outcomes and covariates
}
\description{
Summarize a plpData object
}
