% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewMultiplePlp.R
\name{summaryPlpAnalyses}
\alias{summaryPlpAnalyses}
\title{summarises the multiple PLP results into a dataframe}
\usage{
summaryPlpAnalyses(analysesLocation)
}
\arguments{
\item{analysesLocation}{The directory containing the results (with the analysis_x folders)}
}
\description{
summarises the multiple PLP results into a dataframe
}
\details{
Loads all the study results contained in the analysesLocation and aggregates a summary of the results
}
