% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSplitting.R
\name{timeSplitter}
\alias{timeSplitter}
\title{Split test/train data by time and then partitions training set into random folds stratified by class}
\usage{
timeSplitter(population, test = 0.3, nfold = 3, silent = F, seed = NULL)
}
\arguments{
\item{population}{An object created using createStudyPopulation().}

\item{test}{A real number between 0 and 1 indicating the test set fraction of the data}

\item{nfold}{An integer >= 1 specifying the number of folds used in cross validation}

\item{silent}{Whether to turn off the progress reporting}
}
\value{
A dataframe containing the columns: rowId and index
}
\description{
Split test/train data by time and then partitions training set into random folds stratified by class
}
\details{
Returns a dataframe of rowIds and indexes with a -1 index indicating the rowId belongs to the test set and 
a positive integer index value indicating the rowId's cross valiation fold within the train set.
}

