% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatting.R
\name{toPlpData}
\alias{toPlpData}
\title{Convert matrix into plpData}
\usage{
toPlpData(data, columnNames, columnTimes = NULL, outcomeId, indexTime = 0,
  includeIndexDay = T)
}
\arguments{
\item{data}{An data.frame with column names or matrix.}

\item{columnNames}{A dataframe with two columns, column 1 contains column ids and column 2 contains names for each column id}

\item{columnTimes}{A dataframe with two columns, column 1 contains column ids and column 2 specifies the time prior to index the variables was recorded}

\item{outcomeId}{The column id containing the outcome}

\item{indexTime}{The time defining the index date}

\item{includeIndexDay}{Boolean - whether to include variables recorded on index date}
}
\value{
Returns an object of class plpData
}
\description{
Converts a matrix (rows = people, columns=variables) into the standard plpData
}
\details{
This function converts matrix into plpData
}
\examples{
#TODO

}
