% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportExport.R
\name{transportModel}
\alias{transportModel}
\title{Transports a plpModel to a new location and removes sensitive data}
\usage{
transportModel(plpModel, outputFolder)
}
\arguments{
\item{plpModel}{A trianed model.}

\item{outputFolder}{The folder on the file system where the CSV files will be created. If the
folder does not yet exist it will be created.}
}
\description{
Transports a plpModel to a new location and removes sensitive data
}
\details{
This function is used to
}
\examples{
\dontrun{
transportModel(plpModel, "s:/temp/exportTest")
}
}
