% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportExport.R
\name{transportPlp}
\alias{transportPlp}
\title{Transports a plpResult to a new location and removed sensitive data}
\usage{
transportPlp(plpResult, outputFolder, n = NULL,
  includeEvaluationStatistics = T, includeThresholdSummary = T,
  includeDemographicSummary = T, includeCalibrationSummary = T,
  includePredictionDistribution = T, includeCovariateSummary = T)
}
\arguments{
\item{plpResult}{An object returned by running \code{RunPlp}.}

\item{outputFolder}{The folder on the file system where the CSV files will be created. If the
folder does not yet exist it will be created.}

\item{n}{The minimum number of people required for each result summary to be included}

\item{includeEvaluationStatistics}{}

\item{includeDemographicSummary}{}

\item{includePredictionDistribution}{}
}
\description{
Transports a plpResult to a new location and removed sensitive data
}
\details{
This function is used to
}
\examples{
\dontrun{
transportPlp(plpResult, "s:/temp/exportTest", n=10)
}
}
