% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableSelection.R
\name{univariateSettings}
\alias{univariateSettings}
\title{Create settings using univariate statistics for feature selection}
\usage{
univariateSettings(k = 20, corMethod = "pearson")
}
\arguments{
\item{k}{number of variables to select, default `20`}

\item{corMethod}{which type of correlation to use, `pearson`, `kendall` or `spearman`. default `pearson`}
}
\description{
Create settings using univariate statistics for feature selection
}
