% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePlp.R
\name{varImportance}
\alias{varImportance}
\title{varImportance}
\usage{
varImportance(models.list)
}
\arguments{
\item{models.list}{A list of plp models or simple plp model}
}
\value{
A table containing the variable importances for each model
}
\description{
Plots the variable importance of the patient level prediction model/s and returns a table of variable importances
}
\details{

}

