% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recalibration.R
\name{weakRecalibration}
\alias{weakRecalibration}
\title{weakRecalibration}
\usage{
weakRecalibration(prediction, columnType = "evaluationType")
}
\arguments{
\item{prediction}{A prediction dataframe}

\item{columnType}{The column name where the strata types are specified}
}
\value{
An prediction dataframe with the recalibrated predictions added
}
\description{
Recalibrate a model using the weakRecalibration method which fits a glm model
to the logit of the predicted risks.
Alsi known as Platt scaling/logistic recalibration
}
\keyword{internal}
