% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationSummary.R
\name{averagePrecision}
\alias{averagePrecision}
\title{Calculate the average precision}
\usage{
averagePrecision(prediction)
}
\arguments{
\item{prediction}{A prediction object}
}
\value{
The average precision
}
\description{
Calculate the average precision
}
\details{
Calculates the average precision from a predition object
}
