% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationSummary.R
\name{brierScore}
\alias{brierScore}
\title{brierScore}
\usage{
brierScore(prediction)
}
\arguments{
\item{prediction}{A prediction object}
}
\value{
A list containing the brier score and the scaled brier score
}
\description{
brierScore
}
\details{
Calculates the brierScore from prediction object
}
