% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{brierScore}
\alias{brierScore}
\title{brierScore}
\usage{
brierScore(prediction)
}
\arguments{
\item{prediction}{A prediction object as generated using the
\code{\link{predictProbabilities}} function.}
}
\value{
A list containing the brier score and the scaled brier score
}
\description{
brierScore
}
\details{
Calculates the brierScore from prediction object
}
