% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{calibrationLine}
\alias{calibrationLine}
\title{calibrationLine}
\usage{
calibrationLine(prediction, numberOfStrata = 10)
}
\arguments{
\item{prediction}{A prediction object as generated using the
\code{\link{predictProbabilities}} function.}

\item{numberOfStrata}{The number of groups to split the prediction into}
}
\description{
calibrationLine
}
\details{
Calculates the calibration from prediction object
}
