% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalValidatePlp.R
\name{checkAllSameInModels}
\alias{checkAllSameInModels}
\title{checkAllSameInModels - Check if all settings are the same across models}
\usage{
checkAllSameInModels(settingsList, settingName)
}
\arguments{
\item{settingsList}{A list of settings to check}

\item{settingName}{The name of the setting to check}
}
\description{
checkAllSameInModels - Check if all settings are the same across models
}
\keyword{internal}
