% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalValidatePlp.R
\name{checkValidateExternalInputs}
\alias{checkValidateExternalInputs}
\title{checkValidateExternalInputs - Check the inputs for validateExternal}
\usage{
checkValidateExternalInputs(
  validationDesignList,
  databaseDetails,
  logSettings,
  outputFolder
)
}
\arguments{
\item{validationDesignList}{A list of validationDesign objects}

\item{databaseDetails}{A list of databaseDetails objects}

\item{logSettings}{An object of logSettings}

\item{outputFolder}{The directory to save the validation results to}
}
\value{
A list of inputs that were modified
}
\description{
checkValidateExternalInputs - Check the inputs for validateExternal
}
\keyword{internal}
