% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePlp.R
\name{comparePlp}
\alias{comparePlp}
\title{comparePlp}
\usage{
comparePlp(models)
}
\arguments{
\item{models}{A list of plp models}
}
\value{
A table summarising the performance value comparision and plots.
}
\description{
Compares the performance of two or more patient level prediction models
}
\details{
The function summarises and plots the performance of the input models for comparison
}

