% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CovariatePrevalence.R
\name{computeCovariateMeans}
\alias{computeCovariateMeans}
\title{Compute covariate means}
\usage{
computeCovariateMeans(cohortData, covariateData, outcomeData = NULL,
  cohortId = NULL, outcomeId = NULL)
}
\arguments{
\item{cohortData}{An object of type \code{cohortData}.}

\item{covariateData}{An object of type \code{covariateData}.}

\item{outcomeData}{An object of type \code{outcomeData}. If NULL then only the overall means will
be computed, else the means will also be computed within the group with the outcome
and the group without the outcome.}

\item{cohortId}{The ID of the specific cohort for which to compute the means.}

\item{outcomeId}{The ID of the specific outcome for which to compute the subgroup means.}
}
\description{
Compute covariate means
}

