% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{configurePython}
\alias{configurePython}
\title{Sets up a virtual environment to use for PLP (can be conda or python)}
\usage{
configurePython(envname = "PLP", envtype = NULL)
}
\arguments{
\item{envname}{A string for the name of the virtual environment (default is 'PLP')}

\item{envtype}{An option for specifying the environment as'conda' or 'python'.  If NULL then the default is 'conda' for windows users and 'python' for non-windows users}
}
\description{
Sets up a virtual environment to use for PLP (can be conda or python)
}
\details{
This function creates a virtual environment that can be used by PatientLevelPrediction
and installs all the required package dependancies.  If using python, pip must be set up.
}
