% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndromedaHelperFunctions.R
\name{convertToCyclopsData.FileSystemDataset}
\alias{convertToCyclopsData.FileSystemDataset}
\title{convertToCyclopsData
temporary add cyclops function here so I don't need to fork cyclops}
\usage{
convertToCyclopsData.FileSystemDataset(
  outcomes,
  covariates,
  modelType = "lr",
  addIntercept = TRUE,
  checkSorting = NULL,
  checkRowIds = TRUE,
  normalize = NULL,
  quiet = FALSE,
  floatingPoint = 64
)
}
\arguments{
\item{outcomes}{outcome dataset}

\item{covariates}{covariates dataset}

\item{modelType}{the model type (most often logistic regression)}

\item{addIntercept}{if an intercept should be added to the model}

\item{checkSorting}{check sorting}

\item{checkRowIds}{check row ids}

\item{normalize}{if normalize}

\item{quiet}{sssss....}

\item{floatingPoint}{floats my boat.}
}
\description{
convertToCyclopsData Convert data from two \code{Arrow} FileSystemDatasets
}
