% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boruta.R
\name{createBorutaFeatureSelection}
\alias{createBorutaFeatureSelection}
\title{Create the settings for Boruta feature selection}
\usage{
createBorutaFeatureSelection(
  nJobs = 10L,
  maxDepth = 5L,
  nTrees = "auto",
  verbosity = 2L,
  iterations = 100L,
  randomState = 42L
)
}
\arguments{
\item{nJobs}{How many jobs to do in parallel}

\item{maxDepth}{Max depth of each tree in the RandomForest used}

\item{nTrees}{How many trees to use, default is `auto`}
}
\value{
An object of class \code{featureEngineeringSettings}
}
\description{
Create the settings for Boruta feature selection
}
\details{
From: https://doi.org/10.18637/jss.v036.i11
}
