% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostCurve.R
\name{createCostCurve}
\alias{createCostCurve}
\title{Create setting for DecisionTree with python}
\usage{
createCostCurve(population, plpData, modeltype = "MLPTorch",
  testSplit = "time", testFraction = 0.25, factors = c(1, 2, 4, 6, 8, 10,
  12, 14, 16, 20, 30, 40, 50), splitSeed = NULL, nfold = 3,
  indexes = NULL, save = NULL, saveModel = T,
  verbosity = futile.logger::INFO, timeStamp = FALSE, analysisId = NULL,
  saveFig = FALSE)
}
\arguments{
\item{max_depth}{The maximum depth of the tree}

\item{min_samples_split}{The minimum samples per split}

\item{min_samples_leaf}{The minimum number of samples per leaf}

\item{min_impurity_split}{Threshold for early stopping in tree growth. A node will split if its impurity is above the threshold, otherwise it is a leaf.}

\item{class_weight}{Balance or None}

\item{seed}{The random state seed}
}
\description{
Create setting for DecisionTree with python
}
\examples{
\dontrun{
model.decisionTree <- setDecisionTree(max_depth=10,min_samples_leaf=10, seed=NULL )
}
}
