% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadToDatabase.R
\name{createDatabaseList}
\alias{createDatabaseList}
\title{Create a list with the database details and database meta data entries}
\usage{
createDatabaseList(cdmDatabaseSchemas, cdmDatabaseName, databaseRefId)
}
\arguments{
\item{cdmDatabaseSchemas}{(string vector) A vector of the cdmDatabaseSchemas used in the study}

\item{cdmDatabaseName}{A sharable name for the database}

\item{databaseRefId}{The databaseName or databaseId}
}
\value{
Returns a data.frame with the database details
}
\description{
This function creates a list with the database details and database meta data entries used in the study
}
\details{
This function is used when inserting database details into the PatientLevelPrediction database results schema
}
