% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadToDatabase.R
\name{createDatabaseList}
\alias{createDatabaseList}
\title{Create a list with the database details and database meta data entries}
\usage{
createDatabaseList(cdmDatabaseSchemas, cdmDatabaseNames, databaseRefIds = NULL)
}
\arguments{
\item{cdmDatabaseSchemas}{(string vector) A vector of the cdmDatabaseSchemas used in the study - if the schemas are not unique per database please also specify databaseRefId}

\item{cdmDatabaseNames}{Sharable names for the databases}

\item{databaseRefIds}{(string vector) Unique database identifiers - what you specified as cdmDatabaseId in \code{PatientLevelPrediction::createDatabaseDetails()} when developing the models}
}
\value{
Returns a data.frame with the database details
}
\description{
This function creates a list with the database details and database meta data entries used in the study
}
\details{
This function is used when inserting database details into the PatientLevelPrediction database results schema
}
