% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glm.R
\name{createGlmModel}
\alias{createGlmModel}
\title{createGlmModel}
\usage{
createGlmModel(coefficients, intercept = 0, finalMapping = "logistic")
}
\arguments{
\item{coefficients}{A dataframe containing two columns, coefficients and
covariateId, both of type numeric. The covariateId column must contain
valid covariateIds that match those used in the /code{FeatureExtraction}
package.}

\item{intercept}{A numeric value representing the intercept of the model.}

\item{finalMapping}{A string representing the final mapping from the 
linear predictors to outcome probabilities. For generalized linear models
this is the inverse of the link function. Supported values is only
"logistic" for logistic regression model at the moment.}
}
\value{
A model object containing the model and the prediction function.
}
\description{
Create a generalized linear model that can be used in the 
PatientLevelPrediction package.
}
