% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperparameterSettings.R
\name{createHyperparameterSettings}
\alias{createHyperparameterSettings}
\title{Create Hyperparameter Settings}
\usage{
createHyperparameterSettings(
  search = "grid",
  tuningMetric = aucMetric,
  sampleSize = NULL,
  randomSeed = NULL,
  generator = NULL
)
}
\arguments{
\item{search}{The type of hyperparameter search to perform. Options are "grid" for grid search, "random" for random search, and "custom" for a user-defined search strategy.}

\item{tuningMetric}{The metric to optimize during hyperparameter tuning. Common
choices include \code{aucMetric} and \code{auprcMetric}.}

\item{sampleSize}{Sample size in case of random sampling}

\item{randomSeed}{Random seed for random sampling}

\item{generator}{An object with \code{initialize}, \code{getNext} and \code{finalize} methods
for custom flexible hyperparameter tuning.}
}
\description{
Create Hyperparameter Settings
}
