% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperparameterSettings.R
\name{createHyperparameterSettings}
\alias{createHyperparameterSettings}
\title{Create Hyperparameter Settings}
\usage{
createHyperparameterSettings(
  search = "grid",
  tuningMetric = "auc",
  sampleSize = NULL,
  randomSeed = NULL,
  searchControl = list(),
  generator = NULL
)
}
\arguments{
\item{search}{The type of hyperparameter search to perform. Options are "grid" for grid search, "random" for random search, and "custom" for a user-defined search strategy.}

\item{tuningMetric}{The metric to optimize during hyperparameter tuning.}
}
\description{
Create Hyperparameter Settings
}
