% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearningCurve.R
\name{createLearningCurvePar}
\alias{createLearningCurvePar}
\title{createLearningCurvePar}
\usage{
createLearningCurvePar(population, plpData, modelSettings,
  testSplit = "person", testFraction = 0.25, trainFractions = c(0.25, 0.5,
  0.75), splitSeed = NULL, nfold = 3, indexes = NULL,
  minCovariateFraction = 0.001)
}
\arguments{
\item{population}{The population created using \code{createStudyPopulation()}
that will be used to develop the model.}

\item{plpData}{An object of type \code{plpData} - the patient level
prediction data extracted from the CDM.}

\item{modelSettings}{An object of class \code{modelSettings} created using
one of the function. Currently only one model is supported:
\itemize{
  \item{\code{setLassoLogisticRegression} - a lasso logistic regression
    model}
}}

\item{testSplit}{Specifies the type of evaluation used. Can be either 
\code{'person'} or \code{'time'}. The value \code{'time'} finds the date
that splots the population into the testing and training fractions
provided. Patients with an index after this date are assigned to the test
set and patients with an index prior to this date are assigned to the
training set. The value \code{'person'} splits the data randomly into
testing and training sets according to fractions provided. The split is
stratified by the class label.}

\item{testFraction}{The fraction of the data, which will be used as the 
testing set in the patient split evaluation.}

\item{trainFractions}{A list of training fractions to create models for.}

\item{splitSeed}{The seed used to split the testing and training set when
using a 'person' type split}

\item{nfold}{The number of folds used in the cross validation (default = 
\code{3}).}

\item{indexes}{A dataframe containing a rowId and index column where the 
index value of -1 means in the test set, and positive integer represents
the cross validation fold (default is \code{NULL}).}

\item{minCovariateFraction}{Minimum covariate prevalence in population to
avoid removal during preprocssing.}
}
\value{
A learning curve object containing the various performance measures
 obtained by the model for each training set fraction. It can be plotted
 using \code{plotLearningCurve}.
}
\description{
Creates a learning curve in parallel, which can be plotted using
 the \code{plotLearningCurve()} function. Currently this functionality is
 only supported by Lasso Logistic Regression.
}
\examples{
\dontrun{
# define model
modelSettings = setLassoLogisticRegression()

# register parallel backend
registerParallelBackend()

# create learning curve
learningCurve <- createLearningCurvePar(population,
                                        plpData,
                                        modelSettings)
# plot learning curve
plotLearningCurve(learningCurve)
}

}
