% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportExport.R
\name{createLrSql}
\alias{createLrSql}
\title{Convert logistic regression model to sql code...}
\usage{
createLrSql(models, modelNames, covariateConstructionName = "prediction",
  modelTable = "#model_table", analysisId = 111, e = environment(),
  databaseOutput = NULL)
}
\arguments{
\item{models}{A trianed plp model.}

\item{modelNames}{A name used in the covariate function names (no spaces)}

\item{covariateConstructionName}{the name used for the create covariate function}

\item{modelTable}{The temporary table name storing the model details}

\item{analysisId}{The covariate analysis_id}

\item{e}{The environment to output the covariate setting functions to}

\item{databaseOutput}{If you want to output to go inot a cohort table add the "database.schema.tablename" here}
}
\description{
Convert logistic regression model to sql code...
}
\details{
This function is used to create custom covariates for a logistic regression model
(currently only supports, demographics/conditions/drug/procedures/observations and measurement concepts)
}
