% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePlpDocument.R
\name{createPlpJournalDocument}
\alias{createPlpJournalDocument}
\title{createPlpJournalDocument}
\usage{
createPlpJournalDocument(plpResult = NULL, plpValidation = NULL,
  plpData = NULL, targetName = "<target population>",
  outcomeName = "<outcome>", table1 = F, connectionDetails = NULL,
  includeTrain = FALSE, includeTest = TRUE,
  includePredictionPicture = TRUE, includeAttritionPlot = TRUE,
  outputLocation = file.path(getwd(), "plp_journal_document.docx"),
  save = T)
}
\arguments{
\item{plpResult}{An object of type \code{plpResult} returned by running runPlp()}

\item{plpValidation}{An object of type \code{validatePlp} returned by running externalValidatePlp()}

\item{plpData}{The plpData}

\item{targetName}{A string with the target description name}

\item{outcomeName}{A string with the outcome description name}

\item{table1}{Whether to include table1 (characteristics)}

\item{connectionDetails}{The connection required to calcualte the characteristics}

\item{includeTrain}{Whether to include the train set performance}

\item{includeTest}{Whether to include the test set performance}

\item{includePredictionPicture}{Whether to include a picture detailing the prediction problem}

\item{includeAttritionPlot}{Whether to include the attriction plot}

\item{outputLocation}{The location to write the document to}

\item{save}{If false this fucntion returns the document and does not save to outputLocation}
}
\value{
A work document containing the selected outputs within the user's directory at location specified in outputLocation
}
\description{
Creates a template for a prediction journal paper with the characteristics/results filled in
}
\details{
The function creates a word document containing the analysis details, data summary and prediction model results.
}
