% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering.R
\name{createRandomForestFeatureSelection}
\alias{createRandomForestFeatureSelection}
\title{Create the settings for random foreat based feature selection}
\usage{
createRandomForestFeatureSelection(ntrees = 2000, maxDepth = 17)
}
\arguments{
\item{ntrees}{Number of tree in forest}

\item{maxDepth}{Max depth of each tree}
}
\value{
An object of class \code{featureEngineeringSettings}
}
\description{
Create the settings for random foreat based feature selection
}
\details{
Returns an object of class \code{featureEngineeringSettings} that specifies the sampling function that will be called and the settings
}
