% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering.R
\name{createRareFeatureRemover}
\alias{createRareFeatureRemover}
\title{Create the settings for removing rare features}
\usage{
createRareFeatureRemover(threshold = 0.001)
}
\arguments{
\item{threshold}{The minimum fraction of the training data that must have a
feature for it to be included}
}
\value{
An object of class \code{featureEngineeringSettings}
}
\description{
Create the settings for removing rare features
}
