% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData.R
\name{createRestrictPlpDataSettings}
\alias{createRestrictPlpDataSettings}
\title{createRestrictPlpDataSettings define extra restriction settings when calling getPlpData}
\usage{
createRestrictPlpDataSettings(
  studyStartDate = "",
  studyEndDate = "",
  firstExposureOnly = F,
  washoutPeriod = 0,
  sampleSize = NULL
)
}
\arguments{
\item{studyStartDate}{A calendar date specifying the minimum date that a cohort index
date can appear. Date format is 'yyyymmdd'.}

\item{studyEndDate}{A calendar date specifying the maximum date that a cohort index
date can appear. Date format is 'yyyymmdd'. Important: the study
end data is also used to truncate risk windows, meaning no outcomes
beyond the study end date will be considered.}

\item{firstExposureOnly}{Should only the first exposure per subject be included? Note that
this is typically done in the \code{createStudyPopulation} function,
but can already be done here for efficiency reasons.}

\item{washoutPeriod}{The mininum required continuous observation time prior to index
date for a person to be included in the at risk cohort. Note that
this is typically done in the \code{createStudyPopulation} function,
but can already be done here for efficiency reasons.}

\item{sampleSize}{If not NULL, the number of people to sample from the target cohort}
}
\value{
A setting object of class \code{restrictPlpDataSettings} containing a list getPlpData extra settings
}
\description{
This function creates the settings used to restrict the target cohort when calling getPlpData
}
\details{
Users need to specify the extra restrictions to apply when downloading the target cohort
}
