% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering.R
\name{createStratifiedImputationSettings}
\alias{createStratifiedImputationSettings}
\title{Create the settings for adding a spline for continuous variables}
\usage{
createStratifiedImputationSettings(covariateId, ageSplits = NULL)
}
\arguments{
\item{covariateId}{The covariateId that needs imputed values}

\item{ageSplits}{A vector of age splits in years to create age groups}
}
\value{
An object of class \code{featureEngineeringSettings}
}
\description{
Create the settings for adding a spline for continuous variables
}
\details{
Returns an object of class \code{featureEngineeringSettings} that specifies how to do stratified imputation
}
