% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMultiplePlp.R
\name{createStudyPopulationSettings}
\alias{createStudyPopulationSettings}
\title{create the study population settings}
\usage{
createStudyPopulationSettings(binary = T, includeAllOutcomes = T,
  firstExposureOnly = FALSE, washoutPeriod = 0,
  removeSubjectsWithPriorOutcome = TRUE, priorOutcomeLookback = 99999,
  requireTimeAtRisk = T, minTimeAtRisk = 364, riskWindowStart = 1,
  addExposureDaysToStart = FALSE, riskWindowEnd = 365,
  addExposureDaysToEnd = F, verbosity = "INFO")
}
\arguments{
\item{binary}{Forces the outcomeCount to be 0 or 1 (use for binary prediction problems)}

\item{includeAllOutcomes}{(binary) indicating whether to include people with outcomes who are not observed for the whole at risk period}

\item{firstExposureOnly}{Should only the first exposure per subject be included? Note that
this is typically done in the \code{createStudyPopulation} function,}

\item{washoutPeriod}{The mininum required continuous observation time prior to index
date for a person to be included in the cohort.}

\item{removeSubjectsWithPriorOutcome}{Remove subjects that have the outcome prior to the risk window start?}

\item{priorOutcomeLookback}{How many days should we look back when identifying prior outcomes?}

\item{requireTimeAtRisk}{Should subject without time at risk be removed?}

\item{minTimeAtRisk}{The minimum number of days at risk required to be included}

\item{riskWindowStart}{The start of the risk window (in days) relative to the index date (+
days of exposure if the \code{addExposureDaysToStart} parameter is
specified).}

\item{addExposureDaysToStart}{Add the length of exposure the start of the risk window?}

\item{riskWindowEnd}{The end of the risk window (in days) relative to the index data (+
days of exposure if the \code{addExposureDaysToEnd} parameter is
specified).}

\item{addExposureDaysToEnd}{Add the length of exposure the risk window?}

\item{verbosity}{Sets the level of the verbosity. If the log level is at or higher in priority than the logger threshold, a message will print. The levels are:
\itemize{
\item{DEBUG}{Highest verbosity showing all debug statements}
\item{TRACE}{Showing information about start and end of steps}
\item{INFO}{Show informative information (Default)}
\item{WARN}{Show warning messages}
\item{ERROR}{Show error messages}
\item{FATAL}{Be silent except for fatal errors} 
}#'}
}
\value{
A list containing all the settings required for creating the study population
}
\description{
create the study population settings
}
\details{
Takes as input the inputs to create study population
}
