% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering.R
\name{createUnivariateFeatureSelection}
\alias{createUnivariateFeatureSelection}
\title{Create the settings for defining any feature selection that will be done}
\usage{
createUnivariateFeatureSelection(k = 100)
}
\arguments{
\item{k}{This function returns the K features most associated (univariately) to the outcome}
}
\value{
An object of class \code{featureEngineeringSettings}
}
\description{
Create the settings for defining any feature selection that will be done
}
\details{
Returns an object of class \code{featureEngineeringSettings} that specifies the sampling function that will be called and the settings
}
