% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalValidatePlp.R
\name{createValidationDesign}
\alias{createValidationDesign}
\title{createValidationDesign - Define the validation design for external validation}
\usage{
createValidationDesign(
  targetId,
  outcomeId,
  populationSettings,
  restrictPlpDataSettings,
  plpModelList,
  recalibrate = NULL,
  runCovariateSummary = TRUE
)
}
\arguments{
\item{targetId}{The targetId of the target cohort to validate on}

\item{outcomeId}{The outcomeId of the outcome cohort to validate on}

\item{populationSettings}{A list of population restriction settings created by \code{createPopulationSettings}}

\item{restrictPlpDataSettings}{A list of plpData restriction settings created by \code{createRestrictPlpDataSettings}}

\item{plpModelList}{A list of plpModels objects created by \code{runPlp} or a path to such objects}

\item{recalibrate}{A vector of characters specifying the recalibration method to apply,}

\item{runCovariateSummary}{whether to run the covariate summary for the validation data}
}
\description{
createValidationDesign - Define the validation design for external validation
}
