% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{diagnostic}
\alias{diagnostic}
\title{diagnostic - Investigates the prediction problem settings - use before training a model}
\usage{
diagnostic(
  plpData = NULL,
  cdmDatabaseName = "none",
  cohortName,
  outcomeNames,
  databaseDetails,
  restrictPlpDataSettings,
  populationSettings,
  outputFolder = NULL,
  minCellCount = 5
)
}
\arguments{
\item{plpData}{The data object to do the diagnostic on - if NULL you need to specify the connection settings below}

\item{cdmDatabaseName}{The name of the database being diagnosed}

\item{cohortName}{Name of the target cohort}

\item{outcomeNames}{Vector of outcome names}

\item{databaseDetails}{(only used is plpData is NULL) The database details created using \code{createDatabaseDetails}}

\item{restrictPlpDataSettings}{(only used is plpData is NULL) The restrictPlpDataSettings created using \code{createRestrictPlpDataSettings}}

\item{populationSettings}{The population setting details created using \code{createPopulationSettings}}

\item{outputFolder}{Location to save results for shiny app}

\item{minCellCount}{The minimum count that will be displayed}
}
\value{
An object containing the model or location where the model is save, the data selection settings, the preprocessing
and training settings as well as various performance measures obtained by the model.

\item{distribution}{list for each O of a data.frame containing: i) Time to observation end distribution, ii) Time from observation start distribution, iii) Time to event distribution and iv) Time from last prior event to index distribution (only for patients in T who have O before index) }
\item{incident}{list for each O of incidence of O in T during TAR}
\item{characterization}{list for each O of Characterization of T, TnO, Tn~O}
}
\description{
This function runs a set of prediction diagnoses to help pick a suitable T, O, TAR and determine 
whether the prediction problem is worth executing.
}
\details{
Users can define set of Ts, Os, databases and population settings.  A list of data.frames containing details such as
follow-up time distribution, time-to-event information, characteriszation details, time from last prior event, 
observation time distribution.
}
\examples{
\dontrun{
#******** EXAMPLE 1 ********* 
} 
}
