% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalValidatePlp.R
\name{evaluateExistingModel}
\alias{evaluateExistingModel}
\title{evaluateExistingModel}
\usage{
evaluateExistingModel(modelTable, covariateTable, interceptTable = NULL,
  type = "score", covariateSettings, customCovariates = NULL,
  riskWindowStart = 1, addExposureDaysToEnd = F, riskWindowEnd = 365,
  requireTimeAtRisk = T, minTimeAtRisk = 364, includeAllOutcomes = T,
  removeSubjectsWithPriorOutcome = T, connectionDetails, cdmDatabaseSchema,
  cohortDatabaseSchema, cohortTable, cohortId, outcomeDatabaseSchema,
  outcomeTable, outcomeId, oracleTempSchema = cdmDatabaseSchema,
  modelName = "existingModel", calibrationPopulation = NULL)
}
\arguments{
\item{modelTable}{The model covariates and scores}

\item{covariateTable}{The mapping from model covariates to standard covariates}

\item{interceptTable}{The model intercepts}

\item{type}{Model type (score or logistic)}

\item{covariateSettings}{The standard covariate settings (specify covariate lookback time)}

\item{customCovariates}{A table of covariateId, sql (sql creates the custom covariate)}

\item{riskWindowStart}{The day after index to start predicting the outcome}

\item{addExposureDaysToEnd}{riskWindomEnd relative to the cohort end date instead of the cohort start date?}

\item{riskWindowEnd}{The day after index to stop predicting the outcome}

\item{requireTimeAtRisk}{Do you want to ignore people who leave the database some point between the riskWindowStart and riskWindowEnd}

\item{minTimeAtRisk}{If requireTimeAtRisk is TRUE, how many days must they be observed before leaving to get included (default recommendation is all risk period: riskWindowEnd-riskWindowStart)}

\item{includeAllOutcomes}{Setting this to TRUE means people with the outcome who leave the data during the risk period are still included, so only non-outcome people who leave during the risk period are removed}

\item{removeSubjectsWithPriorOutcome}{Remove people from the target population if they have the outcome prior to target cohort start date}

\item{connectionDetails}{The details to connect to the CDM}

\item{cdmDatabaseSchema}{A string specifying the database containing the cdm}

\item{cohortDatabaseSchema}{A string specifying the database containing the target cohorts}

\item{cohortTable}{A string specifying the table containing the target cohorts}

\item{cohortId}{An iteger specifying the cohort id for the target cohorts}

\item{outcomeDatabaseSchema}{A string specifying the database containing the outcome cohorts}

\item{outcomeTable}{A string specifying the table containing the outcome cohorts}

\item{outcomeId}{An iteger specifying the cohort id for the outcome cohorts}

\item{oracleTempSchema}{The temp oracle schema}

\item{modelName}{The name of the model}

\item{calibrationPopulation}{A data.frame of subjectId, cohortStartDate, indexes used to recalibrate the model on new data}
}
\value{
The performance of the existing model and prediction
}
\description{
This function implements an existing model
}
\details{
Implements an existing model and evaluates its performance
}
