% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{evaluatePlp}
\alias{evaluatePlp}
\title{evaluatePlp}
\usage{
evaluatePlp(prediction, typeColumn = "evaluationType")
}
\arguments{
\item{prediction}{The patient level prediction model's prediction}

\item{typeColumn}{The column name in the prediction object that is used to
stratify the evaluation}
}
\value{
An object of class plpEvaluation containing the following components
- evaluationStatistics: A data frame containing the evaluation statistics'
- thresholdSummary: A data frame containing the threshold summary'
- demographicSummary: A data frame containing the demographic summary'
- calibrationSummary: A data frame containing the calibration summary'
- predictionDistribution: A data frame containing the prediction distribution'
}
\description{
Evaluates the performance of the patient level prediction model
}
\details{
The function calculates various metrics to measure the performance of the model
}
