% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalValidatePlp.R
\name{externalValidateDbPlp}
\alias{externalValidateDbPlp}
\title{externalValidateDbPlp - Validate a model on new databases}
\usage{
externalValidateDbPlp(
  plpModel,
  validationDatabaseDetails = createDatabaseDetails(),
  validationRestrictPlpDataSettings = createRestrictPlpDataSettings(),
  settings = createValidationSettings(recalibrate = "weakRecalibration"),
  logSettings = createLogSettings(verbosity = "INFO", logName = "validatePLP"),
  outputFolder = getwd()
)
}
\arguments{
\item{plpModel}{The model object returned by runPlp() containing the trained model}

\item{validationDatabaseDetails}{A list of objects of class \code{databaseDetails} created using \code{createDatabaseDetails}}

\item{validationRestrictPlpDataSettings}{A list of population restriction settings created by \code{createRestrictPlpDataSettings()}}

\item{settings}{A settings object of class \code{validationSettings} created using \code{createValidationSettings}}

\item{logSettings}{An object of \code{logSettings} created using \code{createLogSettings} 
specifying how the logging is done}

\item{outputFolder}{The directory to save the validation results to (subfolders are created per database in validationDatabaseDetails)}
}
\value{
A list containing the performance for each validation_schema
}
\description{
This function extracts data using a user specified connection and cdm_schema, applied the model and then calcualtes the performance
}
\details{
Users need to input a trained model (the output of runPlp()) and new database connections. The function will return a list of length equal to the 
number of cdm_schemas input with the performance on the new data
}
