% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalValidatePlp.R
\name{externalValidatePlp}
\alias{externalValidatePlp}
\title{externalValidatePlp - Validate a model on new databases}
\usage{
externalValidatePlp(plpResult, connectionDetails, validation_schema_target,
  validation_schema_outcome, validation_schema_cdm,
  validation_table_target = "cohort", validation_table_outcome = "cohort",
  validation_id_target = NULL, validation_id_outcome = NULL,
  verbosity = futile.logger::INFO, keepPrediction = F)
}
\arguments{
\item{plpResult}{The object returned by runPlp() containing the trained model}

\item{connectionDetails}{The connection details for extracting the new data}

\item{validation_schema_target}{A string or list of strings specifying the database containing the target cohorts}

\item{validation_schema_outcome}{A string or list of strings specifying the database containing the outcome cohorts}

\item{validation_schema_cdm}{A string or list of strings specifying the database containing the cdm}

\item{validation_table_target}{A string or list of strings specifying the table containing the target cohorts}

\item{validation_table_outcome}{A string or list of strings specifying the table containing the outcome cohorts}

\item{validation_id_target}{An iteger or list of integers specifying the cohort id for the target cohorts}

\item{validation_id_outcome}{An iteger or list of integers specifying the cohort id for the outcome cohorts}

\item{verbosity}{Sets the level of the verbosity. If the log level is at or higher in priority than the logger threshold, a message will print. The levels are:
\itemize{
\item{DEBUG}{Highest verbosity showing all debug statements}
\item{TRACE}{Showing information about start and end of steps}
\item{INFO}{Show informative information (Default)}
\item{WARN}{Show warning messages}
\item{ERROR}{Show error messages}
\item{FATAL}{Be silent except for fatal errors}
}}

\item{keepPrediction}{Whether to keep the predicitons for the new data}
}
\value{
A list containing the performance for each validation_schema
}
\description{
This function extracts data using a user specified connection and cdm_schema, applied the model and then calcualtes the performance
}
\details{
Users need to input a trained model (the output of runPlp()) and new database connections. The function will return a list of length equal to the 
number of cdm_schemas input with the performance on the new data
}
