% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{extractDatabaseToCsv}
\alias{extractDatabaseToCsv}
\title{Exports all the results from a database into csv files}
\usage{
extractDatabaseToCsv(
  conn = NULL,
  connectionDetails,
  databaseSchemaSettings = createDatabaseSchemaSettings(resultSchema = "main"),
  csvFolder,
  fileAppend = NULL
)
}
\arguments{
\item{conn}{The connection to the database with the results}

\item{connectionDetails}{The connectionDetails for the result database}

\item{databaseSchemaSettings}{The result database schema settings}

\item{csvFolder}{Location to save the csv files}

\item{fileAppend}{If set to a string this will be appended to the start of the csv file names}
}
\description{
Exports all the results from a database into csv files
}
\details{
Extracts the results from a database into a set of csv files
}
