% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{fitPlp}
\alias{fitPlp}
\title{fitPlp}
\usage{
fitPlp(trainData, modelSettings, search = "grid", analysisId)
}
\arguments{
\item{trainData}{An object of type \code{TrainData} created using \code{splitData}
data extracted from the CDM.}

\item{modelSettings}{An object of class \code{modelSettings} created using one of the function:
\itemize{
\item{logisticRegressionModel()}{ A lasso logistic regression model}
\item{GBMclassifier()}{ A gradient boosting machine}
\item{RFclassifier()}{ A random forest model}
\item{GLMclassifier ()}{ A generalised linear model}
\item{KNNclassifier()}{ A KNN model}
}}

\item{search}{The search strategy for the hyper-parameter selection (currently not used)}

\item{analysisId}{The id of the analysis}
}
\value{
An object of class \code{plpModel} containing:

\item{model}{The trained prediction model}
\item{modelLoc}{The path to where the model is saved (if saved)}
\item{trainAuc}{The AUC obtained on the training set}
\item{trainCalibration}{The calibration obtained on the training set}
\item{modelSettings}{A list specifiying the model, preprocessing, outcomeId and cohortId}
\item{metaData}{The model meta data}
\item{trainingTime}{The time taken to train the classifier}
}
\description{
Train various models using a default parameter gird search or user specified parameters
}
\details{
The user can define the machine learning model to train (regularised logistic regression, random forest,
gradient boosting machine, neural network and )
}
