% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalValidatePlp.R
\name{getData}
\alias{getData}
\title{getData - Get the plpData for the validation}
\usage{
getData(design, database, outputFolder, downloadTasks)
}
\arguments{
\item{design}{The validationDesign object}

\item{database}{The databaseDetails object}

\item{outputFolder}{The directory to save the validation results to}

\item{downloadTasks}{A list of download tasks determined by unique
combinations of targetId and restrictPlpDataSettings}
}
\value{
The plpData object
}
\description{
getData - Get the plpData for the validation
}
\keyword{internal}
