% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemographicSummary.R
\name{getDemographicSummary}
\alias{getDemographicSummary}
\title{Get a demographic summary}
\usage{
getDemographicSummary(prediction, predictionType, typeColumn = "evaluation")
}
\arguments{
\item{prediction}{A prediction object}

\item{predictionType}{The type of prediction (binary or survival)}

\item{typeColumn}{A column that is used to stratify the results}
}
\value{
A dataframe with the demographic summary
}
\description{
Get a demographic summary
}
\details{
Generates a data.frame with a prediction summary per each 5 year age group 
and gender group
}
