% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportExport.R
\name{getPlpTable}
\alias{getPlpTable}
\title{Create a dataframe with the summary details of the population cohort for publications}
\usage{
getPlpTable(plpData, population, connectionDetails,
  cohortTable = "#temp_person")
}
\arguments{
\item{plpData}{An object returned by running \code{getPlpData}.}

\item{population}{The population you want the summary table for}

\item{connectionDetails}{The connection details used to connect to the CDM database}

\item{cohortTable}{The name of the temp table that will store the popualtion cohort}
}
\description{
Create a dataframe with the summary details of the population cohort for publications
}
\details{
This function is used to create a summary table for population to be inserted into publications
}
\examples{
\dontrun{
getTable1 (plpData, population, connectionDetails)
}
}
