% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePlpDocument.R
\name{getPlpTable}
\alias{getPlpTable}
\title{Create a dataframe with the summary details of the population cohort for publications}
\usage{
getPlpTable(cdmDatabaseSchema, oracleTempSchema, covariateSettings,
  longTermStartDays = -365, population, connectionDetails,
  cohortTable = "#temp_person")
}
\arguments{
\item{cdmDatabaseSchema}{The schema containing the OMOP CDM data}

\item{oracleTempSchema}{The oracle schema if needed}

\item{covariateSettings}{The covariateSettings if different from default}

\item{longTermStartDays}{How far to look back when looking for the variables in the data}

\item{population}{The population you want the summary table for}

\item{connectionDetails}{The connection details used to connect to the CDM database}

\item{cohortTable}{The name of the temp table that will store the popualtion cohort}
}
\description{
Create a dataframe with the summary details of the population cohort for publications
}
\details{
This function is used to create a summary table for population to be inserted into publications
}
\examples{
\dontrun{
getTable1 (plpData, population, connectionDetails)
}
}
