% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalValidatePlp.R
\name{getPopulation}
\alias{getPopulation}
\title{getPopulation - Get the population for the validationDesign}
\usage{
getPopulation(validationDesign, modelDesigns, plpData)
}
\description{
getPopulation - Get the population for the validationDesign
}
