% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{getPredictionDistribution}
\alias{getPredictionDistribution}
\title{Calculates the prediction distribution}
\usage{
getPredictionDistribution(prediction)
}
\arguments{
\item{prediction}{A prediction object as generated using the
\code{\link{predictProbabilities}} function.}
}
\value{
The 0.00, 0.1, 0.25, 0.5, 0.75, 0.9, 1.00 quantile pf the prediction,
the mean and standard deviation per class
}
\description{
Calculates the prediction distribution
}
\details{
Calculates the quantiles from a predition object
}
