% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePlp.R
\name{getThresholdSummary}
\alias{getThresholdSummary}
\title{Calculate all measures for sparse ROC}
\usage{
getThresholdSummary(prediction)
}
\arguments{
\item{prediction}{A prediction object as generated using the
\code{\link{predictProbabilities}} function.}
}
\value{
A data.frame with all the measures
}
\description{
Calculate all measures for sparse ROC
}
\details{
Calculates the TP, FP, TN, FN, TPR, FPR, accuracy, PPF, FOR and Fmeasure
from a predition object
}
