% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationSummary.R
\name{ici}
\alias{ici}
\title{Calculate the Integrated Calibration Information from Austin and Steyerberg
https://onlinelibrary.wiley.com/doi/full/10.1002/sim.8281}
\usage{
ici(prediction)
}
\arguments{
\item{prediction}{the prediction object found in the plpResult object}
}
\value{
Integrated Calibration Information
}
\description{
Calculate the Integrated Calibration Information from Austin and Steyerberg
https://onlinelibrary.wiley.com/doi/full/10.1002/sim.8281
}
\details{
Calculate the Integrated Calibration Information
}
