% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{listAppend}
\alias{listAppend}
\title{join two lists}
\usage{
listAppend(a, b)
}
\arguments{
\item{a}{A list}

\item{b}{Another list}
}
\description{
join two lists
}
\details{
This function joins two lists
}
