% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetPlpData.R
\name{loadPlpData}
\alias{loadPlpData}
\title{Load the PatientLevelPrediction data from a folder}
\usage{
loadPlpData(file, readOnly = FALSE)
}
\arguments{
\item{file}{The name of the folder containing the data.}

\item{readOnly}{If true, the data is opened read only.}
}
\value{
An object of class PlPData
}
\description{
\code{loadPlPData} loads an object of type \code{plpData} from a folder in the file system.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}
\examples{
# todo
}

