% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{loadPlpModel}
\alias{loadPlpModel}
\title{loads the plp model}
\usage{
loadPlpModel(dirPath)
}
\arguments{
\item{dirPath}{The location of the model}
}
\description{
loads the plp model
}
\details{
Loads a plp model that was saved using \code{savePlpModel()}
}
