% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{loadPlpResult}
\alias{loadPlpResult}
\title{Loads the evalaution dataframe}
\usage{
loadPlpResult(dirPath)
}
\arguments{
\item{dirPath}{The directory where the evaluation was saved}
}
\description{
Loads the evalaution dataframe
}
\details{
Loads the evaluation
}
