% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{loadPlpShareable}
\alias{loadPlpShareable}
\title{Loads the plp result saved as json/csv files for transparent sharing}
\usage{
loadPlpShareable(loadDirectory)
}
\arguments{
\item{loadDirectory}{The directory with the results as json/csv files}
}
\description{
Loads the plp result saved as json/csv files for transparent sharing
}
\details{
Load the main results from json/csv files into a runPlp object
}
