% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{loadPrediction}
\alias{loadPrediction}
\title{Loads the prediciton dataframe to csv}
\usage{
loadPrediction(fileLocation)
}
\arguments{
\item{fileLocation}{The location with the saved prediction}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{                            The prediction data.frame
}\if{html}{\out{</div>}}
}
\description{
Loads the prediciton dataframe to csv
}
\details{
Loads the prediciton  RDS file
}
