% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{loadPrediction}
\alias{loadPrediction}
\title{Loads the prediction dataframe to json}
\usage{
loadPrediction(fileLocation)
}
\arguments{
\item{fileLocation}{The location with the saved prediction}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{                            The prediction data.frame
}\if{html}{\out{</div>}}
}
\description{
Loads the prediction dataframe to json
}
\details{
Loads the prediciton json file
}
\examples{
\donttest{ \dontshow{ # takes too long }
data("simulationProfile")
plpData <- simulatePlpData(simulationProfile, n = 1000)
saveLoc <- file.path(tempdir(), "loadPrediction")
results <- runPlp(plpData, outcomeId = 3, saveDirectory = saveLoc)
savePrediction(results$prediction, saveLoc)
dir(saveLoc)
loadedPrediction <- loadPrediction(file.path(saveLoc, "prediction.json"))
}
}
