% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{loadPrediction}
\alias{loadPrediction}
\title{Loads the prediciton dataframe to csv}
\usage{
loadPrediction(location)
}
\arguments{
\item{location}{The directory to saved the csv}
}
\description{
Loads the prediciton dataframe to csv
}
\details{
Loads the prediciton  csv file
}

