% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plotF1Measure}
\alias{plotF1Measure}
\title{Plot the F1 measure efficiency frontier using the sparse thresholdSummary data frame}
\usage{
plotF1Measure(evaluation, type = "test", fileName = NULL)
}
\arguments{
\item{evaluation}{A prediction object as generated using the
\code{\link{runPlp}} function.}

\item{type}{options: 'train' or test'}

\item{fileName}{Name of the file where the plot should be saved, for example
'plot.png'. See the function \code{ggsave} in the ggplot2 package for
supported file formats.}
}
\value{
A ggplot object. Use the \code{\link[ggplot2]{ggsave}} function to save to file in a different
format.
}
\description{
Plot the F1 measure efficiency frontier using the sparse thresholdSummary data frame
}
\details{
Create a plot showing the F1 measure efficiency frontier using the sparse thresholdSummary data frame
}
