% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plotLearningCurve}
\alias{plotLearningCurve}
\title{Plot the plearning curve using the data frame}
\usage{
plotLearningCurve(learningCurve, title = " ", xlabel = "Training Size",
  fileName = NULL)
}
\arguments{
\item{learningCurve}{A results object as generated using the
\code{\link{createLearningCurve}} function.}

\item{title}{plot title}

\item{xlabel}{The label informaiton of plot x-axis}

\item{fileName}{Name of the file where the plot should be saved, for example
'plot.png'. See the function \code{ggsave} in the ggplot2 package for
supported file formats.}
}
\value{
A ggplot object. Use the \code{\link[ggplot2]{ggsave}} function to save to file in a different
format.
}
\description{
Plot the plearning curve using the data frame
}
\details{
Create a plot showing thelearnign curve using output from createLearningCurve
}
