% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plotLearningCurve}
\alias{plotLearningCurve}
\title{plotLearningCurve}
\usage{
plotLearningCurve(learningCurve, metric = "AUROC",
  abscissa = "observations", plotTitle = "Learning Curve",
  plotSubtitle = NULL, fileName = NULL)
}
\arguments{
\item{learningCurve}{An object returned by \code{\link{createLearningCurve}}
function.}

\item{metric}{Specifies the metric to be plotted:
\itemize{
  \item{\code{'AUROC'} - use the area under the Receiver Operating
    Characteristic curve}
  \item{\code{'AUPRC'} - use the area under the Precision-Recall curve}
  \item{\code{'sBrier'} - use the scaled Brier score}
}}

\item{abscissa}{Specify the abscissa metric to be plotted:
\itemize{
  \item{\code{'observations'} - use number of observations}
  \item{\code{'outcomes'} - use number of positive outcomes}
}}

\item{plotTitle}{Title of the learning curve plot.}

\item{plotSubtitle}{Subtitle of the learning curve plot.}

\item{fileName}{Filename of plot to be saved, for example \code{'plot.png'}.
See the function \code{ggsave} in the ggplot2 package for supported file 
formats.}
}
\value{
A ggplot object. Use the \code{\link[ggplot2]{ggsave}} function to save to 
file in a different format.
}
\description{
Create a plot of the learning curve using the object returned
from \code{createLearningCurve}.
}
\examples{
\dontrun{
# create learning curve object
learningCurve <- createLearningCurve(population,
                                     plpData,
                                     modelSettings)
# plot the learning curve
plotLearningCurve(learningCurve)
}

}
