% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plotNetBenefit}
\alias{plotNetBenefit}
\title{Plot the net benefit}
\usage{
plotNetBenefit(
  plpResult,
  typeColumn = "evaluation",
  saveLocation = NULL,
  fileName = "netBenefit.png",
  evalType = NULL,
  ylim = NULL,
  xlim = NULL
)
}
\arguments{
\item{plpResult}{A plp result object as generated using the \code{\link{runPlp}} function.}

\item{typeColumn}{The name of the column specifying the evaluation type}

\item{saveLocation}{Directory to save plot (if NULL plot is not saved)}

\item{fileName}{Name of the file to save to plot, for example 'plot.png'. See the function \code{ggsave} in the ggplot2 package for supported file formats.}

\item{evalType}{Which evaluation type to plot for. For example `Test`, `Train`. If NULL everything is plotted}

\item{ylim}{The y limits for the plot, if NULL the limits are calculated from the data}

\item{xlim}{The x limits for the plot, if NULL the limits are calculated from the data}
}
\value{
A list of ggplot objects
}
\description{
Plot the net benefit
}
