% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plotPlp}
\alias{plotPlp}
\title{Plot all the PatientLevelPrediction plots}
\usage{
plotPlp(result, filename, type = "test")
}
\arguments{
\item{result}{Object returned by the runPlp() function}

\item{filename}{Name of the file where the plot should be saved, for example
'plot.png'. See the function \code{ggsave} in the ggplot2 package for
supported file formats.}

\item{type}{Evaluation data type either 'test', 'val' or 'train'}
}
\value{
TRUE if it ran
}
\description{
Plot all the PatientLevelPrediction plots
}
\details{
Create a directory with all the plots
}
