% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plotSmoothCalibration}
\alias{plotSmoothCalibration}
\title{Plot the smooth calibration as detailed in Calster et al. "A calibration heirarchy for risk models
was defined: from utopia to empirical data" (2016)}
\usage{
plotSmoothCalibration(result, smooth = c("loess", "rcs"), span = 1,
  nKnots = 5, scatter = F, type = "test", bins = 20, zoom = c("none",
  "deciles", "data"), fileName = NULL)
}
\arguments{
\item{result}{The result of running \code{\link{runPlp}} function. An object containing the
model or location where the model is save, the data selection settings, the
preprocessing and training settings as well as various performance measures
obtained by the model.}

\item{smooth}{options: 'loess' or 'rcs'}

\item{span}{This specifies the width of span used for loess. This will allow for faster
computing and lower memory usage.}

\item{nKnots}{The number of knots to be used by the rcs evaluation. Default is 5}

\item{scatter}{plot the decile calibrations as points on the graph. Default is False}

\item{type}{Whether to use train or test data, default is test.}

\item{bins}{The number of bins for the histogram. Default is 20.}

\item{zoom}{Zoom in on the region containing the deciles or on the data. If not specified
shows the entire space.}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'. See the
function \code{ggsave} in the ggplot2 package for supported file formats.}
}
\value{
A cowplot object. Use the \code{cowplot::save_plot} function to save to file in a different format.
}
\description{
Plot the smooth calibration as detailed in Calster et al. "A calibration heirarchy for risk models
was defined: from utopia to empirical data" (2016)
}
\details{
Create a plot showing the smoothed calibration #'
}
