% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PatientLevelPrediction.R
\docType{data}
\name{plpDataSimulationProfile}
\alias{plpDataSimulationProfile}
\title{A simulation profile}
\usage{
data(plpDataSimulationProfile)
}
\description{
A simulation profile
}
\keyword{datasets}

