% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CyclopsModels.R
\name{predictCyclops}
\alias{predictCyclops}
\title{Create predictive probabilities}
\usage{
predictCyclops(plpModel, data, cohort)
}
\arguments{
\item{plpModel}{An object of type \code{predictiveModel} as generated using
\code{\link{fitPlp}}.}

\item{data}{The new plpData containing the covariateData for the new population}

\item{cohort}{The cohort to calculate the prediction for}
}
\value{
The value column in the result data.frame is: logistic: probabilities of the outcome, poisson:
Poisson rate (per day) of the outome, survival: hazard rate (per day) of the outcome.
}
\description{
Create predictive probabilities
}
\details{
Generates predictions for the population specified in plpData given the model.
}
