% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predictPlp}
\alias{predictPlp}
\title{predictPlp}
\usage{
predictPlp(plpModel, plpData, population, timepoint)
}
\arguments{
\item{plpModel}{An object of type \code{plpModel} - a patient level prediction model}

\item{plpData}{An object of type \code{plpData} - the patient level prediction
data extracted from the CDM.}

\item{population}{The population created using createStudyPopulation() who will have their risks predicted or a cohort without the outcome known}

\item{timepoint}{The timepoint to predict risk (survival models only)}
}
\value{
A dataframe containing the prediction for each person in the population with an attribute metaData containing prediction details.
}
\description{
Predict the risk of the outcome using the input plpModel for the input plpData
}
\details{
The function applied the trained model on the plpData to make predictions
}
