% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predictProbabilities}
\alias{predictProbabilities}
\title{Create predictive probabilities}
\usage{
predictProbabilities(predictiveModel, population, covariateData)
}
\arguments{
\item{predictiveModel}{An object of type \code{predictiveModel} as generated using
\code{\link{fitPlp}}.}

\item{population}{The population to calculate the prediction for}

\item{covariateData}{The covariateData containing the covariates for the population}
}
\value{
The value column in the result data.frame is: logistic: probabilities of the outcome, poisson:
Poisson rate (per day) of the outome, survival: hazard rate (per day) of the outcome.
}
\description{
Create predictive probabilities
}
\details{
Generates predictions for the population specified in plpData given the model.
}
