% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Predict.R
\name{predictProbabilities}
\alias{predictProbabilities}
\title{Create predictive probabilities}
\usage{
predictProbabilities(predictiveModel, plpData)
}
\arguments{
\item{predictiveModel}{An object of type \code{predictiveModel} as generated using
\code{\link{fitPredictiveModel}}.}

\item{plpData}{An object of type \code{plpData} as generated using
\code{\link{getDbPlpData}}.}
}
\value{
The value column in the result data.frame is: logistic: probabilities of the outcome, poisson:
Poisson rate (per day) of the outome, survival: hazard rate (per day) of the outcome.
}
\description{
Create predictive probabilities
}
\details{
Generates predictions for the population specified in plpData given the model.
}

