% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Predict.R
\name{predictProbabilities}
\alias{predictProbabilities}
\title{Create predictive probabilities}
\usage{
predictProbabilities(predictiveModel, cohortData, covariateData)
}
\arguments{
\item{predictiveModel}{An object of type \code{predictiveModel} as generated using
\code{\link{fitPredictiveModel}}.}

\item{cohortData}{An object of type \code{cohortData} as generated using
\code{\link{getDbCohortData}}.}

\item{covariateData}{An object of type \code{covariateData} as generated using
\code{\link{getDbCovariateData}}.}
}
\description{
Create predictive probabilities
}
\details{
Note that the cohortData and covariateData objects need to come from the same population.
}

