% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessingData.R
\name{preprocessData}
\alias{preprocessData}
\title{A function that wraps around FeatureExtraction::tidyCovariateData to normalise the data
and remove rare or redundant features}
\usage{
preprocessData(covariateData, preprocessSettings)
}
\arguments{
\item{covariateData}{The covariate part of the training data created by \code{splitData} after being sampled and having 
any required feature engineering}

\item{preprocessSettings}{The settings for the preprocessing created by \code{createPreprocessSettings}}
}
\value{
The data processed
}
\description{
A function that wraps around FeatureExtraction::tidyCovariateData to normalise the data
and remove rare or redundant features
}
\details{
Returns an object of class \code{covariateData} that has been processed
}
