% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData.R
\name{print.summary.plpData}
\alias{print.summary.plpData}
\title{Print a summary.plpData object}
\usage{
\method{print}{summary.plpData}(x, ...)
}
\value{
A message describing the object
}
\description{
Print a summary.plpData object
}
