% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recalibration.R
\name{recalibratePlp}
\alias{recalibratePlp}
\title{recalibratePlp}
\usage{
recalibratePlp(
  prediction,
  analysisId,
  typeColumn = "evaluationType",
  method = c("recalibrationInTheLarge", "weakRecalibration")
)
}
\arguments{
\item{prediction}{A prediction dataframe}

\item{analysisId}{The model analysisId}

\item{typeColumn}{The column name where the strata types are specified}

\item{method}{Method used to recalibrate ('recalibrationInTheLarge' or 'weakRecalibration' )}
}
\value{
A prediction dataframe with the recalibrated predictions added
}
\description{
Recalibrating a model using the recalibrationInTheLarge or weakRecalibration methods
}
\details{
'recalibrationInTheLarge' calculates a single correction factor for the
average predicted risks to match the average observed risks.
'weakRecalibration' fits a glm model to the logit of the predicted risks,
also known as Platt scaling/logistic recalibration.
}
