% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMultiplePlp.R
\name{runPlpAnalyses}
\alias{runPlpAnalyses}
\title{Run a list of predictions}
\usage{
runPlpAnalyses(connectionDetails, cdmDatabaseSchema, cdmDatabaseName,
  oracleTempSchema = cdmDatabaseSchema,
  cohortDatabaseSchema = cdmDatabaseSchema, cohortTable = "cohort",
  outcomeDatabaseSchema = cdmDatabaseSchema, outcomeTable = "cohort",
  cdmVersion = 5, outputFolder = "./PlpOutput", modelAnalysisList,
  cohortIds, cohortNames, outcomeIds, outcomeNames, washoutPeriod = 0,
  maxSampleSize = NULL, minCovariateFraction = 0, normalizeData = T,
  testSplit = "person", testFraction = 0.25, splitSeed = NULL,
  nfold = 3, verbosity = "INFO")
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance. Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{cdmDatabaseName}{A string with a shareable name of the database (this will be shown to OHDSI researchers if the results get transported)}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you
want all temporary tables to be managed. Requires
create/insert permissions to this database.}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where the
target cohorts are available.  Requires read
permissions to this database.}

\item{cohortTable}{The tablename that contains the target cohorts.  Expectation is cohortTable
has format of COHORT table: COHORT_DEFINITION_ID, SUBJECT_ID,
COHORT_START_DATE, COHORT_END_DATE.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where the
data used to define the outcome cohorts is available. Requires read permissions to
this database.}

\item{outcomeTable}{The tablename that contains the outcome cohorts.  Expectation is
outcomeTable has format of COHORT table: COHORT_DEFINITION_ID,
SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE.}

\item{cdmVersion}{Define the OMOP CDM version used: currently support "4" and
"5".}

\item{outputFolder}{Name of the folder where all the outputs will written to.}

\item{modelAnalysisList}{A list of objects of type \code{modelSettings} as created using
the \code{\link{createPlpModelSettings}} function.}

\item{cohortIds}{A vector of cohortIds that specify all the target cohorts}

\item{cohortNames}{A vector of cohortNames corresponding to the cohortIds}

\item{outcomeIds}{A vector of outcomeIds that specify all the outcome cohorts}

\item{outcomeNames}{A vector of outcomeNames corresponding to the outcomeIds}

\item{washoutPeriod}{Minimum number of prior observation days}

\item{maxSampleSize}{Max number of target people to sample from to develop models}

\item{minCovariateFraction}{Any covariate with an incidence less than this value if ignored}

\item{normalizeData}{Whether to normalize the covariates}

\item{testSplit}{How to split into test/train (time or person)}

\item{testFraction}{Fraction of data to use as test set}

\item{splitSeed}{The seed used for the randomization into test/train}

\item{nfold}{Number of folds used to do cross validation}

\item{verbosity}{The logging level}
}
\value{
A data frame with the following columns: \tabular{ll}{ \verb{analysisId} \tab The unique identifier
for a set of analysis choices.\cr \verb{cohortId} \tab The ID of the target cohort populations.\cr
\verb{outcomeId} \tab The ID of the outcomeId.\cr \verb{plpDataFolder} \tab The location where the plpData was saved\cr \verb{studyPopFile} \tab The
name of the file containing the study population \cr \verb{evaluationFolder} \tab The name of file containing the evaluation saved as a csv
 \cr \verb{modelFolder} \tab The name of the file containing the developed model.\cr }
}
\description{
Run a list of predictions
}
\details{
Run a list of predictions for the target cohorts and outcomes of interest. This function will run all
specified predictions, meaning that the total number of outcome
models is `length(cohortIds) * length(outcomeIds) * length(modelAnalysisList)`.
}
