% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMultiplePlp.R
\name{savePlpAnalysesJson}
\alias{savePlpAnalysesJson}
\title{Save the modelDesignList to a json file}
\usage{
savePlpAnalysesJson(
  modelDesignList = list(createModelDesign(targetId = 1, outcomeId = 2, modelSettings =
    setLassoLogisticRegression()), createModelDesign(targetId = 1, outcomeId = 3,
    modelSettings = setLassoLogisticRegression())),
  saveDirectory = NULL
)
}
\arguments{
\item{modelDesignList}{A list of modelDesigns created using \code{createModelDesign()}}

\item{saveDirectory}{The directory to save the modelDesignList settings}
}
\description{
Save the modelDesignList to a json file
}
\details{
This function creates a json file with the modelDesignList saved
}
\examples{
\dontrun{
savePlpAnalysesJson(
modelDesignList = list(
createModelDesign(targetId = 1, outcomeId = 2, modelSettings = setLassoLogisticRegression()), 
createModelDesign(targetId = 1, outcomeId = 3, modelSettings = setLassoLogisticRegression())
),
saveDirectory = 'C:/bestModels'
)
}

}
