% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadingSaving.R
\name{savePlpData}
\alias{savePlpData}
\title{Save the cohort data to folder}
\usage{
savePlpData(plpData, file, envir = NULL)
}
\arguments{
\item{plpData}{An object of type \code{plpData} as generated using
\code{getDbPlpData}.}

\item{file}{The name of the folder where the data will be written. The folder should
not yet exist.}
}
\description{
\code{savePlpData} saves an object of type plpData to folder.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}
\examples{
# todo

}

