% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{savePlpModel}
\alias{savePlpModel}
\title{Saves the plp model}
\usage{
savePlpModel(plpModel, dirPath)
}
\arguments{
\item{plpModel}{A trained classifier returned by running \code{runPlp()$model}}

\item{dirPath}{A location to save the model to}
}
\description{
Saves the plp model
}
\details{
Saves the plp model to a user specificed folder
}
