% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{savePrediction}
\alias{savePrediction}
\title{Saves the prediciton dataframe to csv}
\usage{
savePrediction(prediction, location)
}
\arguments{
\item{prediction}{The prediciton data.frame}

\item{location}{The directory to save the csv}
}
\description{
Saves the prediciton dataframe to csv
}
\details{
Saves the prediciton data frame returned by predict.R to a csv file
}

