% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{savePrediction}
\alias{savePrediction}
\title{Saves the prediction dataframe to csv}
\usage{
savePrediction(prediction, dirPath)
}
\arguments{
\item{prediction}{The prediciton data.frame}

\item{dirPath}{The directory to save the csv}
}
\description{
Saves the prediction dataframe to csv
}
\details{
Saves the prediction data frame returned by predict.R to a csv file
}
