% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadPlp.R
\name{savePrediction}
\alias{savePrediction}
\title{Saves the prediction dataframe to RDS}
\usage{
savePrediction(prediction, dirPath, fileName = "prediction.rds")
}
\arguments{
\item{prediction}{The prediciton data.frame}

\item{dirPath}{The directory to save the prediction RDS}

\item{fileName}{The name of the RDS file that will be saved in dirPath}
}
\description{
Saves the prediction dataframe to RDS
}
\details{
Saves the prediction data frame returned by predict.R to an RDS file and returns the fileLocation where the prediction is saved
}
