% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaBoost.R
\name{setAdaBoost}
\alias{setAdaBoost}
\title{Create setting for AdaBoost with python}
\usage{
setAdaBoost(n_estimators = 50, learning_rate = 1, seed = NULL)
}
\arguments{
\item{n_estimators}{The maximum number of estimators at which boosting is terminated}

\item{learning_rate}{Learning rate shrinks the contribution of each classifier by learning_rate. There is a trade-off between learning_rate and n_estimators.}

\item{seed}{A seed for the model}
}
\description{
Create setting for AdaBoost with python
}
\examples{
\dontrun{
model.adaBoost <- setAdaBoost(size=4, alpha=0.00001, seed=NULL)
}
}
