% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaBoost.R
\name{setAdaBoost}
\alias{setAdaBoost}
\title{Create setting for AdaBoost with python}
\usage{
setAdaBoost(nEstimators = 50, learningRate = 1, seed = NULL)
}
\arguments{
\item{nEstimators}{The maximum number of estimators at which boosting is terminated}

\item{learningRate}{Learning rate shrinks the contribution of each classifier by learningRate.
There is a trade-off between learningRate and nEstimators .}

\item{seed}{A seed for the model}
}
\description{
Create setting for AdaBoost with python
}
\examples{
\dontrun{
model.adaBoost <- setAdaBoost(size = 4, alpha = 1e-05, seed = NULL)
}
}
