% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CyclopsSettings.R
\name{setBrokenAdaptiveRidge}
\alias{setBrokenAdaptiveRidge}
\title{Create setting for Broken Adaptive Ridge Regression using Cyclops}
\usage{
setBrokenAdaptiveRidge(
  initialRidgeVariance = 1,
  seed = NULL,
  includeCovariateIds = c(),
  noShrinkage = c("(Intercept)"),
  penalty = "BIC",
  threads = -1,
  forceIntercept = F,
  upperLimit = 20,
  lowerLimit = 0.01,
  tolerance = 2e-06,
  maxIterations = 3000,
  threshold = 1e-06
)
}
\arguments{
\item{initialRidgeVariance}{Numeric: prior distribution starting variance}

\item{seed}{An option to add a seed when training the model}

\item{includeCovariateIds}{a set of covariate IDS to limit the analysis to}

\item{noShrinkage}{a set of covariates whcih are to be forced to be included in the final model. default is the intercept}

\item{penalty}{Specifies the BAR penalty; possible values are `BIC` or `AIC` or a numeric value}

\item{threads}{An option to set number of threads when training model}

\item{forceIntercept}{Logical: Force intercept coefficient into prior}

\item{upperLimit}{Numeric: Upper prior variance limit for grid-search}

\item{lowerLimit}{Numeric: Lower prior variance limit for grid-search}

\item{tolerance}{Numeric: maximum relative change in convergence criterion from successive iterations to achieve convergence}

\item{maxIterations}{Integer: maximum iterations of Cyclops to attempt before returning a failed-to-converge error}

\item{threshold}{absolute threshold at which to force coefficient to 0}
}
\description{
See paper: https://doi.org/10.1016/j.jspi.2020.12.001
}
