% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNNTorch.R
\name{setCNNTorch}
\alias{setCNNTorch}
\title{Create setting for CNN model with python}
\usage{
setCNNTorch(nbfilters = c(16, 32), epochs = c(20, 50), seed = 0,
  time_window = 12, class_weight = 0)
}
\arguments{
\item{nbfilters}{The number of filters}

\item{epochs}{The number of epochs}

\item{seed}{A seed for the model}

\item{time_window}{A time window to get temporal data}

\item{class_weight}{weight the class in imblanced data}
}
\description{
Create setting for CNN model with python
}
\examples{
\dontrun{
model.cnnTorch <- setCNNTorch()
}
}
