% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNNTorch.R
\name{setCNNTorch}
\alias{setCNNTorch}
\title{Create setting for CNN model with python}
\usage{
setCNNTorch(
  nbfilters = c(16, 32),
  epochs = c(20, 50),
  seed = 0,
  class_weight = 0,
  type = "CNN"
)
}
\arguments{
\item{nbfilters}{The number of filters}

\item{epochs}{The number of epochs}

\item{seed}{A seed for the model}

\item{class_weight}{The class weight used for imbalanced data: 
 0: Inverse ratio between positives and negatives
-1: Focal loss}

\item{type}{It can be normal 'CNN', 'CNN_LSTM', CNN_MLF' with multiple kernels with different kernel size, 
'CNN_MIX', 'ResNet' and 'CNN_MULTI'}
}
\description{
Create setting for CNN model with python
}
\examples{
\dontrun{
model.cnnTorch <- setCNNTorch()
}
}
