% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\name{setDecisionTree}
\alias{setDecisionTree}
\title{Create setting for DecisionTree with python}
\usage{
setDecisionTree(maxDepth = 10, minSamplesSplit = 2, minSamplesLeaf = 10,
  minImpurityDecrease = 10^-7, seed = NULL, classWeight = "None",
  plot = F)
}
\arguments{
\item{maxDepth}{The maximum depth of the tree}

\item{minSamplesSplit}{The minimum samples per split}

\item{minSamplesLeaf}{The minimum number of samples per leaf}

\item{minImpurityDecrease}{Threshold for early stopping in tree growth. A node will split if its impurity is above the threshold, otherwise it is a leaf.}

\item{seed}{The random state seed}

\item{classWeight}{Balance or None}

\item{plot}{Boolean whether to plot the tree (requires python pydotplus module)}
}
\description{
Create setting for DecisionTree with python
}
\examples{
\dontrun{
model.decisionTree <- setDecisionTree(maxDepth=10,minSamplesLeaf=10, seed=NULL )
}
}
