% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeHardThresholding.R
\name{setIterativeHardThresholding}
\alias{setIterativeHardThresholding}
\title{Create setting for lasso logistic regression}
\usage{
setIterativeHardThresholding(
  K = 10,
  penalty = "bic",
  seed = NULL,
  exclude = NULL,
  forceIntercept = F,
  fitBestSubset = FALSE
)
}
\arguments{
\item{K}{The maximum number of non-zero predictors}

\item{penalty}{Specifies the IHT penalty; possible values are `BIC` or `AIC` or a numeric value}

\item{seed}{An option to add a seed when training the model}

\item{exclude}{A vector of numbers or covariateId names to exclude from prior}

\item{forceIntercept}{Logical: Force intercept coefficient into regularization}

\item{fitBestSubset}{Logical: Fit final subset with no regularization}
}
\description{
Create setting for lasso logistic regression
}
\examples{
model.lr <- setLassoLogisticRegression()
}
