% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlmNetModels.R
\name{setLassoGlmNet}
\alias{setLassoGlmNet}
\title{setLassoGlmNet}
\usage{
setLassoGlmNet(
  nlambda = 100,
  nfolds = 3,
  lambda.min.ratio = 0.01,
  parallel = TRUE,
  measure = "default"
)
}
\arguments{
\item{nlambda}{The number of lambda values}

\item{nfolds}{How many cross validation folds to perform, default=3}

\item{lambda.min.ratio}{Smallest value of lambda, as fraction of lambda.max which is data derived 
to have the first estimated model have zero coefficients.}

\item{parallel}{TRUE to use parallelization across folds (cores=nfolds)}

\item{measure}{measure to use for selecting best lambda, one of:
"default", "mse", "deviance", "class", "auc", "mae", "C"}
}
\description{
Create setting for lasso logistic regression model using glmnet implementation
}
