% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRTorch.R
\name{setLRTorch}
\alias{setLRTorch}
\title{Create setting for logistics regression model with python}
\usage{
setLRTorch(w_decay = c(5e-04, 0.005), epochs = c(20, 50, 100),
  seed = NULL, class_weight = 0, autoencoder = FALSE, vae = FALSE)
}
\arguments{
\item{w_decay}{The l2 regularisation}

\item{epochs}{The number of epochs}

\item{seed}{A seed for the model}

\item{class_weight}{The class weight used for imbalanced data: 
 0: Inverse ratio between positives and negatives
-1: Focal loss}

\item{autoencoder}{First learn stakced autoencoder for input features, then train LR on the encoded features.}

\item{vae}{First learn stakced varational autoencoder for input features, then train LR on the encoded features.}
}
\description{
Create setting for logistics regression model with python
}
\examples{
\dontrun{
model.lrTorch <- setLRTorch()
}
}
