% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLPTorch.R
\name{setMLPTorch}
\alias{setMLPTorch}
\title{Create setting for neural network model with python}
\usage{
setMLPTorch(size = c(100, 200), w_decay = c(0.003, 0.005, 0.007),
  epochs = c(20, 50), seed = 0, class_weight = 0)
}
\arguments{
\item{size}{The number of hidden nodes}

\item{w_decay}{The l2 regularisation}

\item{epochs}{The number of epochs}

\item{seed}{A seed for the model}

\item{class_weight}{weight the class in imblanced data}
}
\description{
Create setting for neural network model with python
}
\examples{
\dontrun{
model.mlpTorch <- setMLPTorch()
}
}
