% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveBayes.R
\name{setNaiveBayes}
\alias{setNaiveBayes}
\title{Create setting for naive bayes model with python}
\usage{
setNaiveBayes()
}
\description{
Create setting for naive bayes model with python
}
\examples{
\dontrun{
model.nb <- setNaiveBayes()
}
}
