% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SklearnClassifierSettings.R
\name{setNaiveBayes}
\alias{setNaiveBayes}
\title{Create setting for naive bayes model with python}
\usage{
setNaiveBayes()
}
\value{
a modelSettings object
}
\description{
Create setting for naive bayes model with python
}
\examples{
\dontrun{ \dontshow{ # requires python's scikit-learn, checkSklearn() will error without it}
plpData <- getEunomiaPlpData()
model <- setNaiveBayes()
analysisId <- "naiveBayes"
saveLocation <- file.path(tempdir(), analysisId)
results <- runPlp(plpData, modelSettings = model,
                  saveDirectory = saveLocation,
                  analysisId = analysisId)
# clean up
unlink(saveLocation, recursive = TRUE)
}
}
