% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{setPythonEnvironment}
\alias{setPythonEnvironment}
\title{Use the virtual environment created using configurePython()}
\usage{
setPythonEnvironment(envname = "PLP", envtype = NULL)
}
\arguments{
\item{envname}{A string for the name of the virtual environment (default is 'PLP')}

\item{envtype}{An option for specifying the environment as'conda' or 'python'.  If NULL then the default is 'conda' for windows users and 'python' for non-windows users}
}
\description{
Use the virtual environment created using configurePython()
}
\details{
This function sets PatientLevelPrediction to use a virtual environment
}
