% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SklearnToJson.R
\name{sklearnFromJson}
\alias{sklearnFromJson}
\title{Loads sklearn python model from json}
\usage{
sklearnFromJson(path)
}
\arguments{
\item{path}{path to the model json file}
}
\value{
a sklearn python model object
}
\description{
Loads sklearn python model from json
}
