% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataSplitting.R
\name{splitData}
\alias{splitData}
\title{Split data into random subsets}
\usage{
splitData(plpData, splits = 2)
}
\arguments{
\item{plpData}{An object of type \code{plpData}.}

\item{splits}{This can be either a single integer, in which case the data will be split up into
                 equally sized parts. If a vector is provided instead, these are interpreted as the
                 relative sizes of each part.}
}
\value{
A list with entries for each part. An entry itself is a plpData object.
}
\description{
Split data into random subsets
}
\details{
Splits cohort, covariate, and outcome data into random subsets, to be used for validation.
}

