% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSplitting.R
\name{timeSplitter}
\alias{timeSplitter}
\title{Split test/train data by time and then partitions training set into random folds stratified by
class}
\usage{
timeSplitter(population, test = 0.3, train = NULL, nfold = 3,
  seed = NULL)
}
\arguments{
\item{population}{An object created using createStudyPopulation().}

\item{test}{A real number between 0 and 1 indicating the test set fraction of the data}

\item{train}{A real number between 0 and 1 indicating the training set fraction of the data}

\item{nfold}{An integer >= 1 specifying the number of folds used in cross validation}

\item{seed}{If set a fixed seed is used, otherwise a random split is performed}
}
\value{
A dataframe containing the columns: rowId and index
}
\description{
Split test/train data by time and then partitions training set into random folds stratified by
class
}
\details{
Returns a dataframe of rowIds and indexes with a -1 index indicating the rowId belongs to the test
set and a positive integer index value indicating the rowId's cross valiation fold within the train
set.
}
