% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportExport.R
\name{toPlpData}
\alias{toPlpData}
\title{Convert matrix into plpData}
\usage{
toPlpData(data, columnInfo, outcomeId, outcomeThreshold = 0.5,
  indexTime = 0, includeIndexDay = T)
}
\arguments{
\item{data}{An data.frame or matrix.}

\item{columnInfo}{A dataframe with three columns, column 1 contains columnId, column 2 contains columnName for each column id and column 3 contains the columnTime - the time prior to index the variable was recorded}

\item{outcomeId}{The column id containing the outcome}

\item{outcomeThreshold}{The outcome value must be higher or equal to this for the person to have the outcome}

\item{indexTime}{The time defining the index date}

\item{includeIndexDay}{Boolean - whether to include variables recorded on index date}
}
\value{
Returns an object of class plpData
}
\description{
Converts a matrix (rows = people, columns=variables) into the standard plpData
}
\details{
This function converts matrix into plpData
}
\examples{
#TODO

}
