% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportExport.R
\name{transportPlp}
\alias{transportPlp}
\title{Transports a plpResult to a new location and removed sensitive data}
\usage{
transportPlp(plpResult, modelName = NULL, dataName = NULL, outputFolder,
  n = NULL, includeEvaluationStatistics = T, includeThresholdSummary = T,
  includeDemographicSummary = T, includeCalibrationSummary = T,
  includePredictionDistribution = T, includeCovariateSummary = T,
  save = T)
}
\arguments{
\item{plpResult}{An object returned by running \code{runPlp}.}

\item{modelName}{A string of the name of the model}

\item{dataName}{A string of the name of the data}

\item{outputFolder}{The folder on the file system where the CSV files will be created. If the
folder does not yet exist it will be created.}

\item{n}{The minimum number of people required for each result summary to be included}

\item{includeEvaluationStatistics}{Whether to include the evaluationStatistics}

\item{includeThresholdSummary}{Whether to include the thresholdSummary}

\item{includeDemographicSummary}{Whether to include the demographicSummary}

\item{includeCalibrationSummary}{Whether to include the calibrationSummary}

\item{includePredictionDistribution}{Whether to include the predictionDistribution}

\item{includeCovariateSummary}{Whether to include the covariateSummary}

\item{save}{Whether to save the result or just return the transportable object}
}
\description{
Transports a plpResult to a new location and removed sensitive data
}
\details{
This function is used to
}
\examples{
\dontrun{
transportPlp(plpResult, "s:/temp/exportTest", n=10)
}
}
