% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalValidatePlp.R
\name{validateExternal}
\alias{validateExternal}
\title{externalValidatePlp - Validate model performance on new data}
\usage{
validateExternal(
  validationDesignList,
  databaseDetails,
  logSettings,
  outputFolder
)
}
\arguments{
\item{validationDesignList}{A list of objects created with \code{createValidationDesign}}

\item{databaseDetails}{A list of objects of class
\code{databaseDetails} created using \code{createDatabaseDetails}}

\item{logSettings}{An object of \code{logSettings} created
using \code{createLogSettings}}

\item{outputFolder}{The directory to save the validation results to
(subfolders are created per database in validationDatabaseDetails)}
}
\description{
externalValidatePlp - Validate model performance on new data
}
