% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMultiplePlp.R
\name{validateMultiplePlp}
\alias{validateMultiplePlp}
\title{externally validate the multiple plp models across new datasets}
\usage{
validateMultiplePlp(
  analysesLocation,
  validationDatabaseDetails,
  validationRestrictPlpDataSettings = createRestrictPlpDataSettings(),
  recalibrate = NULL,
  saveDirectory = NULL
)
}
\arguments{
\item{analysesLocation}{The location where the multiple plp analyses are}

\item{validationDatabaseDetails}{The validation database settings created using \code{createDatabaseDetails()}}

\item{validationRestrictPlpDataSettings}{The settings specifying the extra restriction settings when extracting the data created using \code{createRestrictPlpDataSettings()}.}

\item{recalibrate}{A vector of recalibration methods (currently supports 'RecalibrationintheLarge' and/or 'weakRecalibration')}

\item{saveDirectory}{The location to save to validation results}
}
\description{
This function loads all the models in a multiple plp analysis folder and
validates the models on new data
}
\details{
Users need to input a location where the results of the multiple plp analyses
are found and the connection and database settings for the new data
}
