% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewShinyPlp.R
\name{viewDatabaseResultPlp}
\alias{viewDatabaseResultPlp}
\title{open a local shiny app for viewing the result of a PLP analyses from a database}
\usage{
viewDatabaseResultPlp(
  mySchema,
  myServer,
  myUser,
  myPassword,
  myDbms,
  myPort = NULL,
  myTableAppend
)
}
\arguments{
\item{mySchema}{Database result schema containing the result tables}

\item{myServer}{server with the result database}

\item{myUser}{Username for the connection to the result database}

\item{myPassword}{Password for the connection to the result database}

\item{myDbms}{database management system for the result database}

\item{myPort}{Port for the connection to the result database}

\item{myTableAppend}{A string appended to the results tables (optional)}
}
\description{
open a local shiny app for viewing the result of a PLP analyses from a database
}
\details{
Opens a shiny app for viewing the results of the models from a database
}
