% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewMultiplePlp.R
\name{viewMultiplePlp}
\alias{viewMultiplePlp}
\title{open a local shiny app for viewing the result of a multiple PLP analyses}
\usage{
viewMultiplePlp(analysesLocation)
}
\arguments{
\item{analysesLocation}{The directory containing the results (with the analysis_x folders)}
}
\description{
open a local shiny app for viewing the result of a multiple PLP analyses
}
\details{
Opens a shiny app for viewing the results of the models from various T,O, Tar and settings
settings.
}
