% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewPlp.R
\name{viewPlp}
\alias{viewPlp}
\title{viewPlp - Interactively view the performance and model settings}
\usage{
viewPlp(runPlp, validatePlp = NULL)
}
\arguments{
\item{runPlp}{The output of runPlp() (an object of class 'runPlp')}

\item{validatePlp}{The output of externalValidatePlp (on object of class 'validatePlp')}
}
\value{
Opens a shiny app for interactively viewing the results
}
\description{
This is a shiny app for viewing interactive plots of the performance and the settings
}
\details{
Either the result of runPlp and view the plots
}
